egui::atomics::atoms

Trait IntoAtoms

pub trait IntoAtoms<'a> {
    // Required method
    fn collect(self, atoms: &mut Atoms<'a>);

    // Provided method
    fn into_atoms(self) -> Atoms<'a>
       where Self: Sized { ... }
}

Trait for turning a tuple of Atoms into Atoms.

Many widgets take an impl IntoAtoms parameter, which allows you to easily create atoms from tuples of text, images, and other atoms:

let image = egui::include_image!("../../../eframe/data/icon.png");
ui.button((image, "Click me!"));

Required Methods

fn collect(self, atoms: &mut Atoms<'a>)

Provided Methods

fn into_atoms(self) -> Atoms<'a>
where Self: Sized,