egui::cache
Struct FrameCache

pub struct FrameCache<Value, Computer> { /* private fields */ }

Caches the results of a computation for one frame. If it is still used next frame, it is not recomputed. If it is not used next frame, it is evicted from the cache to save memory.

Implementations

impl<Value, Computer> FrameCache<Value, Computer>

pub fn new(computer: Computer) -> Self

pub fn evict_cache(&mut self)

Must be called once per frame to clear the cache.

impl<Value, Computer> FrameCache<Value, Computer>

pub fn get<Key>(&mut self, key: Key) -> Value
where Key: Copy + Hash, Value: Clone, Computer: ComputerMut<Key, Value>,

Get from cache (if the same key was used last frame) or recompute and store in the cache.