egui::containers

Module old_popup

Old and deprecated API for popups. Use Popup instead.

Enums

AboveOrBelow

Indicate whether a popup will be shown above or below the box.

Functions

popup_above_or_below_widgetDeprecated

Shows a popup above or below another widget.

popup_below_widgetDeprecated

Helper for popup_above_or_below_widget.

show_tooltipDeprecated

Show a tooltip at the current pointer position (if any).

show_tooltip_atDeprecated

Show a tooltip at the given position.

show_tooltip_at_pointerDeprecated

Show a tooltip at the current pointer position (if any).

show_tooltip_forDeprecated

Show a tooltip under the given area.

show_tooltip_textDeprecated

Show some text at the current pointer position (if any).

was_tooltip_open_last_frameDeprecated

Was this tooltip visible last frame?