egui::containers::old_popup

Function show_tooltip_for

pub fn show_tooltip_for<R>(
    ctx: &Context,
    parent_layer: LayerId,
    widget_id: Id,
    widget_rect: &Rect,
    add_contents: impl FnOnce(&mut Ui) -> R,
) -> Option<R>

👎Deprecated: Use egui::Tooltip instead

Show a tooltip under the given area.

If the tooltip does not fit under the area, it tries to place it above it instead.