egui::containers::panel

Struct CentralPanel

pub struct CentralPanel {
    frame: Option<Frame>,
}

A panel that covers the remainder of the screen, i.e. whatever area is left after adding other panels.

This acts very similar to Frame::central_panel, but always expands to use up all available space.

The order in which you add panels matter! The first panel you add will always be the outermost, and the last you add will always be the innermost.

⚠ CentralPanel must be added after all other panels!

NOTE: Any crate::Windows and crate::Areas will cover the top-level CentralPanel.

See the module level docs for more details.

egui::Panel::top("my_panel").show_inside(ui, |ui| {
   ui.label("Hello World! From `Panel`, that must be before `CentralPanel`!");
});
egui::CentralPanel::default().show_inside(ui, |ui| {
   ui.label("Hello World!");
});

Fields
frame: Option<Frame>
Implementations

impl CentralPanel

pub fn no_frame() -> Self

A central panel with no margin or background color

pub fn default_margins() -> Self

A central panel with a background color and some inner margins

pub fn frame(self, frame: Frame) -> Self

Change the background color, margins, etc.

pub fn show_inside<R>( self, ui: &mut Ui, add_contents: impl FnOnce(&mut Ui) -> R, ) -> InnerResponse<R>

Show the panel inside a Ui.

fn show_inside_dyn<'c, R>( self, ui: &mut Ui, add_contents: Box<dyn FnOnce(&mut Ui) -> R + 'c>, ) -> InnerResponse<R>

Show the panel inside a Ui.

pub fn show<R>( self, ctx: &Context, add_contents: impl FnOnce(&mut Ui) -> R, ) -> InnerResponse<R>

👎Deprecated:
Use show_inside() instead

Show the panel at the top level.

fn show_dyn<'c, R>( self, ctx: &Context, add_contents: Box<dyn FnOnce(&mut Ui) -> R + 'c>, ) -> InnerResponse<R>

Show the panel at the top level.