egui::containers::panel

Type Alias TopBottomPanel

pub type TopBottomPanel = Panel;

👎Deprecated:
Use Panel::top or Panel::bottom instead

Aliased Type

pub struct TopBottomPanel {
    side: PanelSide,
    id: Id,
    frame: Option<Frame>,
    resizable: bool,
    show_separator_line: bool,
    default_size: Option<f32>,
    size_range: Rangef,
}

Fields
side: PanelSideid: Idframe: Option<Frame>resizable: boolshow_separator_line: booldefault_size: Option<f32>

The size is defined as being either the width for a Vertical Panel or the height for a Horizontal Panel.

size_range: Rangef

The size is defined as being either the width for a Vertical Panel or the height for a Horizontal Panel.