egui::containers

Module scene

Structs

DragPanButtons

Specifies which pointer buttons can be used to pan the scene by dragging.

Scene

A container that allows you to zoom and pan.

Functions

fit_to_rect_in_scene 🔒

Creates a transformation that fits a given scene rectangle into the available screen size.