egui::containers

Module scroll_area

See ScrollArea for docs.

Structs

Prepared 🔒

ScrollArea

Add vertical and/or horizontal scrolling to a contained Ui.

ScrollAreaOutput

ScrollSource

What is the source of scrolling for a ScrollArea.

ScrollingToTarget 🔒

State

Enums

ScrollBarVisibility

Indicate whether the horizontal and vertical scroll bars must be always visible, hidden or visible when needed.

Functions

paint_fade_areas 🔒

Paint fade-out gradients at the top and/or bottom of a scroll area to indicate that more content is available beyond the visible region.