egui::containers

Struct ComboBox

pub struct ComboBox {
    id_salt: Id,
    label: Option<WidgetText>,
    selected_text: WidgetText,
    width: Option<f32>,
    height: Option<f32>,
    icon: Option<IconPainter>,
    wrap_mode: Option<TextWrapMode>,
    close_behavior: Option<PopupCloseBehavior>,
    popup_style: StyleModifier,
}

A drop-down selection menu with a descriptive label.

let before = selected;
egui::ComboBox::from_label("Select one!")
    .selected_text(format!("{:?}", selected))
    .show_ui(ui, |ui| {
        ui.selectable_value(&mut selected, Enum::First, "First");
        ui.selectable_value(&mut selected, Enum::Second, "Second");
        ui.selectable_value(&mut selected, Enum::Third, "Third");
    }
);

if selected != before {
    // Handle selection change
}

Fields
id_salt: Idlabel: Option<WidgetText>selected_text: WidgetTextwidth: Option<f32>height: Option<f32>icon: Option<IconPainter>wrap_mode: Option<TextWrapMode>close_behavior: Option<PopupCloseBehavior>popup_style: StyleModifier
Implementations

impl ComboBox

pub fn new(id_salt: impl Hash, label: impl Into<WidgetText>) -> Self

Create new ComboBox with id and label

pub fn from_label(label: impl Into<WidgetText>) -> Self

Label shown next to the combo box

pub fn from_id_salt(id_salt: impl Hash) -> Self

Without label.

pub fn from_id_source(id_salt: impl Hash) -> Self

👎Deprecated:
Renamed from_id_salt

Without label.

pub fn width(self, width: f32) -> Self

Set the outer width of the button and menu.

Default is Spacing::combo_width.

pub fn height(self, height: f32) -> Self

Set the maximum outer height of the menu.

Default is Spacing::combo_height.

pub fn selected_text(self, selected_text: impl Into<WidgetText>) -> Self

What we show as the currently selected value

pub fn icon( self, icon_fn: impl FnOnce(&Ui, Rect, &WidgetVisuals, bool) + 'static, ) -> Self

Use the provided function to render a different ComboBox icon. Defaults to a triangle that expands when the cursor is hovering over the ComboBox.

For example:

pub fn filled_triangle(
    ui: &egui::Ui,
    rect: egui::Rect,
    visuals: &egui::style::WidgetVisuals,
    _is_open: bool,
) {
    let rect = egui::Rect::from_center_size(
        rect.center(),
        egui::vec2(rect.width() * 0.6, rect.height() * 0.4),
    );
    ui.painter().add(egui::Shape::convex_polygon(
        vec![rect.left_top(), rect.right_top(), rect.center_bottom()],
        visuals.fg_stroke.color,
        visuals.fg_stroke,
    ));
}

egui::ComboBox::from_id_salt("my-combobox")
    .selected_text(text)
    .icon(filled_triangle)
    .show_ui(ui, |_ui| {});

pub fn wrap_mode(self, wrap_mode: TextWrapMode) -> Self

Controls the wrap mode used for the selected text.

By default, Ui::wrap_mode will be used, which can be overridden with crate::Style::wrap_mode.

Note that any \n in the text will always produce a new line.

pub fn wrap(self) -> Self

Set Self::wrap_mode to TextWrapMode::Wrap.

pub fn truncate(self) -> Self

Set Self::wrap_mode to TextWrapMode::Truncate.

pub fn close_behavior(self, close_behavior: PopupCloseBehavior) -> Self

Controls the close behavior for the popup.

By default, PopupCloseBehavior::CloseOnClick will be used.

pub fn popup_style(self, popup_style: StyleModifier) -> Self

Set the style of the popup menu.

Could for example be used with crate::containers::menu::menu_style to get the frame-less menu button style.

pub fn show_ui<R>( self, ui: &mut Ui, menu_contents: impl FnOnce(&mut Ui) -> R, ) -> InnerResponse<Option<R>>

Show the combo box, with the given ui code for the menu contents.

Returns InnerResponse { inner: None } if the combo box is closed.

fn show_ui_dyn<'c, R>( self, ui: &mut Ui, menu_contents: Box<dyn FnOnce(&mut Ui) -> R + 'c>, ) -> InnerResponse<Option<R>>

pub fn show_index<Text: Into<WidgetText>>( self, ui: &mut Ui, selected: &mut usize, len: usize, get: impl Fn(usize) -> Text, ) -> Response

Show a list of items with the given selected index.

let alternatives = ["a", "b", "c", "d"];
let mut selected = 2;
egui::ComboBox::from_label("Select one!").show_index(
    ui,
    &mut selected,
    alternatives.len(),
    |i| alternatives[i]
);

pub fn is_open(ctx: &Context, id: Id) -> bool

Check if the ComboBox with the given id has its popup menu currently opened.

fn widget_to_popup_id(widget_id: Id) -> Id

Convert a ComboBox id to the id used to store it’s popup state.