Struct egui::containers::ScrollArea

pub struct ScrollArea { /* private fields */ }

Add vertical and/or horizontal scrolling to a contained Ui.

egui::ScrollArea::vertical().show(ui, |ui| {
    // Add a lot of widgets here.
});

You can scroll to an element using Response::scroll_to_me, Ui::scroll_to_cursor and Ui::scroll_to_rect.

Implementations

impl ScrollArea

pub fn horizontal() -> Self

Create a horizontal scroll area.

pub fn vertical() -> Self

Create a vertical scroll area.

pub fn both() -> Self

Create a bi-directional (horizontal and vertical) scroll area.

pub fn neither() -> Self

Create a scroll area where both direction of scrolling is disabled. It’s unclear why you would want to do this.

pub fn new(has_bar: [bool; 2]) -> Self

Create a scroll area where you decide which axis has scrolling enabled. For instance, ScrollAre::new([true, false]) enable horizontal scrolling.

pub fn max_width(self, max_width: f32) -> Self

The maximum width of the outer frame of the scroll area.

Use f32::INFINITY if you want the scroll area to expand to fit the surrounding Ui (default).

See also Self::auto_shrink.

pub fn max_height(self, max_height: f32) -> Self

The maximum height of the outer frame of the scroll area.

Use f32::INFINITY if you want the scroll area to expand to fit the surrounding Ui (default).

See also Self::auto_shrink.

pub fn min_scrolled_width(self, min_scrolled_width: f32) -> Self

The minimum width of a horizontal scroll area which requires scroll bars.

The ScrollArea will only become smaller than this if the content is smaller than this (and so we don’t require scroll bars).

Default: 64.0.

pub fn min_scrolled_height(self, min_scrolled_height: f32) -> Self

The minimum height of a vertical scroll area which requires scroll bars.

The ScrollArea will only become smaller than this if the content is smaller than this (and so we don’t require scroll bars).

Default: 64.0.

pub fn always_show_scroll(self, always_show_scroll: bool) -> Self

If false (default), the scroll bar will be hidden when not needed/ If true, the scroll bar will always be displayed even if not needed.

pub fn id_source(self, id_source: impl Hash) -> Self

A source for the unique Id, e.g. .id_source("second_scroll_area") or .id_source(loop_index).

pub fn scroll_offset(self, offset: Vec2) -> Self

Set the horizontal and vertical scroll offset position.

See also: Self::vertical_scroll_offset, Self::horizontal_scroll_offset, Ui::scroll_to_cursor and Response::scroll_to_me

pub fn vertical_scroll_offset(self, offset: f32) -> Self

Set the vertical scroll offset position.

See also: Self::scroll_offset, Ui::scroll_to_cursor and Response::scroll_to_me

pub fn horizontal_scroll_offset(self, offset: f32) -> Self

Set the horizontal scroll offset position.

See also: Self::scroll_offset, Ui::scroll_to_cursor and Response::scroll_to_me

pub fn hscroll(self, hscroll: bool) -> Self

Turn on/off scrolling on the horizontal axis.

pub fn vscroll(self, vscroll: bool) -> Self

Turn on/off scrolling on the vertical axis.

pub fn scroll2(self, has_bar: [bool; 2]) -> Self

Turn on/off scrolling on the horizontal/vertical axes.

pub fn enable_scrolling(self, enable: bool) -> Self

Control the scrolling behavior If true (default), the scroll area will respond to user scrolling If false, the scroll area will not respond to user scrolling

This can be used, for example, to optionally freeze scrolling while the user is inputing text in a TextEdit widget contained within the scroll area.

This controls both scrolling directions.

pub fn auto_shrink(self, auto_shrink: [bool; 2]) -> Self

For each axis, should the containing area shrink if the content is small?

If true, egui will add blank space outside the scroll area. If false, egui will add blank space inside the scroll area.

Default: [true; 2].

pub fn stick_to_right(self) -> Self

The scroll handle will stick to the rightmost position even while the content size changes dynamically. This can be useful to simulate text scrollers coming in from right hand side. The scroll handle remains stuck until user manually changes position. Once “unstuck” it will remain focused on whatever content viewport the user left it on. If the scroll handle is dragged all the way to the right it will again become stuck and remain there until manually pulled from the end position.

pub fn stick_to_bottom(self) -> Self

The scroll handle will stick to the bottom position even while the content size changes dynamically. This can be useful to simulate terminal UIs or log/info scrollers. The scroll handle remains stuck until user manually changes position. Once “unstuck” it will remain focused on whatever content viewport the user left it on. If the scroll handle is dragged to the bottom it will again become stuck and remain there until manually pulled from the end position.

impl ScrollArea

pub fn show<R>(
self,
ui: &mut Ui,
add_contents: impl FnOnce(&mut Ui) -> R
) -> ScrollAreaOutput<R>

Show the ScrollArea, and add the contents to the viewport.

If the inner area can be very long, consider using Self::show_rows instead.

pub fn show_rows<R>(
self,
ui: &mut Ui,
row_height_sans_spacing: f32,
total_rows: usize,
add_contents: impl FnOnce(&mut Ui, Range<usize>) -> R
) -> ScrollAreaOutput<R>

Efficiently show only the visible part of a large number of rows.

let text_style = egui::TextStyle::Body;
let row_height = ui.text_style_height(&text_style);
// let row_height = ui.spacing().interact_size.y; // if you are adding buttons instead of labels.
let total_rows = 10_000;
egui::ScrollArea::vertical().show_rows(ui, row_height, total_rows, |ui, row_range| {
    for row in row_range {
        let text = format!("Row {}/{}", row + 1, total_rows);
        ui.label(text);
    }
});

pub fn show_viewport<R>(
self,
ui: &mut Ui,
add_contents: impl FnOnce(&mut Ui, Rect) -> R
) -> ScrollAreaOutput<R>

This can be used to only paint the visible part of the contents.

add_contents is past the viewport, which is the relative view of the content. So if the passed rect has min = zero, then show the top left content (the user has not scrolled).