egui::containers

Struct Tooltip

pub struct Tooltip<'a> {
    pub popup: Popup<'a>,
    parent_layer: LayerId,
    parent_widget: Id,
}

Fields
popup: Popup<'a>parent_layer: LayerId

The layer of the parent widget.

parent_widget: Id

The id of the widget that owns this tooltip.

Implementations

impl Tooltip<'_>

pub fn new( parent_widget: Id, ctx: Context, anchor: impl Into<PopupAnchor>, parent_layer: LayerId, ) -> Self

๐Deprecated: Use Tooltip::always_open instead.

Show a tooltip that is always open.

pub fn always_open( ctx: Context, parent_layer: LayerId, parent_widget: Id, anchor: impl Into<PopupAnchor>, ) -> Self

Show a tooltip that is always open.

pub fn for_widget(response: &Response) -> Self

Show a tooltip for a widget. Always open (as long as this function is called).

pub fn for_enabled(response: &Response) -> Self

Show a tooltip when hovering an enabled widget.

pub fn for_disabled(response: &Response) -> Self

Show a tooltip when hovering a disabled widget.

pub fn at_pointer(self) -> Self

Show the tooltip at the pointer position.

pub fn gap(self, gap: f32) -> Self

Set the gap between the tooltip and the anchor

Default: 5.0

pub fn layout(self, layout: Layout) -> Self

Set the layout of the tooltip

pub fn width(self, width: f32) -> Self

Set the width of the tooltip

pub fn show<R>( self, content: impl FnOnce(&mut Ui) -> R, ) -> Option<InnerResponse<R>>

Show the tooltip

fn when_was_a_toolip_last_shown_id() -> Id

pub fn seconds_since_last_tooltip(ctx: &Context) -> f32

fn remember_that_tooltip_was_shown(ctx: &Context)

pub fn next_tooltip_id(ctx: &Context, widget_id: Id) -> Id

What is the id of the next tooltip for this widget?

pub fn tooltip_id(widget_id: Id, tooltip_count: usize) -> Id

pub fn should_show_tooltip( response: &Response, allow_interactive_tooltip: bool, ) -> bool

Should we show a tooltip for this response?

Argument allow_interactive_tooltip controls whether mouse can interact with tooltip that contains interactive widgets

pub fn was_tooltip_open_last_frame(ctx: &Context, widget_id: Id) -> bool

Was this tooltip visible last frame?