Module egui::containers::window

Structs

PossibleInteractions 🔒

TitleBar 🔒

Window

Builder for a floating window which can be dragged, closed, collapsed, resized and scrolled (off by default).

WindowInteraction 🔒

Either a move or resize

Functions

close_button 🔒

Paints the “Close” button of the window and processes clicks on it.

interact 🔒

move_and_resize_window 🔒

paint_frame_interaction 🔒

Fill in parts of the window frame when we resize by dragging that part

paint_resize_corner 🔒

resize_hover 🔒

show_title_bar 🔒

window_interaction 🔒

Returns Some if there is a move or resize