egui

Module context

Structs

Context

Your handle to egui.

ContextImpl 🔒

RepaintCause

What called Context::request_repaint or Context::request_discard?

RequestRepaintInfo

Information given to the backend about when it is time to repaint the ui.

ViewportRepaintInfo 🔒

Per-viewport state related to repaint scheduling.

ViewportState

State stored per viewport.

WrappedTextureManager 🔒

Constants

IMMEDIATE_VIEWPORT_RENDERER 🔒

Functions

warn_if_rect_changes_id 🔒

Check if any Rect appears with different Ids between two passes.