egui::data::input

Struct Modifiers

pub struct Modifiers {
    pub alt: bool,
    pub ctrl: bool,
    pub shift: bool,
    pub mac_cmd: bool,
    pub command: bool,
}

State of the modifier keys. These must be fed to egui.

The best way to compare Modifiers is by using Modifiers::matches_logically or Modifiers::matches_exact.

To access the Modifiers you can use the crate::Context::input function

let modifiers = ctx.input(|i| i.modifiers);

NOTE: For cross-platform uses, ALT+SHIFT is a bad combination of modifiers as on mac that is how you type special characters, so those key presses are usually not reported to egui.

Fields
alt: bool

Either of the alt keys are down (option ⌥ on Mac).

ctrl: bool

Either of the control keys are down. When checking for keyboard shortcuts, consider using Self::command instead.

shift: bool

Either of the shift keys are down.

mac_cmd: bool

The Mac ⌘ Command key. Should always be set to false on other platforms.

command: bool

On Windows and Linux, set this to the same value as ctrl. On Mac, this should be set whenever one of the ⌘ Command keys are down (same as mac_cmd). This is so that egui can, for instance, select all text by checking for command + A and it will work on both Mac and Windows.

Implementations

impl Modifiers

pub const NONE: Self

pub const ALT: Self

pub const CTRL: Self

pub const SHIFT: Self

pub const MAC_CMD: Self

The Mac ⌘ Command key

pub const COMMAND: Self

On Mac: ⌘ Command key, elsewhere: Ctrl key

pub const fn plus(self, rhs: Self) -> Self

assert_eq!(
    Modifiers::CTRL | Modifiers::ALT,
    Modifiers { ctrl: true, alt: true, ..Default::default() }
);
assert_eq!(
    Modifiers::ALT.plus(Modifiers::CTRL),
    Modifiers::CTRL.plus(Modifiers::ALT),
);
assert_eq!(
    Modifiers::CTRL | Modifiers::ALT,
    Modifiers::CTRL.plus(Modifiers::ALT),
);

pub fn is_none(&self) -> bool

pub fn any(&self) -> bool

pub fn all(&self) -> bool

pub fn shift_only(&self) -> bool

Is shift the only pressed button?

pub fn command_only(&self) -> bool

true if only Self::ctrl or only Self::mac_cmd is pressed.

pub fn matches_logically(&self, pattern: Self) -> bool

Checks that the ctrl/cmd matches, and that the shift/alt of the argument is a subset of the pressed key (self).

This means that if the pattern has not set shift, then self can have shift set or not.

The reason is that many logical keys require shift or alt on some keyboard layouts. For instance, in order to press + on an English keyboard, you need to press shift and =, but a Swedish keyboard has dedicated + key. So if you want to make a KeyboardShortcut looking for Cmd + +, it makes sense to ignore the shift key. Similarly, the Alt key is sometimes used to type special characters.

However, if the pattern (the argument) explicitly requires the shift or alt keys to be pressed, then they must be pressed.

Example:

if pressed_modifiers.matches_logically(Modifiers::ALT | Modifiers::SHIFT) {
    // Alt and Shift are pressed, but not ctrl/command
}

Behavior:

assert!(Modifiers::CTRL.matches_logically(Modifiers::CTRL));
assert!(!Modifiers::CTRL.matches_logically(Modifiers::CTRL | Modifiers::SHIFT));
assert!((Modifiers::CTRL | Modifiers::SHIFT).matches_logically(Modifiers::CTRL));
assert!((Modifiers::CTRL | Modifiers::COMMAND).matches_logically(Modifiers::CTRL));
assert!((Modifiers::CTRL | Modifiers::COMMAND).matches_logically(Modifiers::COMMAND));
assert!((Modifiers::MAC_CMD | Modifiers::COMMAND).matches_logically(Modifiers::COMMAND));
assert!(!Modifiers::COMMAND.matches_logically(Modifiers::MAC_CMD));

pub fn matches_exact(&self, pattern: Self) -> bool

Check for equality but with proper handling of Self::command.

self here are the currently pressed modifiers, and the argument the pattern we are testing for.

Note that this will require the shift and alt keys match, even though these modifiers are sometimes required to produce some logical keys. For instance, to press + on an English keyboard, you need to press shift and =, but on a Swedish keyboard you can press the dedicated + key. Therefore, you often want to use Self::matches_logically instead.

Example:

if pressed_modifiers.matches_exact(Modifiers::ALT | Modifiers::SHIFT) {
    // Alt and Shift are pressed, and nothing else
}

Behavior:

assert!(Modifiers::CTRL.matches_exact(Modifiers::CTRL));
assert!(!Modifiers::CTRL.matches_exact(Modifiers::CTRL | Modifiers::SHIFT));
assert!(!(Modifiers::CTRL | Modifiers::SHIFT).matches_exact(Modifiers::CTRL));
assert!((Modifiers::CTRL | Modifiers::COMMAND).matches_exact(Modifiers::CTRL));
assert!((Modifiers::CTRL | Modifiers::COMMAND).matches_exact(Modifiers::COMMAND));
assert!((Modifiers::MAC_CMD | Modifiers::COMMAND).matches_exact(Modifiers::COMMAND));
assert!(!Modifiers::COMMAND.matches_exact(Modifiers::MAC_CMD));

pub fn matches_any(&self, pattern: Self) -> bool

Check if any of the modifiers match exactly.

Returns true if the same modifier is pressed in self as in pattern, for at least one modifier.

Behavior:

assert!(Modifiers::CTRL.matches_any(Modifiers::CTRL));
assert!(Modifiers::CTRL.matches_any(Modifiers::CTRL | Modifiers::SHIFT));
assert!((Modifiers::CTRL | Modifiers::SHIFT).matches_any(Modifiers::CTRL));

pub fn cmd_ctrl_matches(&self, pattern: Self) -> bool

Checks only cmd/ctrl, not alt/shift.

self here are the currently pressed modifiers, and the argument the pattern we are testing for.

This takes care to properly handle the difference between Self::ctrl, Self::command and Self::mac_cmd.

pub fn contains(&self, query: Self) -> bool

Whether another set of modifiers is contained in this set of modifiers with proper handling of Self::command.

assert!(Modifiers::default().contains(Modifiers::default()));
assert!(Modifiers::CTRL.contains(Modifiers::default()));
assert!(Modifiers::CTRL.contains(Modifiers::CTRL));
assert!(Modifiers::CTRL.contains(Modifiers::COMMAND));
assert!(Modifiers::MAC_CMD.contains(Modifiers::COMMAND));
assert!(Modifiers::COMMAND.contains(Modifiers::MAC_CMD));
assert!(Modifiers::COMMAND.contains(Modifiers::CTRL));
assert!(!(Modifiers::ALT | Modifiers::CTRL).contains(Modifiers::SHIFT));
assert!((Modifiers::CTRL | Modifiers::SHIFT).contains(Modifiers::CTRL));
assert!(!Modifiers::CTRL.contains(Modifiers::CTRL | Modifiers::SHIFT));

impl Modifiers

pub fn ui(&self, ui: &mut Ui)