egui

Module debug_text

This is an example of how to create a plugin for egui.

A plugin is a struct that implements the Plugin trait and holds some state. The plugin is registered with the Context using Context::add_plugin to get callbacks on certain events (Plugin::on_begin_pass, Plugin::on_end_pass).

Structs

DebugTextPlugin

A plugin for easily showing debug-text on-screen.

Entry 🔒

Functions

print

Print this text next to the cursor at the end of the pass.