egui

Module hit_test

Structs

WidgetHits

Result of a hit-test against WidgetRects.

Functions

contains_circle 🔒

Returns true if the rectangle contains the whole circle.

find_closest 🔒

find_closest_within 🔒

hit_test

Find the top or closest widgets to the given position, none which is closer than search_radius.

hit_test_on_close 🔒