Module egui::input_state

Re-exports

pub use crate::data::input::Key;

pub use touch_state::MultiTouchInfo;

Modules

touch_state 🔒

Structs

Click 🔒

A pointer (mouse or touch) click.

InputState

Input state that egui updates each frame.

PointerState

Mouse or touch state.

Enums

PointerEvent 🔒

Constants

MAX_CLICK_DIST 🔒

If the pointer moves more than this, it won’t become a click (but it is still a drag)

MAX_CLICK_DURATION 🔒

If the pointer is down for longer than this, it won’t become a click (but it is still a drag)

MAX_DOUBLE_CLICK_DELAY 🔒

The new pointer press must come within this many seconds from previous pointer release