egui::input_state::touch_state

Struct TouchState

pub(crate) struct TouchState {
    device_id: TouchDeviceId,
    active_touches: BTreeMap<TouchId, ActiveTouch>,
    gesture_state: Option<GestureState>,
}

The current state (for a specific touch device) of touch events and gestures.

Fields
device_id: TouchDeviceId

Technical identifier of the touch device. This is used to identify relevant touch events for this TouchState instance.

active_touches: BTreeMap<TouchId, ActiveTouch>

Active touches, if any.

TouchId is the unique identifier of the touch. It is valid as long as the finger/pen touches the surface. The next touch will receive a new unique ID.

Refer to ActiveTouch.

gesture_state: Option<GestureState>

If a gesture has been recognized (i.e. when exactly two fingers touch the surface), this holds state information

Implementations

impl TouchState

pub fn new(device_id: TouchDeviceId) -> Self

pub fn begin_pass( &mut self, time: f64, new: &RawInput, pointer_pos: Option<Pos2>, )

pub fn any_touches(&self) -> bool

Are there currently any fingers touching the surface?

pub fn info(&self) -> Option<MultiTouchInfo>

fn update_gesture(&mut self, time: f64, pointer_pos: Option<Pos2>)

fn calc_dynamic_state(&self) -> Option<DynGestureState>

None if less than two fingers

impl TouchState

pub fn ui(&self, ui: &mut Ui)