egui::memory

Struct Focus

pub(crate) struct Focus {
    focused_widget: Option<FocusWidget>,
    id_previous_frame: Option<Id>,
    id_next_frame: Option<Id>,
    id_requested_by_accesskit: Option<NodeId>,
    give_to_next: bool,
    last_interested: Option<Id>,
    focus_direction: FocusDirection,
    top_modal_layer: Option<LayerId>,
    top_modal_layer_current_frame: Option<LayerId>,
    focus_widgets_cache: IdMap<Rect>,
}

Keeps tracks of what widget has keyboard focus

Fields
focused_widget: Option<FocusWidget>

The widget with keyboard focus (i.e. a text input field).

id_previous_frame: Option<Id>

The ID of a widget that had keyboard focus during the previous frame.

id_next_frame: Option<Id>

The ID of a widget to give the focus to in the next frame.

id_requested_by_accesskit: Option<NodeId>give_to_next: bool

If set, the next widget that is interested in focus will automatically get it. Probably because the user pressed Tab.

last_interested: Option<Id>

The last widget interested in focus.

focus_direction: FocusDirection

Set when looking for widget with navigational keys like arrows, tab, shift+tab.

top_modal_layer: Option<LayerId>

The top-most modal layer from the previous frame.

top_modal_layer_current_frame: Option<LayerId>

The top-most modal layer from the current frame.

focus_widgets_cache: IdMap<Rect>

A cache of widget IDs that are interested in focus with their corresponding rectangles.

Implementations

impl Focus

pub fn focused(&self) -> Option<Id>

Which widget currently has keyboard focus?

fn begin_pass(&mut self, new_input: &RawInput)

pub(crate) fn end_pass(&mut self, used_ids: &IdMap<Rect>)

pub(crate) fn had_focus_last_frame(&self, id: Id) -> bool

fn interested_in_focus(&mut self, id: Id)

fn set_modal_layer(&mut self, layer_id: LayerId)

pub(crate) fn top_modal_layer(&self) -> Option<LayerId>

fn reset_focus(&mut self)

fn find_widget_in_direction(&mut self, new_rects: &IdMap<Rect>) -> Option<Id>