egui::memory

Struct Memory

pub struct Memory {
    pub options: Options,
    pub data: IdTypeMap,
    pub caches: CacheStorage,
    pub(crate) new_font_definitions: Option<FontDefinitions>,
    pub(crate) add_fonts: Vec<FontInsert>,
    pub(crate) viewport_id: ViewportId,
    everything_is_visible: bool,
    pub to_global: HashMap<LayerId, TSTransform>,
    areas: ViewportIdMap<Areas>,
    pub(crate) interactions: ViewportIdMap<InteractionState>,
    pub(crate) focus: ViewportIdMap<Focus>,
    popups: ViewportIdMap<OpenPopup>,
}

The data that egui persists between frames.

This includes window positions and sizes, how far the user has scrolled in a ScrollArea etc.

If you want this to persist when closing your app, you should serialize Memory and store it. For this you need to enable the persistence feature.

If you want to store data for your widgets, you should look at Memory::data

Fields
options: Options

Global egui options.

data: IdTypeMap

This map stores some superficial state for all widgets with custom Ids.

This includes storing whether a crate::CollapsingHeader is open, how far scrolled a crate::ScrollArea is, where the cursor in a crate::TextEdit is, etc.

This is NOT meant to store any important data. Store that in your own structures!

Each read clones the data, so keep your values cheap to clone. If you want to store a lot of data, you should wrap it in Arc<Mutex<…>> so it is cheap to clone.

This will be saved between different program runs if you use the persistence feature.

To store a state common for all your widgets (a singleton), use Id::NULL as the key.

caches: CacheStorage

Can be used to cache computations from one frame to another.

This is for saving CPU time when you have something that may take 1-100ms to compute. Very slow operations (>100ms) should instead be done async (i.e. in another thread) so as not to lock the UI thread.

use egui::cache::{ComputerMut, FrameCache};

#[derive(Default)]
struct CharCounter {}
impl ComputerMut<&str, usize> for CharCounter {
    fn compute(&mut self, s: &str) -> usize {
        s.chars().count() // you probably want to cache something more expensive than this
    }
}
type CharCountCache<'a> = FrameCache<usize, CharCounter>;

ctx.memory_mut(|mem| {
    let cache = mem.caches.cache::<CharCountCache<'_>>();
    assert_eq!(*cache.get("hello"), 5);
});

new_font_definitions: Option<FontDefinitions>

new fonts that will be applied at the start of the next frame

add_fonts: Vec<FontInsert>

add new font that will be applied at the start of the next frame

viewport_id: ViewportIdeverything_is_visible: boolto_global: HashMap<LayerId, TSTransform>

Transforms per layer.

Instead of using this directly, use:

crate::Context::set_transform_layer

crate::Context::layer_transform_to_global

crate::Context::layer_transform_from_global

areas: ViewportIdMap<Areas>interactions: ViewportIdMap<InteractionState>focus: ViewportIdMap<Focus>popups: ViewportIdMap<OpenPopup>

Which popup-window is open on a viewport (if any)? Could be a combo box, color picker, menu, etc. Optionally stores the position of the popup (usually this would be the position where the user clicked). If position is None, the popup position will be calculated based on some configuration (e.g. relative to some other widget).

Implementations

impl Memory

pub(crate) fn begin_pass( &mut self, new_raw_input: &RawInput, viewports: &ViewportIdSet, )

pub(crate) fn end_pass(&mut self, used_ids: &IdMap<Rect>)

pub(crate) fn set_viewport_id(&mut self, viewport_id: ViewportId)

pub fn areas(&self) -> &Areas

Access memory of the Areas, such as Windows.

pub fn areas_mut(&mut self) -> &mut Areas

Access memory of the Areas, such as Windows.

pub fn layer_id_at(&self, pos: Pos2) -> Option<LayerId>

Top-most layer at the given position.

pub fn layer_transforms(&self, layer_id: LayerId) -> Option<TSTransform>

👎Deprecated:
Use Context::layer_transform_to_global instead

The currently set transform of a layer.

pub fn layer_ids(&self) -> impl ExactSizeIterator<Item = LayerId> + '_

An iterator over all layers. Back-to-front, top is last.

pub fn had_focus_last_frame(&self, id: Id) -> bool

Check if the layer had focus last frame. returns true if the layer had focus last frame, but not this one.

pub(crate) fn lost_focus(&self, id: Id) -> bool

Check if the layer lost focus last frame. returns true if the layer lost focus last frame, but not this one.

pub(crate) fn gained_focus(&self, id: Id) -> bool

Check if the layer gained focus this frame. returns true if the layer gained focus this frame, but not last one.

pub fn has_focus(&self, id: Id) -> bool

Does this widget have keyboard focus?

This function does not consider whether the UI as a whole (e.g. window) has the keyboard focus. That makes this function suitable for deciding widget state that should not be disrupted if the user moves away from the window and back.

pub fn focused(&self) -> Option<Id>

Which widget has keyboard focus?

pub fn set_focus_lock_filter(&mut self, id: Id, event_filter: EventFilter)

Set an event filter for a widget.

This allows you to control whether the widget will loose focus when the user presses tab, arrow keys, or escape.

You must first give focus to the widget before calling this.

pub fn request_focus(&mut self, id: Id)

Give keyboard focus to a specific widget. See also crate::Response::request_focus.

pub fn surrender_focus(&mut self, id: Id)

Surrender keyboard focus for a specific widget. See also crate::Response::surrender_focus.

pub fn move_focus(&mut self, direction: FocusDirection)

Move keyboard focus in a specific direction.

pub fn is_above_modal_layer(&self, layer_id: LayerId) -> bool

Returns true if

this layer is the top-most modal layer or above it

there is no modal layer

pub fn allows_interaction(&self, layer_id: LayerId) -> bool

Does this layer allow interaction? Returns true if

the layer is not behind a modal layer

the Order allows interaction

pub fn interested_in_focus(&mut self, id: Id, layer_id: LayerId)

Register this widget as being interested in getting keyboard focus. This will allow the user to select it with tab and shift-tab. This is normally done automatically when handling interactions, but it is sometimes useful to pre-register interest in focus, e.g. before deciding which type of underlying widget to use, as in the crate::DragValue widget, so a widget can be focused and rendered correctly in a single frame.

Pass in the layer_id of the layer that the widget is in.

pub fn set_modal_layer(&mut self, layer_id: LayerId)

Limit focus to widgets on the given layer and above. If this is called multiple times per frame, the top layer wins.

pub fn top_modal_layer(&self) -> Option<LayerId>

Get the top modal layer (from the previous frame).

pub fn stop_text_input(&mut self)

Stop editing the active TextEdit (if any).

pub fn reset_areas(&mut self)

Forget window positions, sizes etc. Can be used to auto-layout windows.

pub fn area_rect(&self, id: impl Into<Id>) -> Option<Rect>

Obtain the previous rectangle of an area.

pub(crate) fn interaction(&self) -> &InteractionState

pub(crate) fn interaction_mut(&mut self) -> &mut InteractionState

pub(crate) fn focus(&self) -> Option<&Focus>

pub(crate) fn focus_mut(&mut self) -> &mut Focus

impl Memory

Deprecated popup API

Use crate::Popup instead.

pub fn is_popup_open(&self, popup_id: Id) -> bool

👎Deprecated:
Use Popup::is_id_open instead

Is the given popup open?

pub fn any_popup_open(&self) -> bool

👎Deprecated:
Use Popup::is_any_open instead

Is any popup open?

pub fn open_popup(&mut self, popup_id: Id)

👎Deprecated:
Use Popup::open_id instead

Open the given popup and close all others.

Note that you must call keep_popup_open on subsequent frames as long as the popup is open.

pub fn keep_popup_open(&mut self, popup_id: Id)

👎Deprecated:
Use Popup::show instead

Popups must call this every frame while open.

This is needed because in some cases popups can go away without close_popup being called. For example, when a context menu is open and the underlying widget stops being rendered.

pub fn open_popup_at(&mut self, popup_id: Id, pos: impl Into<Option<Pos2>>)

👎Deprecated:
Use Popup with PopupAnchor::Position instead

Open the popup and remember its position.

pub fn popup_position(&self, id: Id) -> Option<Pos2>

👎Deprecated:
Use Popup::position_of_id instead

Get the position for this popup.

pub fn close_all_popups(&mut self)

👎Deprecated:
Use Popup::close_all instead

Close any currently open popup.

pub fn close_popup(&mut self, popup_id: Id)

👎Deprecated:
Use Popup::close_id instead

Close the given popup, if it is open.

See also Self::close_all_popups if you want to close any / all currently open popups.

pub fn toggle_popup(&mut self, popup_id: Id)

👎Deprecated:
Use Popup::toggle_id instead

Toggle the given popup between closed and open.

Note: At most, only one popup can be open at a time.

impl Memory

pub fn everything_is_visible(&self) -> bool

If true, all windows, menus, tooltips, etc., will be visible at once.

This is useful for testing, benchmarking, pre-caching, etc.

Experimental feature!

pub fn set_everything_is_visible(&mut self, value: bool)

If true, all windows, menus, tooltips etc are to be visible at once.

This is useful for testing, benchmarking, pre-caching, etc.

Experimental feature!