egui

Module menu

👎Deprecated:
Use egui::containers::menu instead

Deprecated menu API - Use crate::containers::menu instead.

Usage:

fn show_menu(ui: &mut egui::Ui) {
    use egui::{menu, Button};

    menu::bar(ui, |ui| {
        ui.menu_button("File", |ui| {
            if ui.button("Open").clicked() {
                // …
            }
        });
    });
}

Structs

BarStateDeprecated

What is saved between frames.

MenuRootDeprecated

Menu root associated with an Id from a Response

MenuRootManagerDeprecated

Stores the state for the context menu.

MenuStateDeprecated

Components of menu state, public for advanced usage.

SubMenuDeprecated

SubMenuButtonDeprecated

Enums

MenuResponseDeprecated

Constants

CONTEXT_MENU_ID_STR 🔒 Deprecated

Functions

barDeprecated

The menu bar goes well in a crate::Panel::top, but can also be placed in a crate::Window. In the latter case you may want to wrap it in Frame.

context_menuDeprecated

Response to secondary clicks (right-clicks) by showing the given menu.

context_menu_openedDeprecated

Returns true if the context menu is opened for this widget.

menu_buttonDeprecated

Construct a top level menu in a menu bar. This would be e.g. “File”, “Edit” etc.

menu_custom_buttonDeprecated

Construct a top level menu with a custom button in a menu bar.

menu_image_buttonDeprecated

Construct a top level menu with an image in a menu bar. This would be e.g. “File”, “Edit” etc.

menu_popup 🔒 Deprecated

wrapper for the contents of every menu.

set_menu_style 🔒 Deprecated

stationary_menu_button_impl 🔒 Deprecated

Build a top level menu with an image button.

stationary_menu_impl 🔒 Deprecated

Build a top level menu with a button.

submenu_buttonDeprecated

Construct a nested sub menu in another menu.