Enum egui::output::CursorIcon

pub enum CursorIcon {

    Default,
    None,
    ContextMenu,
    Help,
    PointingHand,
    Progress,
    Wait,
    Cell,
    Crosshair,
    Text,
    VerticalText,
    Alias,
    Copy,
    Move,
    NoDrop,
    NotAllowed,
    Grab,
    Grabbing,
    AllScroll,
    ResizeHorizontal,
    ResizeNeSw,
    ResizeNwSe,
    ResizeVertical,
    ResizeEast,
    ResizeSouthEast,
    ResizeSouth,
    ResizeSouthWest,
    ResizeWest,
    ResizeNorthWest,
    ResizeNorth,
    ResizeNorthEast,
    ResizeColumn,
    ResizeRow,
    ZoomIn,
    ZoomOut,

}

A mouse cursor icon.

egui emits a CursorIcon in PlatformOutput each frame as a request to the integration.

Loosely based on https://developer.mozilla.org/en-US/docs/Web/CSS/cursor.

Variants

Default

Normal cursor icon, whatever that is.

None

Show no cursor

ContextMenu

A context menu is available

Help

Question mark

PointingHand

Pointing hand, used for e.g. web links

Progress

Shows that processing is being done, but that the program is still interactive.

Wait

Not yet ready, try later.

Cell

Hover a cell in a table

Crosshair

For precision work

Text

Text caret, e.g. “Click here to edit text”

VerticalText

Vertical text caret, e.g. “Click here to edit vertical text”

Alias

Indicated an alias, e.g. a shortcut

Copy

Indicate that a copy will be made

Move

Omnidirectional move icon (e.g. arrows in all cardinal directions)

NoDrop

Can’t drop here

NotAllowed

Forbidden

Grab

The thing you are hovering can be grabbed

Grabbing

You are grabbing the thing you are hovering

AllScroll

Something can be scrolled in any direction (panned).

ResizeHorizontal

Horizontal resize - to make something wider or more narrow (left to/from right)

ResizeNeSw

Diagonal resize / (right-up to/from left-down)

ResizeNwSe

Diagonal resize \ (left-up to/from right-down)

ResizeVertical

Vertical resize | (up-down or down-up)

ResizeEast

Resize something rightwards (e.g. when dragging the right-most edge of something)

ResizeSouthEast

Resize something down and right (e.g. when dragging the bottom-right corner of something)

ResizeSouth

Resize something downwards (e.g. when dragging the bottom edge of something)

ResizeSouthWest

Resize something down and left (e.g. when dragging the bottom-left corner of something)

ResizeWest

Resize something leftwards (e.g. when dragging the left edge of something)

ResizeNorthWest

Resize something up and left (e.g. when dragging the top-left corner of something)

ResizeNorth

Resize something up (e.g. when dragging the top edge of something)

ResizeNorthEast

Resize something up and right (e.g. when dragging the top-right corner of something)

ResizeColumn

Resize a column

ResizeRow

Resize a row

ZoomIn

Enhance!

ZoomOut

Let’s get a better overview

Implementations

impl CursorIcon

pub const ALL: [CursorIcon; 35]