Struct egui::output::FullOutput

pub struct FullOutput {
    pub platform_output: PlatformOutput,
    pub textures_delta: TexturesDelta,
    pub shapes: Vec<ClippedShape>,
    pub pixels_per_point: f32,
    pub viewport_output: ViewportIdMap<ViewportOutput>,
}

What egui emits each frame from crate::Context::run.

The backend should use this.

Fields
platform_output: PlatformOutput

Non-rendering related output.

textures_delta: TexturesDelta

Texture changes since last frame (including the font texture).

The backend needs to apply crate::TexturesDelta::set before painting, and free any texture in crate::TexturesDelta::free after painting.

It is assumed that all egui viewports share the same painter and texture namespace.

shapes: Vec<ClippedShape>

What to paint.

You can use crate::Context::tessellate to turn this into triangles.

pixels_per_point: f32

The number of physical pixels per logical ui point, for the viewport that was updated.

You can pass this to crate::Context::tessellate together with Self::shapes.

viewport_output: ViewportIdMap<ViewportOutput>

All the active viewports, including the root.

It is up to the integration to spawn a native window for each viewport, and to close any window that no longer has a viewport in this map.

Implementations

impl FullOutput

pub fn append(&mut self, newer: Self)

Add on new output.