Struct egui::response::Response

pub struct Response {
    pub ctx: Context,
    pub layer_id: LayerId,
    pub id: Id,
    pub rect: Rect,
    pub sense: Sense,
    /* private fields */
}

The result of adding a widget to a Ui.

A Response lets you know whether or not a widget is being hovered, clicked or dragged. It also lets you easily show a tooltip on hover.

Whenever something gets added to a Ui, a Response object is returned. [ui.add] returns a Response, as does [ui.button], and all similar shortcuts.

Fields
ctx: Context

Used for optionally showing a tooltip and checking for more interactions.

layer_id: LayerId

Which layer the widget is part of.

id: Id

The Id of the widget/area this response pertains.

rect: Rect

The area of the screen we are talking about.

sense: Sense

The senses (click and/or drag) that the widget was interested in (if any).

Implementations

impl Response

pub fn clicked(&self) -> bool

Returns true if this widget was clicked this frame by the primary button.

A click is registered when the mouse or touch is released within a certain amount of time and distance from when and where it was pressed.

Note that the widget must be sensing clicks with Sense::click. crate::Button senses clicks; crate::Label does not (unless you call crate::Label::sense).

You can use Self::interact to sense more things after adding a widget.

pub fn clicked_by(&self, button: PointerButton) -> bool

Returns true if this widget was clicked this frame by the given button.

pub fn secondary_clicked(&self) -> bool

Returns true if this widget was clicked this frame by the secondary mouse button (e.g. the right mouse button).

pub fn middle_clicked(&self) -> bool

Returns true if this widget was clicked this frame by the middle mouse button.

pub fn double_clicked(&self) -> bool

Returns true if this widget was double-clicked this frame by the primary button.

pub fn triple_clicked(&self) -> bool

Returns true if this widget was triple-clicked this frame by the primary button.

pub fn double_clicked_by(&self, button: PointerButton) -> bool

Returns true if this widget was double-clicked this frame by the given button.

pub fn triple_clicked_by(&self, button: PointerButton) -> bool

Returns true if this widget was triple-clicked this frame by the given button.

pub fn clicked_elsewhere(&self) -> bool

true if there was a click outside this widget this frame.

pub fn enabled(&self) -> bool

Was the widget enabled? If false, there was no interaction attempted and the widget should be drawn in a gray disabled look.

pub fn hovered(&self) -> bool

The pointer is hovering above this widget or the widget was clicked/tapped this frame.

Note that this is slightly different from checking response.rect.contains(pointer_pos). For one, the hover rectangle is slightly larger, by half of the current item spacing (to make it easier to click things). But hovered also checks that no other area is covering this response rectangle.

pub fn has_focus(&self) -> bool

This widget has the keyboard focus (i.e. is receiving key presses).

This function only returns true if the UI as a whole (e.g. window) also has the keyboard focus. That makes this function suitable for style choices, e.g. a thicker border around focused widgets.

pub fn gained_focus(&self) -> bool

True if this widget has keyboard focus this frame, but didn’t last frame.

pub fn lost_focus(&self) -> bool

The widget had keyboard focus and lost it, either because the user pressed tab or clicked somewhere else, or (in case of a crate::TextEdit) because the user pressed enter.

let response = ui.text_edit_singleline(&mut my_text);
if response.lost_focus() && ui.input(|i| i.key_pressed(egui::Key::Enter)) {
    do_request(&my_text);
}

pub fn request_focus(&self)

Request that this widget get keyboard focus.

pub fn surrender_focus(&self)

Surrender keyboard focus for this widget.

pub fn dragged(&self) -> bool

The widgets is being dragged.

To find out which button(s), query crate::PointerState::button_down (ui.input(|i| i.pointer.button_down(…))).

Note that the widget must be sensing drags with Sense::drag. crate::DragValue senses drags; crate::Label does not (unless you call crate::Label::sense).

You can use Self::interact to sense more things after adding a widget.

pub fn dragged_by(&self, button: PointerButton) -> bool

pub fn drag_started(&self) -> bool

Did a drag on this widgets begin this frame?

pub fn drag_started_by(&self, button: PointerButton) -> bool

Did a drag on this widgets by the button begin this frame?

pub fn drag_released(&self) -> bool

The widget was being dragged, but now it has been released.

pub fn drag_released_by(&self, button: PointerButton) -> bool

The widget was being dragged by the button, but now it has been released.

pub fn drag_delta(&self) -> Vec2

If dragged, how many points were we dragged and in what direction?

pub fn interact_pointer_pos(&self) -> Option<Pos2>

Where the pointer (mouse/touch) were when when this widget was clicked or dragged. None if the widget is not being interacted with.

pub fn hover_pos(&self) -> Option<Pos2>

If it is a good idea to show a tooltip, where is pointer? None if the pointer is outside the response area.

pub fn is_pointer_button_down_on(&self) -> bool

Is the pointer button currently down on this widget? This is true if the pointer is pressing down or dragging a widget

pub fn changed(&self) -> bool

What the underlying data changed?

e.g. the slider was dragged, text was entered in a TextEdit etc. Always false for something like a Button.

Can sometimes be true even though the data didn’t changed (e.g. if the user entered a character and erased it the same frame).

This is not set if the view of the data was changed. For instance, moving the cursor in a TextEdit does not set this to true.

pub fn mark_changed(&mut self)

Report the data shown by this widget changed.

This must be called by widgets that represent some mutable data, e.g. checkboxes, sliders etc.

This should be called when the content changes, but not when the view does. So we call this when the text of a crate::TextEdit, but not when the cursors changes.

pub fn on_hover_ui(self, add_contents: impl FnOnce(&mut Ui)) -> Self

Show this UI if the widget was hovered (i.e. a tooltip).

The text will not be visible if the widget is not enabled. For that, use Self::on_disabled_hover_ui instead.

If you call this multiple times the tooltips will stack underneath the previous ones.

pub fn on_disabled_hover_ui(self, add_contents: impl FnOnce(&mut Ui)) -> Self

Show this UI when hovering if the widget is disabled.

pub fn on_hover_ui_at_pointer(self, add_contents: impl FnOnce(&mut Ui)) -> Self

Like on_hover_ui, but show the ui next to cursor.

pub fn is_tooltip_open(&self) -> bool

Was the tooltip open last frame?

fn should_show_hover_ui(&self) -> bool

pub fn on_hover_text_at_pointer(self, text: impl Into<WidgetText>) -> Self

Like on_hover_text, but show the text next to cursor.

pub fn on_hover_text(self, text: impl Into<WidgetText>) -> Self

Show this text if the widget was hovered (i.e. a tooltip).

The text will not be visible if the widget is not enabled. For that, use Self::on_disabled_hover_text instead.

If you call this multiple times the tooltips will stack underneath the previous ones.

pub fn highlight(self) -> Self

Highlight this widget, to make it look like it is hovered, even if it isn’t.

The highlight takes on frame to take effect if you call this after the widget has been fully rendered.

See also Context::highlight_widget.

pub fn on_disabled_hover_text(self, text: impl Into<WidgetText>) -> Self

Show this text when hovering if the widget is disabled.

pub fn on_hover_cursor(self, cursor: CursorIcon) -> Self

When hovered, use this icon for the mouse cursor.

pub fn on_hover_and_drag_cursor(self, cursor: CursorIcon) -> Self

When hovered or dragged, use this icon for the mouse cursor.

pub fn interact(&self, sense: Sense) -> Self

Check for more interactions (e.g. sense clicks on a Response returned from a label).

Note that this call will not add any hover-effects to the widget, so when possible it is better to give the widget a Sense instead, e.g. using crate::Label::sense.

let response = ui.label("hello");
assert!(!response.clicked()); // labels don't sense clicks by default
let response = response.interact(egui::Sense::click());
if response.clicked() { /* … */ }

pub fn scroll_to_me(&self, align: Option<Align>)

Adjust the scroll position until this UI becomes visible.

If align is None, it’ll scroll enough to bring the UI into view.

See also: Ui::scroll_to_cursor, Ui::scroll_to_rect. Ui::scroll_with_delta.

egui::ScrollArea::vertical().show(ui, |ui| {
    for i in 0..1000 {
        let response = ui.button("Scroll to me");
        if response.clicked() {
            response.scroll_to_me(Some(egui::Align::Center));
        }
    }
});

pub fn widget_info(&self, make_info: impl Fn() -> WidgetInfo)

For accessibility.

Call after interacting and potential calls to Self::mark_changed.

pub fn output_event(&self, event: OutputEvent)

pub fn labelled_by(self, id: Id) -> Self

Associate a label with a control for accessibility.

Example

ui.horizontal(|ui| {
    let label = ui.label("Your name: ");
    ui.text_edit_singleline(&mut text).labelled_by(label.id);
});

pub fn context_menu(self, add_contents: impl FnOnce(&mut Ui)) -> Self

Response to secondary clicks (right-clicks) by showing the given menu.

let response = ui.add(Label::new("Right-click me!").sense(Sense::click()));
response.context_menu(|ui| {
    if ui.button("Close the menu").clicked() {
        ui.close_menu();
    }
});

See also: Ui::menu_button and Ui::close_menu.

impl Response

pub fn union(&self, other: Self) -> Self

A logical “or” operation. For instance a.union(b).hovered means “was either a or b hovered?”.

The resulting Self::id will come from the first (self) argument.

impl Response

pub fn with_new_rect(self, rect: Rect) -> Self

Returns a response with a modified Self::rect.