Struct egui::CtxRef

pub struct CtxRef(_);

A wrapper around Arc<Context>. This is how you will normally create and access a Context.

Almost all methods are marked &self, Context has interior mutability (protected by mutexes).

CtxRef is cheap to clone, and any clones refers to the same mutable data.

A CtxRef is only valid for the duration of a frame, and so you should not store a CtxRef between frames. A new CtxRef is created each frame by calling Self::run.

Example:

let mut ctx = egui::CtxRef::default();

// Game loop:
loop {
    let raw_input = egui::RawInput::default();
    let (output, shapes) = ctx.run(raw_input, |ctx| {
        egui::CentralPanel::default().show(&ctx, |ui| {
            ui.label("Hello world!");
            if ui.button("Click me").clicked() {
                // take some action here
            }
        });
    });
    let clipped_meshes = ctx.tessellate(shapes); // create triangles to paint
    handle_output(output);
    paint(clipped_meshes);
}

Implementations

impl CtxRef

pub fn run(
&mut self,
new_input: RawInput,
run_ui: impl FnOnce(&CtxRef)
) -> (Output, Vec<ClippedShape>)

Run the ui code for one frame.

Put your widgets into a SidePanel, TopBottomPanel, CentralPanel, Window or Area.

This will modify the internal reference to point to a new generation of Context. Any old clones of this CtxRef will refer to the old Context, which will not get new input.

You can alternatively run [Self::begin_single_pass_frame] and [Self::end_single_pass_frame].

// One egui context that you keep reusing:
let mut ctx = egui::CtxRef::default();

// Each frame:
let input = egui::RawInput::default();
let (output, shapes) = ctx.run(input, |ctx| {
    egui::CentralPanel::default().show(&ctx, |ui| {
        ui.label("Hello egui!");
    });
});
// handle output, paint shapes

pub fn begin_frame(&mut self, new_input: RawInput)

An alternative to calling Self::run.

// One egui context that you keep reusing:
let mut ctx = egui::CtxRef::default();

// Each frame:
let input = egui::RawInput::default();
ctx.begin_frame(input);

egui::CentralPanel::default().show(&ctx, |ui| {
    ui.label("Hello egui!");
});

let (output, shapes) = ctx.end_frame();
// handle output, paint shapes

pub fn layer_painter(&self, layer_id: LayerId) -> Painter

Get a full-screen painter for a new or existing layer

pub fn debug_painter(&self) -> Painter

Paint on top of everything else