Struct egui::RichText

pub struct RichText { /* private fields */ }

Text and optional style choices for it.

The style choices (font, color) are applied to the entire text. For more detailed control, use crate::text::LayoutJob instead.

A RichText can be used in most widgets and helper functions, e.g. Ui::label and Ui::button.

Example

use egui::{RichText, Color32};

RichText::new("Plain");
RichText::new("colored").color(Color32::RED);
RichText::new("Large and underlined").size(20.0).underline();

Implementations

impl RichText

pub fn new(text: impl Into<String>) -> Self

pub fn is_empty(&self) -> bool

pub fn text(&self) -> &str

pub fn size(self, size: f32) -> Self

Select the font size (in points). This overrides the value from Self::text_style.

pub fn family(self, family: FontFamily) -> Self

Select the font family.

This overrides the value from Self::text_style.

Only the families available in crate::FontDefinitions::families may be used.

pub fn font(self, font_id: FontId) -> Self

Select the font and size. This overrides the value from Self::text_style.

pub fn text_style(self, text_style: TextStyle) -> Self

Override the TextStyle.

pub fn fallback_text_style(self, text_style: TextStyle) -> Self

Set the TextStyle unless it has already been set

pub fn heading(self) -> Self

Use TextStyle::Heading.

pub fn monospace(self) -> Self

Use TextStyle::Monospace.

pub fn code(self) -> Self

Monospace label with different background color.

pub fn strong(self) -> Self

Extra strong text (stronger color).

pub fn weak(self) -> Self

Extra weak text (fainter color).

pub fn underline(self) -> Self

Draw a line under the text.

If you want to control the line color, use LayoutJob instead.

pub fn strikethrough(self) -> Self

Draw a line through the text, crossing it out.

If you want to control the strikethrough line color, use LayoutJob instead.

pub fn italics(self) -> Self

Tilt the characters to the right.

pub fn small(self) -> Self

Smaller text.

pub fn small_raised(self) -> Self

For e.g. exponents.

pub fn raised(self) -> Self

Align text to top. Only applicable together with Self::small().

pub fn background_color(self, background_color: impl Into<Color32>) -> Self

Fill-color behind the text.

pub fn color(self, color: impl Into<Color32>) -> Self

Override text color.

pub fn font_height(&self, fonts: &Fonts, style: &Style) -> f32

Read the font height of the selected text style.