Struct egui::Vec2

#[repr(C)]
pub struct Vec2 {
    pub x: f32,
    pub y: f32,
}

A vector has a direction and length. A Vec2 is often used to represent a size.

emath represents positions using crate::Pos2.

Normally the units are points (logical pixels).

Fields
x: f32

Rightwards. Width.

y: f32

Downwards. Height.

Implementations

impl Vec2

pub const X: Vec2

pub const Y: Vec2

pub const RIGHT: Vec2

pub const LEFT: Vec2

pub const UP: Vec2

pub const DOWN: Vec2

pub const ZERO: Vec2

pub const INFINITY: Vec2

pub const fn new(x: f32, y: f32) -> Vec2

pub const fn splat(v: f32) -> Vec2

Set both x and y to the same value.

pub fn to_pos2(self) -> Pos2

Treat this vector as a position. v.to_pos2() is equivalent to Pos2::default() + v.

pub fn normalized(self) -> Vec2

Safe normalize: returns zero if input is zero.

pub fn rot90(self) -> Vec2

Rotates the vector by 90°, i.e positive X to positive Y (clockwise in egui coordinates).

pub fn length(self) -> f32

pub fn length_sq(self) -> f32

pub fn angle(self) -> f32

Measures the angle of the vector.

use std::f32::consts::TAU;

assert_eq!(Vec2::ZERO.angle(), 0.0);
assert_eq!(Vec2::angled(0.0).angle(), 0.0);
assert_eq!(Vec2::angled(1.0).angle(), 1.0);
assert_eq!(Vec2::X.angle(), 0.0);
assert_eq!(Vec2::Y.angle(), 0.25 * TAU);

assert_eq!(Vec2::RIGHT.angle(), 0.0);
assert_eq!(Vec2::DOWN.angle(), 0.25 * TAU);
assert_eq!(Vec2::UP.angle(), -0.25 * TAU);

pub fn angled(angle: f32) -> Vec2

Create a unit vector with the given CW angle (in radians).

An angle of zero gives the unit X axis.

An angle of 𝞃/4 = 90° gives the unit Y axis.

use std::f32::consts::TAU;

assert_eq!(Vec2::angled(0.0), Vec2::X);
assert!((Vec2::angled(0.25 * TAU) - Vec2::Y).length() < 1e-5);

pub fn floor(self) -> Vec2

pub fn round(self) -> Vec2

pub fn ceil(self) -> Vec2

pub fn abs(self) -> Vec2

pub fn is_finite(self) -> bool

True if all members are also finite.

pub fn any_nan(self) -> bool

True if any member is NaN.

pub fn min(self, other: Vec2) -> Vec2

pub fn max(self, other: Vec2) -> Vec2

pub fn dot(self, other: Vec2) -> f32

The dot-product of two vectors.

pub fn min_elem(self) -> f32

Returns the minimum of self.x and self.y.

pub fn max_elem(self) -> f32

Returns the maximum of self.x and self.y.

pub fn clamp(self, min: Vec2, max: Vec2) -> Vec2