egui

Struct ViewportInfo

pub struct ViewportInfo {
    pub parent: Option<ViewportId>,
    pub title: Option<String>,
    pub events: Vec<ViewportEvent>,
    pub native_pixels_per_point: Option<f32>,
    pub monitor_size: Option<Vec2>,
    pub inner_rect: Option<Rect>,
    pub outer_rect: Option<Rect>,
    pub minimized: Option<bool>,
    pub maximized: Option<bool>,
    pub fullscreen: Option<bool>,
    pub focused: Option<bool>,
    pub occluded: Option<bool>,
}

Information about the current viewport, given as input each frame.

None means “unknown”.

All units are in ui “points”, which can be calculated from native physical pixels using pixels_per_point = crate::Context::zoom_factor * [Self::native_pixels_per_point];

Fields
parent: Option<ViewportId>

Parent viewport, if known.

title: Option<String>

Name of the viewport, if known.

events: Vec<ViewportEvent>native_pixels_per_point: Option<f32>

The OS native pixels-per-point.

This should always be set, if known.

On web this takes browser scaling into account, and corresponds to window.devicePixelRatio in JavaScript.

monitor_size: Option<Vec2>

Current monitor size in egui points.

inner_rect: Option<Rect>

The inner rectangle of the native window, in monitor space and ui points scale.

This is the content rectangle of the viewport.

eframe notes:

On Android / Wayland, this will always be None since getting the position of the window is not possible.

outer_rect: Option<Rect>

The outer rectangle of the native window, in monitor space and ui points scale.

This is the content rectangle plus decoration chrome.

eframe notes:

On Android / Wayland, this will always be None since getting the position of the window is not possible.

minimized: Option<bool>

Are we minimized?

maximized: Option<bool>

Are we maximized?

fullscreen: Option<bool>

Are we in fullscreen mode?

focused: Option<bool>

Is the window focused and able to receive input?

This should be the same as RawInput::focused.

occluded: Option<bool>

Is the window fully occluded (completely covered) by another window?

Not all platforms support this. On platforms that don’t, this will be None or Some(false).

Implementations

impl ViewportInfo

pub fn visible(&self) -> Option<bool>

Is the window considered visible for rendering purposes?

A window is not visible if it is minimized or occluded. When not visible, the UI is not painted and rendering is skipped, but application logic may still be executed by some integrations.

pub fn close_requested(&self) -> bool

This viewport has been told to close.

If this is the root viewport, the application will exit after this frame unless you send a crate::ViewportCommand::CancelClose command.

If this is not the root viewport, it is up to the user to hide this viewport the next frame.

pub fn take(&mut self) -> Self

Helper: move Self::events, clone the other fields.

pub fn ui(&self, ui: &mut Ui)