Struct egui::style::Spacing

pub struct Spacing {

    pub item_spacing: Vec2,
    pub window_margin: Margin,
    pub button_padding: Vec2,
    pub menu_margin: Margin,
    pub indent: f32,
    pub interact_size: Vec2,
    pub slider_width: f32,
    pub combo_width: f32,
    pub text_edit_width: f32,
    pub icon_width: f32,
    pub icon_width_inner: f32,
    pub icon_spacing: f32,
    pub tooltip_width: f32,
    pub indent_ends_with_horizontal_line: bool,
    pub combo_height: f32,
    pub scroll_bar_width: f32,
    pub scroll_handle_min_length: f32,
    pub scroll_bar_inner_margin: f32,
    pub scroll_bar_outer_margin: f32,

}

Controls the sizes and distances between widgets.

Fields
item_spacing: Vec2

Horizontal and vertical spacing between widgets.

To add extra space between widgets, use Ui::add_space.

item_spacing is inserted after adding a widget, so to increase the spacing between widgets A and B you need to change item_spacing before adding A.

window_margin: Margin

Horizontal and vertical margins within a window frame.

button_padding: Vec2

Button size is text size plus this on each side

menu_margin: Margin

Horizontal and vertical margins within a menu frame.

indent: f32

Indent collapsing regions etc by this much.

interact_size: Vec2

Minimum size of a DragValue, color picker button, and other small widgets. interact_size.y is the default height of button, slider, etc. Anything clickable should be (at least) this size.

slider_width: f32

Default width of a Slider.

combo_width: f32

Default (minimum) width of a ComboBox.

text_edit_width: f32

Default width of a TextEdit.

icon_width: f32

Checkboxes, radio button and collapsing headers have an icon at the start. This is the width/height of the outer part of this icon (e.g. the BOX of the checkbox).

icon_width_inner: f32

Checkboxes, radio button and collapsing headers have an icon at the start. This is the width/height of the inner part of this icon (e.g. the check of the checkbox).

icon_spacing: f32

Checkboxes, radio button and collapsing headers have an icon at the start. This is the spacing between the icon and the text

tooltip_width: f32

Width of a tooltip (on_hover_ui, on_hover_text etc).

indent_ends_with_horizontal_line: bool

End indented regions with a horizontal line

combo_height: f32

Height of a combo-box before showing scroll bars.

scroll_bar_width: f32scroll_handle_min_length: f32

Make sure the scroll handle is at least this big

scroll_bar_inner_margin: f32

Margin between contents and scroll bar.

scroll_bar_outer_margin: f32

Margin between scroll bar and the outer container (e.g. right of a vertical scroll bar).

Implementations

impl Spacing

pub fn icon_rectangles(&self, rect: Rect) -> (Rect, Rect)

Returns small icon rectangle and big icon rectangle

impl Spacing

pub fn ui(&mut self, ui: &mut Ui)