Struct egui::text::Galley

pub struct Galley {
    pub job: Arc<LayoutJob>,
    pub rows: Vec<Row, Global>,
    pub rect: Rect,
    pub mesh_bounds: Rect,
    pub num_vertices: usize,
    pub num_indices: usize,
}

Text that has been layed out, ready for painting.

You can create a Galley using crate::Fonts::layout_job;

Fields
job: Arc<LayoutJob>

The job that this galley is the result of. Contains the original string and style sections.

rows: Vec<Row, Global>

Rows of text, from top to bottom. The number of characters in all rows sum up to job.text.chars().count(). Note that each paragraph (pieces of text separated with \n) can be split up into multiple rows.

rect: Rect

Bounding rect.

rect.top() is always 0.0.

With LayoutJob::halign:

Align::LEFT: rect.left() == 0.0

Align::Center: rect.center() == 0.0

Align::RIGHT: rect.right() == 0.0

mesh_bounds: Rect

Tight bounding box around all the meshes in all the rows. Can be used for culling.

num_vertices: usize

Total number of vertices in all the row meshes.

num_indices: usize

Total number of indices in all the row meshes.

Implementations

impl Galley

pub fn is_empty(&self) -> bool

pub fn text(&self) -> &str

pub fn size(&self) -> Vec2

impl Galley

Physical positions

pub fn pos_from_pcursor(&self, pcursor: PCursor) -> Rect

Returns a 0-width Rect.

pub fn pos_from_cursor(&self, cursor: &Cursor) -> Rect

Returns a 0-width Rect.

pub fn cursor_from_pos(&self, pos: Vec2) -> Cursor

Cursor at the given position within the galley

impl Galley

Cursor positions

pub fn end(&self) -> Cursor

Cursor to one-past last character.

pub fn end_rcursor(&self) -> RCursor

impl Galley

Cursor conversions

pub fn from_ccursor(&self, ccursor: CCursor) -> Cursor

pub fn from_rcursor(&self, rcursor: RCursor) -> Cursor

pub fn from_pcursor(&self, pcursor: PCursor) -> Cursor

impl Galley

Cursor positions

pub fn cursor_left_one_character(&self, cursor: &Cursor) -> Cursor

pub fn cursor_right_one_character(&self, cursor: &Cursor) -> Cursor

pub fn cursor_up_one_row(&self, cursor: &Cursor) -> Cursor

pub fn cursor_down_one_row(&self, cursor: &Cursor) -> Cursor

pub fn cursor_begin_of_row(&self, cursor: &Cursor) -> Cursor

pub fn cursor_end_of_row(&self, cursor: &Cursor) -> Cursor