egui
Module util

Miscellaneous tools used by the rest of egui.

Re-exports

pub use id_type_map::IdTypeMap;

Modules

cache

Computing the same thing each frame can be expensive, so often you want to save the result from the previous frame and reuse it.

id_type_map

undoer

Structs

History

This struct tracks recent values of some time series.

Functions

hash

Hash the given value with a predictable hasher.

hash_with

Hash the given value with the given hasher.