Struct egui::util::cache::FrameCache

pub struct FrameCache<Value, Computer> {
    generation: u32,
    computer: Computer,
    cache: IntMap<u64, (u32, Value)>,
}

Caches the results of a computation for one frame. If it is still used next frame, it is not recomputed. If it is not used next frame, it is evicted from the cache to save memory.

Fields
generation: u32computer: Computercache: IntMap<u64, (u32, Value)>
Implementations

impl<Value, Computer> FrameCache<Value, Computer>

pub fn new(computer: Computer) -> Self

pub fn evice_cache(&mut self)

Must be called once per frame to clear the cache.

impl<Value, Computer> FrameCache<Value, Computer>

pub fn get<Key>(&mut self, key: Key) -> Valuewhere Key: Copy + Hash, Value: Clone, Computer: ComputerMut<Key, Value>,

Get from cache (if the same key was used last frame) or recompute and store in the cache.