Struct egui::util::id_type_map::IdTypeMap

pub struct IdTypeMap(_);

Stores values identified by an Id AND a the std::any::TypeId of the value.

so it maps (Id, TypeId) to any value you want.

Values can either be “persisted” (serializable) or “temporary” (cleared when egui is shut down).

You can store state using the key Id::null. The state will then only be identified by its type.

let a = Id::new("a");
let b = Id::new("b");
let mut map: IdTypeMap = Default::default();

// `a` associated with an f64 and an i32
map.insert_persisted(a, 3.14);
map.insert_temp(a, 42);

// `b` associated with an f64 and a `&'static str`
map.insert_persisted(b, 6.28);
map.insert_temp(b, "Hello World".to_string());

// we can retrieve all four values:
assert_eq!(map.get_temp::<f64>(a), Some(3.14));
assert_eq!(map.get_temp::<i32>(a), Some(42));
assert_eq!(map.get_temp::<f64>(b), Some(6.28));
assert_eq!(map.get_temp::<String>(b), Some("Hello World".to_string()));

// we can retrieve them like so also:
assert_eq!(map.get_persisted::<f64>(a), Some(3.14));
assert_eq!(map.get_persisted::<i32>(a), Some(42));
assert_eq!(map.get_persisted::<f64>(b), Some(6.28));
assert_eq!(map.get_temp::<String>(b), Some("Hello World".to_string()));

Implementations

impl IdTypeMap

pub fn insert_temp<T: 'static + Any + Clone + Send + Sync>(
&mut self,
id: Id,
value: T
)

Insert a value that will not be persisted.

pub fn insert_persisted<T: SerializableAny>(&mut self, id: Id, value: T)

Insert a value that will be persisted next time you start the app.

pub fn get_temp<T: 'static + Clone>(&mut self, id: Id) -> Option<T>

Read a value without trying to deserialize a persisted value.

pub fn get_persisted<T: SerializableAny>(&mut self, id: Id) -> Option<T>

Read a value, optionally deserializing it if available.

pub fn get_temp_mut_or<T: 'static + Any + Clone + Send + Sync>(
&mut self,
id: Id,
or_insert: T
) -> &mut T

pub fn get_persisted_mut_or<T: SerializableAny>(
&mut self,
id: Id,
or_insert: T
) -> &mut T

pub fn get_temp_mut_or_default<T: 'static + Any + Clone + Send + Sync + Default>(
&mut self,
id: Id
) -> &mut T

pub fn get_persisted_mut_or_default<T: SerializableAny + Default>(
&mut self,
id: Id
) -> &mut T

pub fn get_temp_mut_or_insert_with<T: 'static + Any + Clone + Send + Sync>(
&mut self,
id: Id,
insert_with: impl FnOnce() -> T
) -> &mut T

pub fn get_persisted_mut_or_insert_with<T: SerializableAny>(
&mut self,
id: Id,
insert_with: impl FnOnce() -> T
) -> &mut T

pub fn remove<T: 'static>(&mut self, id: Id)

Remove the state of this type an id.

pub fn remove_by_type<T: 'static>(&mut self)

Note all state of the given type.

pub fn clear(&mut self)

pub fn is_empty(&mut self) -> bool

pub fn len(&mut self) -> usize

pub fn count_serialized(&mut self) -> usize

Count how many values are stored but not yet deserialized.

pub fn count<T: 'static>(&mut self) -> usize

Count the number of values are stored with the given type.