egui::util::undoer

Struct Undoer

pub struct Undoer<State> {
    settings: Settings,
    undos: VecDeque<State>,
    redos: Vec<State>,
    flux: Option<Flux<State>>,
}

Automatic undo system.

Every frame you feed it the most recent state. The Undoer compares it with the latest undo point and if there is a change it may create a new undo point.

Undoer follows two simple rules:

If the state has changed since the latest undo point, but has remained stable for stable_time seconds, an new undo point is created.

If the state does not stabilize within auto_save_interval seconds, an undo point is created.

Rule 1) will make sure an undo point is not created until you stop dragging that slider. Rule 2) will make sure that you will get some undo points even if you are constantly changing the state.

Fields
settings: Settingsundos: VecDeque<State>

New undoes are added to the back. Two adjacent undo points are never equal. The latest undo point may (often) be the current state.

redos: Vec<State>

Stores redos immediately after a sequence of undos. Gets cleared every time the state changes. Does not need to be a deque, because there can only be up to undos.len() redos, which is already limited to settings.max_undos.

flux: Option<Flux<State>>
Implementations

impl<State> Undoer<State>
where State: Clone + PartialEq,

pub fn with_settings(settings: Settings) -> Self

Create a new Undoer with the given Settings.

pub fn has_undo(&self, current_state: &State) -> bool

Do we have an undo point different from the given state?

pub fn has_redo(&self, current_state: &State) -> bool

pub fn is_in_flux(&self) -> bool

Return true if the state is currently changing

pub fn undo(&mut self, current_state: &State) -> Option<&State>

pub fn redo(&mut self, current_state: &State) -> Option<&State>

pub fn add_undo(&mut self, current_state: &State)

Add an undo point if, and only if, there has been a change since the latest undo point.

pub fn feed_state(&mut self, current_time: f64, current_state: &State)

Call this as often as you want (e.g. every frame) and Undoer will determine if a new undo point should be created.

current_time: current time in seconds.