egui::widgets

Module color_picker

Color picker widgets.

Enums

Alpha

What options to show for alpha

Constants

N 🔒

Number of vertices per dimension in the color sliders. We need at least 6 for hues, and more for smooth 2D areas. Should always be a multiple of 6 to hit the peak hues in HSV/HSL (every 60°).

Functions

background_checkers 🔒

color_button 🔒

color_cache_get 🔒

color_cache_set 🔒

color_edit_button_hsva

color_edit_button_rgb

Shows a button with the given color. If the user clicks the button, a full color picker is shown.

color_edit_button_rgba

Shows a button with the given color. If the user clicks the button, a full color picker is shown.

color_edit_button_srgb

Shows a button with the given color. If the user clicks the button, a full color picker is shown. The given color is in sRGB space.

color_edit_button_srgba

Shows a button with the given color. If the user clicks the button, a full color picker is shown.

color_picker_color32

Shows a color picker where the user can change the given Color32 color.

color_picker_hsva_2d

Shows a color picker where the user can change the given Hsva color.

color_picker_hsvag_2d 🔒

color_slider_1d 🔒

color_slider_2d 🔒

Arguments

contrast_color 🔒

input_type_button_ui 🔒

is_additive_alpha 🔒

We use a negative alpha for additive colors within this file (a bit ironic).

rgba_edit_ui 🔒

Shows 4 DragValue widgets to be used to edit the RGBA f32 values. Alpha’s DragValue is hidden when Alpha::Opaque.

show_color

Show a color with background checkers to demonstrate transparency (if any).

show_color32 🔒

show_color_at

Show a color with background checkers to demonstrate transparency (if any).

srgba_edit_ui 🔒

Shows 4 DragValue widgets to be used to edit the RGBA u8 values. Alpha’s DragValue is hidden when Alpha::Opaque.

use_color_cache 🔒