egui::widgets
Enum ImageSource

pub enum ImageSource<'a> {
    Uri(Cow<'a, str>),
    Texture(SizedTexture),
    Bytes {
        uri: Cow<'static, str>,
        bytes: Bytes,
    },
}

This type tells the Ui how to load an image.

This is used by Image::new and Ui::image.

Variants

Uri(Cow<'a, str>)

Load the image from a URI, e.g. https://example.com/image.png.

This could be a file:// path, https:// url, bytes:// identifier, or some other scheme.

How the URI will be turned into a texture for rendering purposes is up to the registered loaders to handle.

See crate::load for more information.

Texture(SizedTexture)

Load the image from an existing texture.

The user is responsible for loading the texture, determining its size, and allocating a crate::TextureId for it.

Bytes

Load the image from some raw bytes.

The Bytes may be:

'static, obtained from include_bytes! or similar

Anything that can be converted to Arc<[u8]>

This instructs the Ui to cache the raw bytes, which are then further processed by any registered loaders.

See also crate::include_image for an easy way to load and display static images.

See crate::load for more information.

Fields

uri: Cow<'static, str>

The unique identifier for this image, e.g. bytes://my_logo.png.

You should use a proper extension (.jpg, .png, .svg, etc) for the image to load properly.

Use the bytes:// scheme for the URI for better error messages.

bytes: Bytes

Implementations

impl<'a> ImageSource<'a>

pub fn texture_size(&self) -> Option<Vec2>

Size of the texture, if known.

pub fn load( self, ctx: &Context, texture_options: TextureOptions, size_hint: SizeHint, ) -> TextureLoadResult

Errors

Failure to load the texture.

pub fn uri(&self) -> Option<&str>

Get the uri that this image was constructed from.

This will return None for Self::Texture.