egui::widgets

Module image

Structs

FrameDurations

Stores the durations between each frame of an animated image

Image

A widget which displays an image.

ImageOptions

ImageSize

This type determines the constraints on how the size of an image should be calculated.

Enums

ImageFit

This type determines how the image should try to fit within the UI.

ImageSource

This type tells the Ui how to load an image.

Functions

animated_image_frame_index 🔒

Calculates at which frame the animated image is

are_animated_image_bytes 🔒

decode_animated_image_uri

Extracts uri and frame index

encode_animated_image_uri 🔒

Animated image uris contain the uri & the frame that will be displayed

has_gif_magic_header

Checks if bytes are gifs

has_webp_header

Checks if bytes are webp

is_animated_image_uri 🔒

is_gif_uri 🔒

Checks if uri is a gif file

is_webp_uri 🔒

Checks if uri is a webp file

paint_texture_at

paint_texture_load_result

scale_to_fit 🔒

texture_load_result_response

Attach tooltips like “Loading…” or “Failed loading: …”.