Module egui::widgets::plot::items

Contains items that can be added to a plot.

Re-exports

pub use bar::Bar;

pub use box_elem::BoxElem;

pub use box_elem::BoxSpread;

pub use values::LineStyle;

pub use values::MarkerShape;

pub use values::Orientation;

pub use values::PlotPoint;

pub use values::PlotPoints;

Modules

bar 🔒

box_elem 🔒

rect_elem 🔒

values 🔒

Structs

Arrows

A set of arrows.

BarChart

A bar chart.

BoxPlot

A diagram containing a series of BoxElem elements.

HLine

A horizontal line in a plot, filling the full width

Line

A series of values forming a path.

PlotConfig 🔒

Container to pass-through several parameters related to plot visualization

PlotImage

An image in the plot.

Points

A set of points.

Polygon

A convex polygon.

Text

Text inside the plot.

VLine

A vertical line in a plot, filling the full width

Constants

DEFAULT_FILL_ALPHA 🔒

Traits

PlotItem 🔒

Trait shared by things that can be drawn in the plot.

Functions

add_rulers_and_text 🔒

find_closest_rect 🔒

horizontal_line 🔒

rulers_at_value 🔒

Draws a cross of horizontal and vertical ruler at the pointer position. value is used to for text displaying X/Y coordinates.

rulers_color 🔒

vertical_line 🔒

y_intersection 🔒

Returns the x-coordinate of a possible intersection between a line segment from p1 to p2 and a horizontal line at the given y-coordinate.