Struct egui::widgets::plot::items::bar::Bar

pub struct Bar {
    pub name: String,
    pub orientation: Orientation,
    pub argument: f64,
    pub value: f64,
    pub base_offset: Option<f64>,
    pub bar_width: f64,
    pub stroke: Stroke,
    pub fill: Color32,
}

One bar in a BarChart. Potentially floating, allowing stacked bar charts. Width can be changed to allow variable-width histograms.

Fields
name: String

Name of plot element in the diagram (annotated by default formatter)

orientation: Orientation

Which direction the bar faces in the diagram

argument: f64

Position on the argument (input) axis – X if vertical, Y if horizontal

value: f64

Position on the value (output) axis – Y if vertical, X if horizontal

base_offset: Option<f64>

For stacked bars, this denotes where the bar starts. None if base axis

bar_width: f64

Thickness of the bar

stroke: Stroke

Line width and color

fill: Color32

Fill color

Implementations

impl Bar

pub fn new(argument: f64, height: f64) -> Bar

Create a bar. Its orientation is set by its BarChart parent.

argument: Position on the argument axis (X if vertical, Y if horizontal).

value: Height of the bar (if vertical).

By default the bar is vertical and its base is at zero.

pub fn name(self, name: impl ToString) -> Self

Name of this bar chart element.

pub fn stroke(self, stroke: impl Into<Stroke>) -> Self

Add a custom stroke.

pub fn fill(self, color: impl Into<Color32>) -> Self

Add a custom fill color.

pub fn base_offset(self, offset: f64) -> Self

Offset the base of the bar. This offset is on the Y axis for a vertical bar and on the X axis for a horizontal bar.

pub fn width(self, width: f64) -> Self

Set the bar width.

pub fn vertical(self) -> Self

Set orientation of the element as vertical. Argument axis is X.

pub fn horizontal(self) -> Self

Set orientation of the element as horizontal. Argument axis is Y.

pub(super) fn lower(&self) -> f64

pub(super) fn upper(&self) -> f64

pub(super) fn add_shapes( &self, transform: &PlotTransform, highlighted: bool, shapes: &mut Vec<Shape> )

pub(super) fn add_rulers_and_text( &self, parent: &BarChart, plot: &PlotConfig<'_>, shapes: &mut Vec<Shape>, cursors: &mut Vec<Cursor> )