Enum egui::widgets::plot::items::values::PlotPoints

pub enum PlotPoints {
    Owned(Vec<PlotPoint>),
    Generator(ExplicitGenerator),
}

Represents many PlotPoints.

These can be an owned Vec or generated with a function.

Variants

Owned(Vec<PlotPoint>)

Generator(ExplicitGenerator)

Implementations

impl PlotPoints

pub fn new(points: Vec<[f64; 2]>) -> Self

pub fn points(&self) -> &[PlotPoint]

pub fn from_explicit_callback( function: impl Fn(f64) -> f64 + 'static, x_range: impl RangeBounds<f64>, points: usize ) -> Self

Draw a line based on a function y=f(x), a range (which can be infinite) for x and the number of points.

pub fn from_parametric_callback( function: impl Fn(f64) -> (f64, f64), t_range: impl RangeBounds<f64>, points: usize ) -> Self

Draw a line based on a function (x,y)=f(t), a range for t and the number of points. The range may be specified as start..end or as start..=end.

pub fn from_ys_f32(ys: &[f32]) -> Self

From a series of y-values. The x-values will be the indices of these values

pub fn from_ys_f64(ys: &[f64]) -> Self

From a series of y-values. The x-values will be the indices of these values

pub(crate) fn is_empty(&self) -> bool

Returns true if there are no data points available and there is no function to generate any.

pub(super) fn generate_points(&mut self, x_range: RangeInclusive<f64>)

If initialized with a generator function, this will generate n evenly spaced points in the given range.

fn range_intersection( range1: &RangeInclusive<f64>, range2: &RangeInclusive<f64> ) -> Option<RangeInclusive<f64>>

Returns the intersection of two ranges if they intersect.

pub(super) fn bounds(&self) -> PlotBounds