Struct egui::widgets::plot::BarChart

pub struct BarChart { /* private fields */ }

A bar chart.

Implementations

impl BarChart

pub fn new(bars: Vec<Bar>) -> BarChart

Create a bar chart. It defaults to vertically oriented elements.

pub fn color(self, color: impl Into<Color32>) -> Self

Set the default color. It is set on all elements that do not already have a specific color. This is the color that shows up in the legend. It can be overridden at the bar level (see [Bar]). Default is Color32::TRANSPARENT which means a color will be auto-assigned.

pub fn name(self, name: impl ToString) -> Self

Name of this chart.

This name will show up in the plot legend, if legends are turned on. Multiple charts may share the same name, in which case they will also share an entry in the legend.

pub fn vertical(self) -> Self

Set all elements to be in a vertical orientation. Argument axis will be X and bar values will be on the Y axis.

pub fn horizontal(self) -> Self

Set all elements to be in a horizontal orientation. Argument axis will be Y and bar values will be on the X axis.

pub fn width(self, width: f64) -> Self

Set the width (thickness) of all its elements.

pub fn highlight(self, highlight: bool) -> Self

Highlight all plot elements.

pub fn element_formatter(
self,
formatter: Box<dyn Fn(&Bar, &BarChart) -> String>
) -> Self

Add a custom way to format an element. Can be used to display a set number of decimals or custom labels.

pub fn stack_on(self, others: &[&BarChart]) -> Self

Stacks the bars on top of another chart. Positive values are stacked on top of other positive values. Negative values are stacked below other negative values.