Module egui::widgets::slider

Structs

Slider

Control a number with a slider.

SliderSpec 🔒

Enums

SliderOrientation

Specifies the orientation of a Slider.

Constants

INF_RANGE_MAGNITUDE 🔒

When the user asks for an infinitely large range (e.g. logarithmic from zero), give a scale that this many orders of magnitude in size.

Functions

get 🔒

logarithmic_zero_cutoff 🔒

where to put the zero cutoff for logarithmic sliders that crosses zero ?

normalized_from_value 🔒

range_log10 🔒

set 🔒

value_from_normalized 🔒

Type Aliases

GetSetValue 🔒

Combined into one function (rather than two) to make it easier for the borrow checker.

NumFormatter 🔒

NumParser 🔒