egui::widgets

Struct Button

pub struct Button<'a> {
    layout: AtomLayout<'a>,
    fill: Option<Color32>,
    stroke: Option<Stroke>,
    small: bool,
    frame: Option<bool>,
    frame_when_inactive: bool,
    min_size: Vec2,
    corner_radius: Option<CornerRadius>,
    selected: bool,
    image_tint_follows_text_color: bool,
    limit_image_size: bool,
}

Clickable button with text.

See also Ui::button.

if ui.add(egui::Button::new("Click me")).clicked() {
    do_stuff();
}

// A greyed-out and non-interactive button:
if ui.add_enabled(false, egui::Button::new("Can't click this")).clicked() {
    unreachable!();
}

Fields
layout: AtomLayout<'a>fill: Option<Color32>stroke: Option<Stroke>small: boolframe: Option<bool>frame_when_inactive: boolmin_size: Vec2corner_radius: Option<CornerRadius>selected: boolimage_tint_follows_text_color: boollimit_image_size: bool
Implementations

impl<'a> Button<'a>

pub fn new(atoms: impl IntoAtoms<'a>) -> Self

pub fn selectable(selected: bool, atoms: impl IntoAtoms<'a>) -> Self

Show a selectable button.

Equivalent to:

let selected = true;
ui.add(Button::new("toggle me").selected(selected).frame_when_inactive(!selected).frame(true));

See also:

Ui::selectable_value

Ui::selectable_label

pub fn image(image: impl Into<Image<'a>>) -> Self

Creates a button with an image. The size of the image as displayed is defined by the provided size.

Note: In contrast to Button::new, this limits the image size to the default font height (using crate::AtomExt::atom_max_height_font_size).

pub fn image_and_text( image: impl Into<Image<'a>>, text: impl Into<WidgetText>, ) -> Self

Creates a button with an image to the left of the text.

Note: In contrast to Button::new, this limits the image size to the default font height (using crate::AtomExt::atom_max_height_font_size).

pub fn opt_image_and_text( image: Option<Image<'a>>, text: Option<WidgetText>, ) -> Self

Create a button with an optional image and optional text.

Note: In contrast to Button::new, this limits the image size to the default font height (using crate::AtomExt::atom_max_height_font_size).

pub fn wrap_mode(self, wrap_mode: TextWrapMode) -> Self

Set the wrap mode for the text.

By default, crate::Ui::wrap_mode will be used, which can be overridden with crate::Style::wrap_mode.

Note that any \n in the text will always produce a new line.

pub fn wrap(self) -> Self

Set Self::wrap_mode to TextWrapMode::Wrap.

pub fn truncate(self) -> Self

Set Self::wrap_mode to TextWrapMode::Truncate.

pub fn fill(self, fill: impl Into<Color32>) -> Self

Override background fill color. Note that this will override any on-hover effects. Calling this will also turn on the frame.

pub fn stroke(self, stroke: impl Into<Stroke>) -> Self

Override button stroke. Note that this will override any on-hover effects. Calling this will also turn on the frame.

pub fn small(self) -> Self

Make this a small button, suitable for embedding into text.

pub fn frame(self, frame: bool) -> Self

Turn off the frame

pub fn frame_when_inactive(self, frame_when_inactive: bool) -> Self

If false, the button will not have a frame when inactive.

Default: true.

Note: When Self::frame (or ui.visuals().button_frame) is false, this setting has no effect.

pub fn sense(self, sense: Sense) -> Self

By default, buttons senses clicks. Change this to a drag-button with Sense::drag().

pub fn min_size(self, min_size: Vec2) -> Self

Set the minimum size of the button.

pub fn corner_radius(self, corner_radius: impl Into<CornerRadius>) -> Self

Set the rounding of the button.

pub fn rounding(self, corner_radius: impl Into<CornerRadius>) -> Self

👎Deprecated:
Renamed to corner_radius

pub fn image_tint_follows_text_color( self, image_tint_follows_text_color: bool, ) -> Self

If true, the tint of the image is multiplied by the widget text color.

This makes sense for images that are white, that should have the same color as the text color. This will also make the icon color depend on hover state.

Default: false.

pub fn shortcut_text(self, shortcut_text: impl IntoAtoms<'a>) -> Self

Show some text on the right side of the button, in weak color.

Designed for menu buttons, for setting a keyboard shortcut text (e.g. Ctrl+S).

The text can be created with crate::Context::format_shortcut.

See also Self::right_text.

pub fn left_text(self, left_text: impl IntoAtoms<'a>) -> Self

Show some text on the left side of the button.

pub fn right_text(self, right_text: impl IntoAtoms<'a>) -> Self

Show some text on the right side of the button.

pub fn selected(self, selected: bool) -> Self

If true, mark this button as “selected”.

pub fn gap(self, gap: f32) -> Self

Set the gap between atoms.

pub fn atom_ui(self, ui: &mut Ui) -> AtomLayoutResponse

Show the button and return a AtomLayoutResponse for painting custom contents.