egui::widgets

Struct ProgressBar

pub struct ProgressBar {
    progress: f32,
    desired_width: Option<f32>,
    desired_height: Option<f32>,
    text: Option<ProgressBarText>,
    fill: Option<Color32>,
    animate: bool,
    corner_radius: Option<CornerRadius>,
}

A simple progress bar.

See also: crate::Spinner.

Fields
progress: f32desired_width: Option<f32>desired_height: Option<f32>text: Option<ProgressBarText>fill: Option<Color32>animate: boolcorner_radius: Option<CornerRadius>
Implementations

impl ProgressBar

pub fn new(progress: f32) -> Self

Progress in the [0, 1] range, where 1 means “completed”.

pub fn desired_width(self, desired_width: f32) -> Self

The desired width of the bar. Will use all horizontal space if not set.

pub fn desired_height(self, desired_height: f32) -> Self

The desired height of the bar. Will use the default interaction size if not set.

pub fn fill(self, color: Color32) -> Self

The fill color of the bar.

pub fn text(self, text: impl Into<WidgetText>) -> Self

A custom text to display on the progress bar.

pub fn show_percentage(self) -> Self

Show the progress in percent on the progress bar.

pub fn animate(self, animate: bool) -> Self

Whether to display a loading animation when progress < 1. Note that this will cause the UI to be redrawn. Defaults to false.

If Self::corner_radius and Self::animate are used simultaneously, the animation is not rendered, since it requires a perfect circle to render correctly. However, the UI is still redrawn.

pub fn corner_radius(self, corner_radius: impl Into<CornerRadius>) -> Self

Set the rounding of the progress bar.

If Self::corner_radius and Self::animate are used simultaneously, the animation is not rendered, since it requires a perfect circle to render correctly. However, the UI is still redrawn.

pub fn rounding(self, corner_radius: impl Into<CornerRadius>) -> Self

👎Deprecated:
Renamed to corner_radius