Type Alias egui::widgets::text_edit::state::Undoer

type Undoer = Undoer<(CCursorRange, String)>;

Aliased Type

struct Undoer {
    settings: Settings,
    undos: VecDeque<(CCursorRange, String), Global>,
    flux: Option<Flux<(CCursorRange, String)>>,
}

Fields
settings: Settingsundos: VecDeque<(CCursorRange, String), Global>

New undoes are added to the back. Two adjacent undo points are never equal. The latest undo point may (often) be the current state.

flux: Option<Flux<(CCursorRange, String)>>
Implementations

impl<State> Undoer<State>where State: Clone + PartialEq,

pub fn has_undo(&self, current_state: &State) -> bool

Do we have an undo point different from the given state?

pub fn is_in_flux(&self) -> bool

Return true if the state is currently changing

pub fn undo(&mut self, current_state: &State) -> Option<&State>

pub fn add_undo(&mut self, current_state: &State)

Add an undo point if, and only if, there has been a change since the latest undo point.

time: current time in seconds.

pub fn feed_state(&mut self, current_time: f64, current_state: &State)

Call this as often as you want (e.g. every frame) and Undoer will determine if a new undo point should be created.

current_time: current time in seconds.