Module egui::containers

Containers are pieces of the UI which wraps other pieces of UI. Examples: Window, ScrollArea, Resize, etc.

For instance, a Frame adds a frame and background to some contained UI.

Re-exports

pub use popup::*;

Modules

popup
Show popup windows, tooltips, context menus etc.

Structs

Area
An area on the screen that can be moved by dragging.

CentralPanel
A panel that covers the remainder of the screen, i.e. whatever area is left after adding other panels.

CollapsingHeader
A header which can be collapsed/expanded, revealing a contained Ui region.

CollapsingResponse
The response from showing a CollapsingHeader.

ComboBox
A drop-down selection menu with a descriptive label.

Frame
Color and margin of a rectangular background of a Ui.

Resize
A region that can be resized by dragging the bottom right corner.

ScrollArea
Add vertical scrolling to a contained Ui.

SidePanel
A panel that covers the entire left side of the screen.

TopPanel
A panel that covers the entire top side of the screen.

Window
Builder for a floating window which can be dragged, closed, collapsed, resized and scrolled (off by default).

Functions

combo_box_with_label	Deprecated
A drop-down selection menu with a descriptive label.