egui

Module containers

Containers are pieces of the UI which wraps other pieces of UI. Examples: Window, ScrollArea, Resize, SidePanel, etc.

For instance, a Frame adds a frame and background to some contained UI.

Re-exports

pub use collapsing_header::CollapsingHeader;

pub use collapsing_header::CollapsingResponse;

pub use frame::Frame;

pub use modal::Modal;

pub use modal::ModalResponse;

pub use panel::CentralPanel;

pub use panel::SidePanel;

pub use panel::TopBottomPanel;

pub use scroll_area::ScrollArea;

pub use old_popup::*;

Modules

area 🔒

Area is a Ui that has no parent, it floats on the background. It has no frame or own size. It is potentially movable. It is the foundation for windows and popups.

close_tag 🔒

collapsing_header

combo_box 🔒

frame

Frame container

menu

Popup menus, context menus and menu bars.

modal

old_popup

Old and deprecated API for popups. Use Popup instead.

panel

Panels are Ui regions taking up e.g. the left side of a Ui or screen.

popup 🔒

resize 🔒

scene 🔒

scroll_area

sides 🔒

tooltip 🔒

window 🔒

Structs

Area

An area on the screen that can be moved by dragging.

AreaState

State of an Area that is persisted between frames.

ClosableTag

A tag to mark a container as closable.

ComboBox

A drop-down selection menu with a descriptive label.

DragPanButtons

Specifies which pointer buttons can be used to pan the scene by dragging.

Popup

A popup container.

Resize

A region that can be resized by dragging the bottom right corner.

Scene

A container that allows you to zoom and pan.

Sides

Put some widgets on the left and right sides of a ui.

Tooltip

Window

Builder for a floating window which can be dragged, closed, collapsed, resized and scrolled (off by default).

Enums

PopupAnchor

What should we anchor the popup to?

PopupCloseBehavior

Determines popup’s close behavior

PopupKind

Is the popup a popup, tooltip or menu?

SetOpenCommand

Type Aliases

IconPainter

A function that paints the ComboBox icon