Module egui::containers

Containers are pieces of the UI which wraps other pieces of UI. Examples: Window, ScrollArea, Resize, SidePanel, etc.

For instance, a Frame adds a frame and background to some contained UI.

Re-exports

pub use collapsing_header::CollapsingHeader;

pub use collapsing_header::CollapsingResponse;

pub use panel::CentralPanel;

pub use panel::SidePanel;

pub use panel::TopBottomPanel;

pub use scroll_area::ScrollArea;

pub use popup::*;

Modules

collapsing_header

panel

Panels are Ui regions taking up e.g. the left side of a Ui or screen.

popup

Show popup windows, tooltips, context menus etc.

scroll_area

Structs

Area

An area on the screen that can be moved by dragging.

AreaState

State of an Area that is persisted between frames.

ComboBox

A drop-down selection menu with a descriptive label.

Frame

Add a background, frame and/or margin to a rectangular background of a Ui.

Resize

A region that can be resized by dragging the bottom right corner.

Window

Builder for a floating window which can be dragged, closed, collapsed, resized and scrolled (off by default).

Enums

AboveOrBelow

Indicate whether a popup will be shown above or below the box.

Type Aliases

IconPainter

A function that paints the ComboBox icon