Module egui::containers

Containers are pieces of the UI which wraps other pieces of UI. Examples: Window, ScrollArea, Resize, SidePanel, etc.

For instance, a Frame adds a frame and background to some contained UI.

Re-exports

pub use panel::CentralPanel;

pub use panel::SidePanel;

pub use panel::TopBottomPanel;

pub use popup::*;

Modules

panel

Panels are Ui regions taking up e.g. the left side of a Ui or screen.

popup

Show popup windows, tooltips, context menus etc.

Structs

Area

An area on the screen that can be moved by dragging.

CollapsingHeader

A header which can be collapsed/expanded, revealing a contained Ui region.

CollapsingResponse

The response from showing a CollapsingHeader.

ComboBox

A drop-down selection menu with a descriptive label.

Frame

Color and margin of a rectangular background of a Ui.

Resize

A region that can be resized by dragging the bottom right corner.

ScrollArea

Add vertical and/or horizontal scrolling to a contained Ui.

Window

Builder for a floating window which can be dragged, closed, collapsed, resized and scrolled (off by default).