Module egui::containers::panel

Panels are Ui regions taking up e.g. the left side of a Ui or screen.

Panels can either be a child of a Ui (taking up a portion of the parent) or be top-level (taking up a portion of the whole screen).

Together with Window and Area:s, top-level panels are the only places where you can put you widgets.

The order in which you add panels matter! The first panel you add will always be the outermost, and the last you add will always be the innermost.

You must never open one top-level panel from within another panel. Add one panel, then the next.

Always add any CentralPanel last.

Add your Window:s after any top-level panels.

Structs

CentralPanel

A panel that covers the remainder of the screen, i.e. whatever area is left after adding other panels.

SidePanel

A panel that covers the entire left or right side of a Ui or screen.

TopBottomPanel

A panel that covers the entire top or bottom of a Ui or screen.

Enums

Side

Left or Right

TopBottomSide

Top or Bottom