egui::containers::scene

Function fit_to_rect_in_scene

fn fit_to_rect_in_scene(
    rect_in_global: Rect,
    rect_in_scene: Rect,
    zoom_range: Rangef,
) -> TSTransform

Creates a transformation that fits a given scene rectangle into the available screen size.

The resulting visual scene bounds can be larger, due to letterboxing.

Returns the transformation from scene to global coordinates.