Struct egui::containers::scroll_area::Prepared

struct Prepared {

    id: Id,
    state: State,
    auto_shrink: Vec2b,
    scroll_enabled: Vec2b,
    show_bars_factor: Vec2,
    current_bar_use: Vec2,
    scroll_bar_visibility: ScrollBarVisibility,
    scroll_bar_rect: Option<Rect>,
    inner_rect: Rect,
    content_ui: Ui,
    viewport: Rect,
    scrolling_enabled: bool,
    stick_to_end: Vec2b,
    animated: bool,

}

Fields
id: Idstate: Stateauto_shrink: Vec2bscroll_enabled: Vec2b

Does this ScrollArea have horizontal/vertical scrolling enabled?

show_bars_factor: Vec2

Smoothly interpolated boolean of whether or not to show the scroll bars.

current_bar_use: Vec2

How much horizontal and vertical space are used up by the width of the vertical bar, and the height of the horizontal bar?

This is always zero for floating scroll bars.

Note that this is a yx swizzling of Self::show_bars_factor times the maximum bar with. That’s because horizontal scroll uses up vertical space, and vice versa.

scroll_bar_visibility: ScrollBarVisibilityscroll_bar_rect: Option<Rect>inner_rect: Rect

Where on the screen the content is (excludes scroll bars).

content_ui: Uiviewport: Rect

Relative coordinates: the offset and size of the view of the inner UI. viewport.min == ZERO means we scrolled to the top.

scrolling_enabled: boolstick_to_end: Vec2banimated: bool
Implementations

impl Prepared

fn end(self, ui: &mut Ui) -> (Vec2, State)

Returns content size and state