Struct egui::containers::Area

pub struct Area { /* fields omitted */ }

An area on the screen that can be moved by dragging.

This forms the base of the Window container.

egui::Area::new("my_area")
    .fixed_pos(egui::pos2(32.0, 32.0))
    .show(ctx, |ui| {
        ui.label("Floating text!");
    });

Implementations

impl Area

pub fn new(id_source: impl Hash) -> Self

pub fn id(self, id: Id) -> Self

pub fn layer(&self) -> LayerId

pub fn enabled(self, enabled: bool) -> Self

If false, no content responds to click and widgets will be shown grayed out. You won't be able to move the window. Default: true.

pub fn movable(self, movable: bool) -> Self

moveable by dragging the area?

pub fn is_enabled(&self) -> bool

pub fn is_movable(&self) -> bool

pub fn interactable(self, interactable: bool) -> Self

If false, clicks goes straight through to what is behind us. Good for tooltips etc.

pub fn order(self, order: Order) -> Self

order(Order::Foreground) for an Area that should always be on top

pub fn default_pos(self, default_pos: impl Into<Pos2>) -> Self

pub fn fixed_pos(self, fixed_pos: impl Into<Pos2>) -> Self

Positions the window and prevents it from being moved

pub fn current_pos(self, current_pos: impl Into<Pos2>) -> Self

Positions the window but you can still move it.

pub fn drag_bounds(self, bounds: Rect) -> Self

Constrain the area up to which the window can be dragged.

impl Area

pub fn show(self, ctx: &CtxRef, add_contents: impl FnOnce(&mut Ui)) -> Response

pub fn show_open_close_animation(
&self,
ctx: &CtxRef,
frame: &Frame,
is_open: bool
)