Struct egui::containers::Area

pub struct Area { /* private fields */ }

An area on the screen that can be moved by dragging.

This forms the base of the Window container.

egui::Area::new(egui::Id::new("my_area"))
    .fixed_pos(egui::pos2(32.0, 32.0))
    .show(ctx, |ui| {
        ui.label("Floating text!");
    });

The previous rectangle used by this area can be obtained through crate::Memory::area_rect().

Implementations

impl Area

pub fn new(id: Id) -> Self

The id must be globally unique.

pub fn id(self, id: Id) -> Self

Let’s you change the id that you assigned in Self::new.

The id must be globally unique.

pub fn layer(&self) -> LayerId

pub fn enabled(self, enabled: bool) -> Self

If false, no content responds to click and widgets will be shown grayed out. You won’t be able to move the window. Default: true.

pub fn movable(self, movable: bool) -> Self

Moveable by dragging the area?

pub fn is_enabled(&self) -> bool

pub fn is_movable(&self) -> bool

pub fn interactable(self, interactable: bool) -> Self

If false, clicks goes straight through to what is behind us. Good for tooltips etc.

pub fn sense(self, sense: Sense) -> Self

Explicitly set a sense.

If not set, this will default to Sense::drag() if movable, Sense::click() if interactable, and Sense::hover() otherwise.

pub fn order(self, order: Order) -> Self

order(Order::Foreground) for an Area that should always be on top

pub fn default_pos(self, default_pos: impl Into<Pos2>) -> Self

pub fn fixed_pos(self, fixed_pos: impl Into<Pos2>) -> Self

Positions the window and prevents it from being moved

pub fn constrain(self, constrain: bool) -> Self

Constrains this area to the screen bounds.

pub fn constrain_to(self, constrain_rect: Rect) -> Self

Constrain the movement of the window to the given rectangle.

For instance: .constrain_to(ctx.screen_rect()).

pub fn pivot(self, pivot: Align2) -> Self

Where the “root” of the area is.

For instance, if you set this to Align2::RIGHT_TOP then Self::fixed_pos will set the position of the right-top corner of the area.

Default: Align2::LEFT_TOP.

pub fn current_pos(self, current_pos: impl Into<Pos2>) -> Self

Positions the window but you can still move it.

pub fn anchor(self, align: Align2, offset: impl Into<Vec2>) -> Self

Set anchor and distance.

An anchor of Align2::RIGHT_TOP means “put the right-top corner of the window in the right-top corner of the screen”.

The offset is added to the position, so e.g. an offset of [-5.0, 5.0] would move the window left and down from the given anchor.

Anchoring also makes the window immovable.

It is an error to set both an anchor and a position.

impl Area

pub fn show<R>( self, ctx: &Context, add_contents: impl FnOnce(&mut Ui) -> R ) -> InnerResponse<R>

pub fn show_open_close_animation( &self, ctx: &Context, frame: &Frame, is_open: bool )