egui::containers
Struct Area

pub struct Area { /* private fields */ }

An area on the screen that can be moved by dragging.

This forms the base of the crate::Window container.

egui::Area::new(egui::Id::new("my_area"))
    .fixed_pos(egui::pos2(32.0, 32.0))
    .show(ctx, |ui| {
        ui.label("Floating text!");
    });

The previous rectangle used by this area can be obtained through crate::Memory::area_rect().

Implementations

impl Area

pub fn new(id: Id) -> Self

The id must be globally unique.

pub fn id(self, id: Id) -> Self

Letโs you change the id that you assigned in Self::new.

The id must be globally unique.

pub fn kind(self, kind: UiKind) -> Self

Change the UiKind of the arena.

Default to UiKind::GenericArea.

pub fn layer(&self) -> LayerId

pub fn enabled(self, enabled: bool) -> Self

If false, no content responds to click and widgets will be shown grayed out. You wonโt be able to move the window. Default: true.

pub fn movable(self, movable: bool) -> Self

Moveable by dragging the area?

pub fn is_enabled(&self) -> bool

pub fn is_movable(&self) -> bool

pub fn interactable(self, interactable: bool) -> Self

If false, clicks goes straight through to what is behind us.

Can be used for semi-invisible areas that the user should be able to click through.

Default: true.

pub fn sense(self, sense: Sense) -> Self

Explicitly set a sense.

If not set, this will default to Sense::drag() if movable, Sense::click() if interactable, and Sense::hover() otherwise.

pub fn order(self, order: Order) -> Self

order(Order::Foreground) for an Area that should always be on top

pub fn default_pos(self, default_pos: impl Into<Pos2>) -> Self

pub fn default_size(self, default_size: impl Into<Vec2>) -> Self

The size used for the Ui::max_rect the first frame.

Text will wrap at this width, and images that expand to fill the available space will expand to this size.

If the contents are smaller than this size, the area will shrink to fit the contents. If the contents overflow, the area will grow.

If not set, crate::style::Spacing::default_area_size will be used.

pub fn default_width(self, default_width: f32) -> Self

See Self::default_size.

pub fn default_height(self, default_height: f32) -> Self

See Self::default_size.

pub fn fixed_pos(self, fixed_pos: impl Into<Pos2>) -> Self

Positions the window and prevents it from being moved

pub fn constrain(self, constrain: bool) -> Self

Constrains this area to Context::screen_rect?

Default: true.

pub fn constrain_to(self, constrain_rect: Rect) -> Self

Constrain the movement of the window to the given rectangle.

For instance: .constrain_to(ctx.screen_rect()).

pub fn pivot(self, pivot: Align2) -> Self

Where the โrootโ of the area is.

For instance, if you set this to Align2::RIGHT_TOP then Self::fixed_pos will set the position of the right-top corner of the area.

Default: Align2::LEFT_TOP.

pub fn current_pos(self, current_pos: impl Into<Pos2>) -> Self

Positions the window but you can still move it.

pub fn anchor(self, align: Align2, offset: impl Into<Vec2>) -> Self

Set anchor and distance.

An anchor of Align2::RIGHT_TOP means โput the right-top corner of the window in the right-top corner of the screenโ.

The offset is added to the position, so e.g. an offset of [-5.0, 5.0] would move the window left and down from the given anchor.

Anchoring also makes the window immovable.

It is an error to set both an anchor and a position.

pub fn fade_in(self, fade_in: bool) -> Self

If true, quickly fade in the area.

Default: true.

impl Area

pub fn show<R>( self, ctx: &Context, add_contents: impl FnOnce(&mut Ui) -> R, ) -> InnerResponse<R>