Struct egui::containers::ComboBox

pub struct ComboBox {
    id_source: Id,
    label: Option<WidgetText>,
    selected_text: WidgetText,
    width: Option<f32>,
    icon: Option<IconPainter>,
    wrap_enabled: bool,
}

A drop-down selection menu with a descriptive label.

egui::ComboBox::from_label("Select one!")
    .selected_text(format!("{:?}", selected))
    .show_ui(ui, |ui| {
        ui.selectable_value(&mut selected, Enum::First, "First");
        ui.selectable_value(&mut selected, Enum::Second, "Second");
        ui.selectable_value(&mut selected, Enum::Third, "Third");
    }
);

Fields
id_source: Idlabel: Option<WidgetText>selected_text: WidgetTextwidth: Option<f32>icon: Option<IconPainter>wrap_enabled: bool
Implementations

impl ComboBox

pub fn new(id_source: impl Hash, label: impl Into<WidgetText>) -> Self

Create new ComboBox with id and label

pub fn from_label(label: impl Into<WidgetText>) -> Self

Label shown next to the combo box

pub fn from_id_source(id_source: impl Hash) -> Self

Without label.

pub fn width(self, width: f32) -> Self

Set the outer width of the button and menu.

pub fn selected_text(self, selected_text: impl Into<WidgetText>) -> Self

What we show as the currently selected value

pub fn icon( self, icon_fn: impl FnOnce(&Ui, Rect, &WidgetVisuals, bool, AboveOrBelow) + 'static ) -> Self

Use the provided function to render a different ComboBox icon. Defaults to a triangle that expands when the cursor is hovering over the ComboBox.

For example:

pub fn filled_triangle(
    ui: &egui::Ui,
    rect: egui::Rect,
    visuals: &egui::style::WidgetVisuals,
    _is_open: bool,
    _above_or_below: egui::AboveOrBelow,
) {
    let rect = egui::Rect::from_center_size(
        rect.center(),
        egui::vec2(rect.width() * 0.6, rect.height() * 0.4),
    );
    ui.painter().add(egui::Shape::convex_polygon(
        vec![rect.left_top(), rect.right_top(), rect.center_bottom()],
        visuals.fg_stroke.color,
        visuals.fg_stroke,
    ));
}

egui::ComboBox::from_id_source("my-combobox")
    .selected_text(text)
    .icon(filled_triangle)
    .show_ui(ui, |_ui| {});

pub fn wrap(self, wrap: bool) -> Self

Controls whether text wrap is used for the selected text

pub fn show_ui<R>( self, ui: &mut Ui, menu_contents: impl FnOnce(&mut Ui) -> R ) -> InnerResponse<Option<R>>

Show the combo box, with the given ui code for the menu contents.

Returns InnerResponse { inner: None } if the combo box is closed.

fn show_ui_dyn<'c, R>( self, ui: &mut Ui, menu_contents: Box<dyn FnOnce(&mut Ui) -> R + 'c> ) -> InnerResponse<Option<R>>

pub fn show_index<Text: Into<WidgetText>>( self, ui: &mut Ui, selected: &mut usize, len: usize, get: impl Fn(usize) -> Text ) -> Response

Show a list of items with the given selected index.

let alternatives = ["a", "b", "c", "d"];
let mut selected = 2;
egui::ComboBox::from_label("Select one!").show_index(
    ui,
    &mut selected,
    alternatives.len(),
    |i| alternatives[i]
);