Struct egui::containers::ComboBox

pub struct ComboBox { /* private fields */ }

A drop-down selection menu with a descriptive label.

egui::ComboBox::from_label("Select one!")
    .selected_text(format!("{:?}", selected))
    .show_ui(ui, |ui| {
        ui.selectable_value(&mut selected, Enum::First, "First");
        ui.selectable_value(&mut selected, Enum::Second, "Second");
        ui.selectable_value(&mut selected, Enum::Third, "Third");
    }
);

Implementations

impl ComboBox

pub fn from_label(label: impl Into<WidgetText>) -> Self

Label shown next to the combo box

pub fn from_id_source(id_source: impl Hash) -> Self

Without label.

pub fn width(self, width: f32) -> Self

Set the width of the button and menu

pub fn selected_text(self, selected_text: impl Into<WidgetText>) -> Self

What we show as the currently selected value

pub fn show_ui<R>(
self,
ui: &mut Ui,
menu_contents: impl FnOnce(&mut Ui) -> R
) -> InnerResponse<Option<R>>

Show the combo box, with the given ui code for the menu contents.

Returns InnerResponse { inner: None } if the combo box is closed.

pub fn show_index(
self,
ui: &mut Ui,
selected: &mut usize,
len: usize,
get: impl Fn(usize) -> String
) -> Response

Show a list of items with the given selected index.

let alternatives = ["a", "b", "c", "d"];
let mut selected = 2;
egui::ComboBox::from_label("Select one!").show_index(
    ui,
    &mut selected,
    alternatives.len(),
    |i| alternatives[i].to_owned()
);