egui::containers
Struct Window

pub struct Window<'open> { /* private fields */ }

Builder for a floating window which can be dragged, closed, collapsed, resized and scrolled (off by default).

You can customize:

title

default, minimum, maximum and/or fixed size, collapsed/expanded

if the window has a scroll area (off by default)

if the window can be collapsed (minimized) to just the title bar (yes, by default)

if there should be a close button (none by default)

egui::Window::new("My Window").show(ctx, |ui| {
   ui.label("Hello World!");
});

The previous rectangle used by this window can be obtained through crate::Memory::area_rect().

Note that this is NOT a native OS window. To create a new native OS window, use crate::Context::show_viewport_deferred.

Implementations

impl<'open> Window<'open>

pub fn new(title: impl Into<WidgetText>) -> Self

The window title is used as a unique Id and must be unique, and should not change. This is true even if you disable the title bar with .title_bar(false). If you need a changing title, you must call window.id(โฆ) with a fixed id.

pub fn id(self, id: Id) -> Self

Assign a unique id to the Window. Required if the title changes, or is shared with another window.

pub fn open(self, open: &'open mut bool) -> Self

Call this to add a close-button to the window title bar.

If *open == false, the window will not be visible.

If *open == true, the window will have a close button.

If the close button is pressed, *open will be set to false.

pub fn enabled(self, enabled: bool) -> Self

If false the window will be grayed out and non-interactive.

pub fn interactable(self, interactable: bool) -> Self

If false, clicks goes straight through to what is behind us.

Can be used for semi-invisible areas that the user should be able to click through.

Default: true.

pub fn movable(self, movable: bool) -> Self

If false the window will be immovable.

pub fn order(self, order: Order) -> Self

order(Order::Foreground) for a Window that should always be on top

pub fn fade_in(self, fade_in: bool) -> Self

If true, quickly fade in the Window when it first appears.

Default: true.

pub fn fade_out(self, fade_out: bool) -> Self

If true, quickly fade out the Window when it closes.

This only works if you use Self::open to close the window.

Default: true.

pub fn mutate(self, mutate: impl Fn(&mut Self)) -> Self

Usage: Window::new(โฆ).mutate(|w| w.resize = w.resize.auto_expand_width(true))

pub fn resize(self, mutate: impl Fn(Resize) -> Resize) -> Self

Usage: Window::new(โฆ).resize(|r| r.auto_expand_width(true))

pub fn frame(self, frame: Frame) -> Self

Change the background color, margins, etc.

pub fn min_width(self, min_width: f32) -> Self

Set minimum width of the window.

pub fn min_height(self, min_height: f32) -> Self

Set minimum height of the window.

pub fn min_size(self, min_size: impl Into<Vec2>) -> Self

Set minimum size of the window, equivalent to calling both min_width and min_height.

pub fn max_width(self, max_width: f32) -> Self

Set maximum width of the window.

pub fn max_height(self, max_height: f32) -> Self

Set maximum height of the window.

pub fn max_size(self, max_size: impl Into<Vec2>) -> Self

Set maximum size of the window, equivalent to calling both max_width and max_height.

pub fn current_pos(self, current_pos: impl Into<Pos2>) -> Self

Set current position of the window. If the window is movable it is up to you to keep track of where it moved to!

pub fn default_pos(self, default_pos: impl Into<Pos2>) -> Self

Set initial position of the window.

pub fn fixed_pos(self, pos: impl Into<Pos2>) -> Self

Sets the window position and prevents it from being dragged around.

pub fn constrain(self, constrain: bool) -> Self

Constrains this window to Context::screen_rect.

To change the area to constrain to, use Self::constrain_to.

Default: true.

pub fn constrain_to(self, constrain_rect: Rect) -> Self

Constrain the movement of the window to the given rectangle.

For instance: .constrain_to(ctx.screen_rect()).

pub fn pivot(self, pivot: Align2) -> Self

Where the โrootโ of the window is.

For instance, if you set this to Align2::RIGHT_TOP then Self::fixed_pos will set the position of the right-top corner of the window.

Default: Align2::LEFT_TOP.

pub fn anchor(self, align: Align2, offset: impl Into<Vec2>) -> Self

Set anchor and distance.

An anchor of Align2::RIGHT_TOP means โput the right-top corner of the window in the right-top corner of the screenโ.

The offset is added to the position, so e.g. an offset of [-5.0, 5.0] would move the window left and down from the given anchor.

Anchoring also makes the window immovable.

It is an error to set both an anchor and a position.

pub fn default_open(self, default_open: bool) -> Self

Set initial collapsed state of the window

pub fn default_size(self, default_size: impl Into<Vec2>) -> Self

Set initial size of the window.

pub fn default_width(self, default_width: f32) -> Self

Set initial width of the window.

pub fn default_height(self, default_height: f32) -> Self

Set initial height of the window.

pub fn fixed_size(self, size: impl Into<Vec2>) -> Self

Sets the window size and prevents it from being resized by dragging its edges.

pub fn default_rect(self, rect: Rect) -> Self

Set initial position and size of the window.

pub fn fixed_rect(self, rect: Rect) -> Self

Sets the window pos and size and prevents it from being moved and resized by dragging its edges.

pub fn resizable(self, resizable: impl Into<Vec2b>) -> Self

Can the user resize the window by dragging its edges?

Note that even if you set this to false the window may still auto-resize.

You can set the window to only be resizable in one direction by using e.g. [true, false] as the argument, making the window only resizable in the x-direction.

Default is true.

pub fn collapsible(self, collapsible: bool) -> Self

Can the window be collapsed by clicking on its title?

pub fn title_bar(self, title_bar: bool) -> Self

Show title bar on top of the window? If false, the window will not be collapsible nor have a close-button.

pub fn auto_sized(self) -> Self

Not resizable, just takes the size of its contents. Also disabled scrolling. Text will not wrap, but will instead make your window width expand.

pub fn scroll(self, scroll: impl Into<Vec2b>) -> Self

Enable/disable horizontal/vertical scrolling. false by default.

You can pass in false, true, [false, true] etc.

pub fn scroll2(self, scroll: impl Into<Vec2b>) -> Self

๐Deprecated: Renamed to scroll

Enable/disable horizontal/vertical scrolling. false by default.

pub fn hscroll(self, hscroll: bool) -> Self

Enable/disable horizontal scrolling. false by default.

pub fn vscroll(self, vscroll: bool) -> Self

Enable/disable vertical scrolling. false by default.

pub fn drag_to_scroll(self, drag_to_scroll: bool) -> Self

Enable/disable scrolling on the window by dragging with the pointer. true by default.

See ScrollArea::drag_to_scroll for more.

pub fn scroll_bar_visibility(self, visibility: ScrollBarVisibility) -> Self

Sets the ScrollBarVisibility of the window.

impl<'open> Window<'open>

pub fn show<R>( self, ctx: &Context, add_contents: impl FnOnce(&mut Ui) -> R, ) -> Option<InnerResponse<Option<R>>>

Returns None if the window is not open (if Window::open was called with &mut false). Returns Some(InnerResponse { inner: None }) if the window is collapsed.