egui::containers::window

Function close_button

fn close_button(ui: &mut Ui, rect: Rect) -> Response

Paints the “Close” button of the window and processes clicks on it.

The close button is just an X symbol painted by a current stroke for foreground elements (such as a label text).

Parameters

ui:

rect: The rectangular area to fit the button in

Returns the result of a click on a button if it was pressed