Module egui::data::input

The input needed by egui.

Structs

DroppedFile

A file dropped into egui.

HoveredFile

A file about to be dropped into egui.

KeyboardShortcut

A keyboard shortcut, e.g. Ctrl+Alt+W.

ModifierNames

Names of different modifier keys.

Modifiers

State of the modifier keys. These must be fed to egui.

RawInput

What the integrations provides to egui at the start of each frame.

TouchDeviceId

this is a u64 as values of this kind can always be obtained by hashing

TouchId

Unique identification of a touch occurrence (finger or pen or …). A Touch ID is valid until the finger is lifted. A new ID is used for the next touch.

Enums

Event

An input event generated by the integration.

Key

Keyboard keys.

MouseWheelUnit

The unit associated with the numeric value of a mouse wheel event

PointerButton

Mouse button (or similar for touch input)

TouchPhase

In what phase a touch event is in.

Constants

NUM_POINTER_BUTTONS

Number of pointer buttons supported by egui, i.e. the number of possible states of PointerButton.