Enum egui::WidgetText

pub enum WidgetText {
    RichText(RichText),
    LayoutJob(LayoutJob),
    Galley(Arc<Galley>),
}

This is how you specify text for a widget.

A lot of widgets use impl Into<WidgetText> as an argument, allowing you to pass in String, RichText, LayoutJob, and more.

Often a WidgetText is just a simple String, but it can be a RichText (text with color, style, etc), a LayoutJob (for when you want full control of how the text looks) or text that has already been layed out in a Galley.

Variants

RichText(RichText)

LayoutJob(LayoutJob)

Use this LayoutJob when laying out the text.

Only LayoutJob::text and LayoutJob::sections are guaranteed to be respected.

LayoutJob::wrap_width, LayoutJob::halign, LayoutJob::justify and LayoutJob::first_row_min_height will likely be determined by the crate::Layout of the Ui the widget is placed in. If you want all parts of the LayoutJob respected, then convert it to a Galley and use Self::Galley instead.

Galley(Arc<Galley>)

Use exactly this galley when painting the text.

Implementations

impl WidgetText

pub fn is_empty(&self) -> bool

pub fn text(&self) -> &str

pub fn text_style(self, text_style: TextStyle) -> Self

Override the TextStyle if, and only if, this is a RichText.

Prefer using RichText directly!

pub fn fallback_text_style(self, text_style: TextStyle) -> Self

Set the TextStyle unless it has already been set

Prefer using RichText directly!

pub fn color(self, color: impl Into<Color32>) -> Self

Override text color if, and only if, this is a RichText.

Prefer using RichText directly!

pub fn heading(self) -> Self

Prefer using RichText directly!

pub fn monospace(self) -> Self

Prefer using RichText directly!

pub fn code(self) -> Self

Prefer using RichText directly!

pub fn strong(self) -> Self

Prefer using RichText directly!

pub fn weak(self) -> Self

Prefer using RichText directly!

pub fn underline(self) -> Self

Prefer using RichText directly!

pub fn strikethrough(self) -> Self

Prefer using RichText directly!

pub fn italics(self) -> Self

Prefer using RichText directly!

pub fn small(self) -> Self

Prefer using RichText directly!

pub fn small_raised(self) -> Self

Prefer using RichText directly!

pub fn raised(self) -> Self

Prefer using RichText directly!

pub fn background_color(self, background_color: impl Into<Color32>) -> Self

Prefer using RichText directly!

pub fn into_text_job(
self,
style: &Style,
default_text_style: TextStyle,
default_valign: Align
) -> WidgetTextJob

pub fn into_galley(
self,
ui: &Ui,
wrap: Option<bool>,
available_width: f32,
default_text_style: TextStyle
) -> WidgetTextGalley

Layout with wrap mode based on the containing Ui.

wrap: override for Ui::wrap_text.