Module egui::layers

Handles paint layers, i.e. how things are sometimes painted behind or in front of other things.

Structs

LayerId
An identifier for a paint layer. Also acts as an identifier for Area:s.

PaintList
A list of Shapes paired with a clip rectangle.

ShapeIdx
A unique identifier of a specific Shape in a PaintList.

Enums

Order
Different layer categories