egui

Module memory

Re-exports

pub use theme::Theme;

pub use theme::ThemePreference;

Modules

theme 🔒

Structs

Areas

Keeps track of Areas, which are free-floating Uis. These Areas can be in any Order.

Focus 🔒

Keeps tracks of what widget has keyboard focus

FocusWidget 🔒

The widget with focus.

InteractionState 🔒

The state of the interaction in egui, i.e. what is being dragged.

Memory

The data that egui persists between frames.

OpenPopup 🔒

State of an open popup.

Options

Some global options that you can read and write.

Enums

FocusDirection

A direction in which to move the keyboard focus.

Type Aliases

OrderMap 🔒

Map containing the index of each layer in the order list, for quick lookups.