Module egui::output

All the data egui returns to the backend at the end of each frame.

Structs

OpenUrl

Output
What egui emits each frame. The backend should use this.

WidgetInfo
Describes a widget such as a crate::Button or a crate::TextEdit.

Enums

CursorIcon
A mouse cursor icon.

OutputEvent
Things that happened during this frame that the integration may be interested in.

WidgetEvent

WidgetType
The different types of built-in widgets in egui