egui::output

Struct PlatformOutput

pub struct PlatformOutput {
    pub commands: Vec<OutputCommand>,
    pub cursor_icon: CursorIcon,
    pub events: Vec<OutputEvent>,
    pub mutable_text_under_cursor: bool,
    pub ime: Option<IMEOutput>,
    pub accesskit_update: Option<TreeUpdate>,
    pub num_completed_passes: usize,
    pub request_discard_reasons: Vec<RepaintCause>,
}

The non-rendering part of what egui emits each frame.

You can access (and modify) this with crate::Context::output.

The backend should use this.

Fields
commands: Vec<OutputCommand>

Commands that the egui integration should execute at the end of a frame.

cursor_icon: CursorIcon

Set the cursor to this icon.

events: Vec<OutputEvent>

Events that may be useful to e.g. a screen reader.

mutable_text_under_cursor: bool

Is there a mutable TextEdit under the cursor? Use by eframe web to show/hide mobile keyboard and IME agent.

ime: Option<IMEOutput>

This is set if, and only if, the user is currently editing text.

Useful for IME.

accesskit_update: Option<TreeUpdate>

The difference in the widget tree since last frame.

NOTE: this needs to be per-viewport.

num_completed_passes: usize

How many ui passes is this the sum of?

See crate::Context::request_discard for details.

This is incremented at the END of each frame, so this will be 0 for the first pass.

request_discard_reasons: Vec<RepaintCause>

Was crate::Context::request_discard called during the latest pass?

If so, what was the reason(s) for it?

If empty, there was never any calls.

Implementations

impl PlatformOutput

pub fn events_description(&self) -> String

This can be used by a text-to-speech system to describe the events (if any).

pub fn append(&mut self, newer: Self)

Add on new output.

pub fn take(&mut self) -> Self

Take everything ephemeral (everything except cursor_icon currently)

pub fn requested_discard(&self) -> bool

Was crate::Context::request_discard called?