Module egui::special_emojis

The default egui fonts supports around 1216 emojis in total. Here are some of the most useful: ∞⊗⎗⎘⎙⏏⏴⏵⏶⏷ ⏩⏪⏭⏮⏸⏹⏺■▶📾🔀🔁🔃 ☀☁★☆☐☑☜☝☞☟⛃⛶✔ ↺↻⟲⟳⬅➡⬆⬇⬈⬉⬊⬋⬌⬍⮨⮩⮪⮫ ♡ 📅📆 📈📉📊 📋📌📎📤📥🔆 🔈🔉🔊🔍🔎🔗🔘 🕓🖧🖩🖮🖱🖴🖵🖼🗀🗁🗋🗐🗑🗙🚫❓

NOTE: In egui all emojis are monochrome!

You can explore them all in the Font Book in the online demo.

In addition, egui supports a few special emojis that are not part of the unicode standard. This module contains some of them:

Constants

GIT

The word git.

GITHUB

The Github logo.

OS_ANDROID

The Android logo.

OS_APPLE

The Apple logo.

OS_LINUX

Tux, the Linux penguin.

OS_WINDOWS

The Windows logo.

TWITTER

The Twitter bird.