Struct egui::Memory

pub struct Memory {
    pub options: Options,
    pub data: IdTypeMap,
    pub caches: CacheStorage,
    /* private fields */
}

The data that egui persists between frames.

This includes window positions and sizes, how far the user has scrolled in a ScrollArea etc.

If you want this to persist when closing your app you should serialize Memory and store it. For this you need to enable the persistence.

If you want to store data for your widgets, you should look at Memory::data

Fields
options: Optionsdata: IdTypeMap

This map stores current states for all widgets with custom Ids.

This will be saved between different program runs if you use the persistence feature.

To store a state common for all your widgets (a singleton), use Id::null as the key.

caches: CacheStorage

Can be used to cache computations from one frame to another.

This is for saving CPU when you have something that may take 1-100ms to compute. Things that are very slow (>100ms) should instead be done async (i.e. in another thread) so as not to lock the UI thread.

use egui::util::cache::{ComputerMut, FrameCache};

#[derive(Default)]
struct CharCounter {}
impl ComputerMut<&str, usize> for CharCounter {
    fn compute(&mut self, s: &str) -> usize {
        s.chars().count() // you probably want to cache something more expensive than this
    }
}
type CharCountCache<'a> = FrameCache<usize, CharCounter>;

let mut memory = ctx.memory();
let cache = memory.caches.cache::<CharCountCache<'_>>();
assert_eq!(cache.get("hello"), 5);

Implementations

impl Memory

pub fn layer_id_at(
&self,
pos: Pos2,
resize_interact_radius_side: f32
) -> Option<LayerId>

Top-most layer at the given position.

pub fn has_focus(&self, id: Id) -> bool

Does this widget have keyboard focus?

pub fn focus(&self) -> Option<Id>

Which widget has keyboard focus?

pub fn lock_focus(&mut self, id: Id, lock_focus: bool)

Prevent keyboard focus from moving away from this widget even if users presses the tab key. You must first give focus to the widget before calling this.

pub fn has_lock_focus(&mut self, id: Id) -> bool

Is the keyboard focus locked on this widget? If so the focus won’t move even if the user presses the tab key.

pub fn request_focus(&mut self, id: Id)

Give keyboard focus to a specific widget. See also crate::Response::request_focus.

pub fn surrender_focus(&mut self, id: Id)

Surrender keyboard focus for a specific widget. See also crate::Response::surrender_focus.

pub fn stop_text_input(&mut self)

Stop editing of active TextEdit (if any).

pub fn is_anything_being_dragged(&self) -> bool

pub fn is_being_dragged(&self, id: Id) -> bool

pub fn reset_areas(&mut self)

Forget window positions, sizes etc. Can be used to auto-layout windows.

impl Memory

Popups

Popups are things like combo-boxes, color pickers, menus etc. Only one can be be open at a time.

pub fn is_popup_open(&mut self, popup_id: Id) -> bool

pub fn open_popup(&mut self, popup_id: Id)

pub fn close_popup(&mut self)

pub fn toggle_popup(&mut self, popup_id: Id)

pub fn everything_is_visible(&self) -> bool

If true, all windows, menus, tooltips etc are to be visible at once.

This is useful for testing, benchmarking, pre-caching, etc.

Experimental feature!

pub fn set_everything_is_visible(&mut self, value: bool)

If true, all windows, menus, tooltips etc are to be visible at once.

This is useful for testing, benchmarking, pre-caching, etc.

Experimental feature!