Struct egui::Memory

pub struct Memory {
    pub options: Options,
    pub data: TypeMap,
    pub data_temp: TypeMap,
    pub id_data: AnyMap<Id>,
    pub id_data_temp: AnyMap<Id>,
    // some fields omitted
}

The data that egui persists between frames.

This includes window positions and sizes, how far the user has scrolled in a ScrollArea etc.

If you want this to persist when closing your app you should serialize Memory and store it.

If you want to store data for your widgets, you should look at data/data_temp and id_data/id_data_temp fields, and read the documentation of any module.

Fields
options: Optionsdata: TypeMap

This map stores current states for widgets that don’t require Id. This will be saved between different program runs if you use the persistence feature.

data_temp: TypeMap

Same as data, but this data will not be saved between runs.

id_data: AnyMap<Id>

This map stores current states for all widgets with custom Ids. This will be saved between different program runs if you use the persistence feature.

id_data_temp: AnyMap<Id>

Same as id_data, but this data will not be saved between runs.

Implementations

impl Memory

pub fn layer_id_at(
&self,
pos: Pos2,
resize_interact_radius_side: f32
) -> Option<LayerId>

pub fn has_focus(&self, id: Id) -> bool

Does this widget have keyboard focus?

pub fn focus(&self) -> Option<Id>

Which widget has keyboard focus?

pub fn request_focus(&mut self, id: Id)

Give keyboard focus to a specific widget. See also crate::Response::request_focus.

pub fn surrender_focus(&mut self, id: Id)

Surrender keyboard focus for a specific widget. See also crate::Response::surrender_focus.

pub fn stop_text_input(&mut self)

Stop editing of active TextEdit (if any).

pub fn is_anything_being_dragged(&self) -> bool

pub fn is_being_dragged(&self, id: Id) -> bool

pub fn reset_areas(&mut self)

Forget window positions, sizes etc. Can be used to auto-layout windows.

impl Memory

Popups

Popups are things like combo-boxes, color pickers, menus etc. Only one can be be open at a time.

pub fn is_popup_open(&mut self, popup_id: Id) -> bool

pub fn open_popup(&mut self, popup_id: Id)

pub fn close_popup(&mut self)

pub fn toggle_popup(&mut self, popup_id: Id)

pub fn everything_is_visible(&self) -> bool

If true, all windows, menus, tooltips etc are to be visible at once.

This is useful for testing, benchmarking, pre-caching, etc.

Experimental feature!

pub fn set_everything_is_visible(&mut self, value: bool)

If true, all windows, menus, tooltips etc are to be visible at once.

This is useful for testing, benchmarking, pre-caching, etc.

Experimental feature!