Struct egui::Options

pub struct Options {
    pub zoom_factor: f32,
    pub zoom_with_keyboard: bool,
    pub tessellation_options: TessellationOptions,
    pub repaint_on_widget_change: bool,
    pub screen_reader: bool,
    pub preload_font_glyphs: bool,
    pub warn_on_id_clash: bool,
    /* private fields */
}

Some global options that you can read and write.

See also crate::style::DebugOptions.

Fields
zoom_factor: f32

Global zoom factor of the UI.

This is used to calculate the pixels_per_point for the UI as pixels_per_point = zoom_fator * native_pixels_per_point.

The default is 1.0. Make larger to make everything larger.

Please call crate::Context::set_zoom_factor instead of modifying this directly!

zoom_with_keyboard: bool

If true, egui will change the scale of the ui (crate::Context::zoom_factor) when the user presses Cmd+Plus, Cmd+Minus or Cmd+0, just like in a browser.

This is true by default.

tessellation_options: TessellationOptions

Controls the tessellator.

repaint_on_widget_change: bool

If any widget moves or changes id, repaint everything.

It is recommended you keep this OFF, because it is know to cause endless repaints, for unknown reasons (https://github.com/rerun-io/rerun/issues/5018).

screen_reader: bool

This is a signal to any backend that we want the crate::PlatformOutput::events read out loud.

The only change to egui is that labels can be focused by pressing tab.

Screen readers is an experimental feature of egui, and not supported on all platforms.

eframe supports it only on web, but you should consider using AccessKit instead, which eframe supports.

preload_font_glyphs: bool

If true, the most common glyphs (ASCII) are pre-rendered to the texture atlas.

Only the fonts in Style::text_styles will be pre-cached.

This can lead to fewer texture operations, but may use up the texture atlas quicker if you are changing Style::text_styles, of have a lot of text styles.

warn_on_id_clash: bool

Check reusing of Ids, and show a visual warning on screen when one is found.

By default this is true in debug builds.

Implementations

impl Options

pub fn ui(&mut self, ui: &mut Ui)

Show the options in the ui.