Struct egui::Painter

pub struct Painter { /* fields omitted */ }

Helper to paint shapes and text to a specific region on a specific layer.

All coordinates are screen coordinates in the unit points (one point can consist of many physical pixels).

Implementations

impl Painter

pub fn new(ctx: CtxRef, layer_id: LayerId, clip_rect: Rect) -> Self

pub fn with_layer_id(self, layer_id: LayerId) -> Self

pub fn set_layer_id(&mut self, layer_id: LayerId)

redirect

pub fn sub_region(&self, rect: Rect) -> Self

Create a painter for a sub-region of this Painter.

The clip-rect of the returned Painter will be the intersection of the given rectangle and the clip_rect() of this Painter.

impl Painter

Accessors etc

pub fn ctx(&self) -> &CtxRef

Get a reference to the parent CtxRef.

pub fn fonts(&self) -> &Fonts

Available fonts.

pub fn layer_id(&self) -> LayerId

Where we paint

pub fn clip_rect(&self) -> Rect

Everything painted in this Painter will be clipped against this. This means nothing outside of this rectangle will be visible on screen.

pub fn set_clip_rect(&mut self, clip_rect: Rect)

Everything painted in this Painter will be clipped against this. This means nothing outside of this rectangle will be visible on screen.

pub fn round_to_pixel(&self, point: f32) -> f32

Useful for pixel-perfect rendering.

pub fn round_vec_to_pixels(&self, vec: Vec2) -> Vec2

Useful for pixel-perfect rendering.

pub fn round_pos_to_pixels(&self, pos: Pos2) -> Pos2

Useful for pixel-perfect rendering.

impl Painter

Low level

pub fn add(&self, shape: Shape) -> ShapeIdx

It is up to the caller to make sure there is room for this. Can be used for free painting. NOTE: all coordinates are screen coordinates!

pub fn extend(&self, shapes: Vec<Shape>)

Add many shapes at once.

Calling this once is generally faster than calling Self::add multiple times.

pub fn set(&self, idx: ShapeIdx, shape: Shape)

Modify an existing Shape.

impl Painter

Debug painting

pub fn debug_rect(&mut self, rect: Rect, color: Color32, text: impl ToString)

pub fn error(&self, pos: Pos2, text: impl Display) -> Rect

pub fn debug_text(
&self,
pos: Pos2,
anchor: Align2,
color: Color32,
text: impl ToString
) -> Rect

text with a background

impl Painter

Paint different primitives

pub fn line_segment(&self, points: [Pos2; 2], stroke: impl Into<Stroke>)

pub fn circle(
&self,
center: Pos2,
radius: f32,
fill_color: impl Into<Color32>,
stroke: impl Into<Stroke>
)

pub fn circle_filled(
&self,
center: Pos2,
radius: f32,
fill_color: impl Into<Color32>
)

pub fn circle_stroke(
&self,
center: Pos2,
radius: f32,
stroke: impl Into<Stroke>
)

pub fn rect(
&self,
rect: Rect,
corner_radius: f32,
fill_color: impl Into<Color32>,
stroke: impl Into<Stroke>
)

pub fn rect_filled(
&self,
rect: Rect,
corner_radius: f32,
fill_color: impl Into<Color32>
)

pub fn rect_stroke(
&self,
rect: Rect,
corner_radius: f32,
stroke: impl Into<Stroke>
)

pub fn arrow(&self, origin: Pos2, vec: Vec2, stroke: Stroke)

Show an arrow starting at origin and going in the direction of vec, with the length vec.length().

impl Painter

Text

pub fn text(
&self,
pos: Pos2,
anchor: Align2,
text: impl ToString,
text_style: TextStyle,
text_color: Color32
) -> Rect

Lay out and paint some text.

To center the text at the given position, use anchor: (Center, Center).

To find out the size of text before painting it, use Self::layout_no_wrap or Self::layout_multiline.

Returns where the text ended up.

pub fn layout_no_wrap(&self, text_style: TextStyle, text: String) -> Arc<Galley>

Will line break at \n.

Paint the results with Self::galley. Always returns at least one row.

pub fn layout_multiline(
&self,
text_style: TextStyle,
text: String,
max_width_in_points: f32
) -> Arc<Galley>

Will wrap text at the given width and line break at \n.

Paint the results with Self::galley. Always returns at least one row.

pub fn galley(&self, pos: Pos2, galley: Arc<Galley>, color: Color32)

Paint text that has already been layed out in a Galley.

You can create the Galley with Self::layout_no_wrap or Self::layout_multiline.

pub fn galley_with_italics(
&self,
pos: Pos2,
galley: Arc<Galley>,
color: Color32,
fake_italics: bool
)