Struct egui::PlatformOutput

pub struct PlatformOutput {
    pub cursor_icon: CursorIcon,
    pub open_url: Option<OpenUrl>,
    pub copied_text: String,
    pub events: Vec<OutputEvent>,
    pub mutable_text_under_cursor: bool,
    pub ime: Option<IMEOutput>,
    pub accesskit_update: Option<TreeUpdate>,
}

The non-rendering part of what egui emits each frame.

You can access (and modify) this with crate::Context::output.

The backend should use this.

Fields
cursor_icon: CursorIcon

Set the cursor to this icon.

open_url: Option<OpenUrl>

If set, open this url.

copied_text: String

If set, put this text in the system clipboard. Ignore if empty.

This is often a response to crate::Event::Copy or crate::Event::Cut.

if ui.button("📋").clicked() {
    ui.output_mut(|o| o.copied_text = "some_text".to_string());
}

events: Vec<OutputEvent>

Events that may be useful to e.g. a screen reader.

mutable_text_under_cursor: bool

Is there a mutable TextEdit under the cursor? Use by eframe web to show/hide mobile keyboard and IME agent.

ime: Option<IMEOutput>

This is et if, and only if, the user is currently editing text.

Useful for IME.

accesskit_update: Option<TreeUpdate>

The difference in the widget tree since last frame.

NOTE: this needs to be per-viewport.

Implementations

impl PlatformOutput

pub fn events_description(&self) -> String

This can be used by a text-to-speech system to describe the events (if any).

pub fn append(&mut self, newer: Self)

Add on new output.

pub fn take(&mut self) -> Self

Take everything ephemeral (everything except cursor_icon currently)