egui
Struct PointerState

pub struct PointerState { /* private fields */ }

Mouse or touch state.

Implementations

impl PointerState

pub fn delta(&self) -> Vec2

How much the pointer moved compared to last frame, in points.

pub fn motion(&self) -> Option<Vec2>

How much the mouse moved since the last frame, in unspecified units. Represents the actual movement of the mouse, without acceleration or clamped by screen edges. May be unavailable on some integrations.

pub fn velocity(&self) -> Vec2

Current velocity of pointer.

This is smoothed over a few frames, but can be ZERO when frame-rate is bad.

pub fn direction(&self) -> Vec2

Current direction of the pointer.

This is less sensitive to bad framerate than Self::velocity.

pub fn press_origin(&self) -> Option<Pos2>

Where did the current click/drag originate? None if no mouse button is down.

pub fn press_start_time(&self) -> Option<f64>

When did the current click/drag originate? None if no mouse button is down.

pub fn latest_pos(&self) -> Option<Pos2>

Latest reported pointer position. When tapping a touch screen, this will be None.

pub fn hover_pos(&self) -> Option<Pos2>

If it is a good idea to show a tooltip, where is pointer?

pub fn interact_pos(&self) -> Option<Pos2>

If you detect a click or drag and wants to know where it happened, use this.

Latest position of the mouse, but ignoring any Event::PointerGone if there were interactions this frame. When tapping a touch screen, this will be the location of the touch.

pub fn has_pointer(&self) -> bool

Do we have a pointer?

false if the mouse is not over the egui area, or if no touches are down on touch screens.

pub fn is_still(&self) -> bool

Is the pointer currently still? This is smoothed so a few frames of stillness is required before this returns true.

pub fn is_moving(&self) -> bool

Is the pointer currently moving? This is smoothed so a few frames of stillness is required before this returns false.

pub fn time_since_last_movement(&self) -> f32

How long has it been (in seconds) since the pointer was last moved?

pub fn time_since_last_click(&self) -> f32

How long has it been (in seconds) since the pointer was clicked?

pub fn any_pressed(&self) -> bool

Was any pointer button pressed (!down -> down) this frame?

This can sometimes return true even if any_down() == false because a press can be shorted than one frame.

pub fn any_released(&self) -> bool

Was any pointer button released (down -> !down) this frame?

pub fn button_pressed(&self, button: PointerButton) -> bool

Was the button given pressed this frame?

pub fn button_released(&self, button: PointerButton) -> bool

Was the button given released this frame?

pub fn primary_pressed(&self) -> bool

Was the primary button pressed this frame?

pub fn secondary_pressed(&self) -> bool

Was the secondary button pressed this frame?

pub fn primary_released(&self) -> bool

Was the primary button released this frame?

pub fn secondary_released(&self) -> bool

Was the secondary button released this frame?

pub fn any_down(&self) -> bool

Is any pointer button currently down?

pub fn any_click(&self) -> bool

Were there any type of click this frame?

pub fn button_clicked(&self, button: PointerButton) -> bool

Was the given pointer button given clicked this frame?

Returns true on double- and triple- clicks too.

pub fn button_double_clicked(&self, button: PointerButton) -> bool

Was the button given double clicked this frame?

pub fn button_triple_clicked(&self, button: PointerButton) -> bool

Was the button given triple clicked this frame?

pub fn primary_clicked(&self) -> bool

Was the primary button clicked this frame?

pub fn secondary_clicked(&self) -> bool

Was the secondary button clicked this frame?

pub fn button_down(&self, button: PointerButton) -> bool

Is this button currently down?

pub fn could_any_button_be_click(&self) -> bool

If the pointer button is down, will it register as a click when released?

See also Self::is_decidedly_dragging.

pub fn is_decidedly_dragging(&self) -> bool

Just because the mouse is down doesnโt mean we are dragging. We could be at the start of a click. But if the mouse is down long enough, or has moved far enough, then we consider it a drag.

This function can return true on the same frame the drag is released, but NOT on the first frame it was started.

See also Self::could_any_button_be_click.

pub fn primary_down(&self) -> bool

Is the primary button currently down?

pub fn secondary_down(&self) -> bool

Is the secondary button currently down?

pub fn middle_down(&self) -> bool

Is the middle button currently down?

impl PointerState

pub fn ui(&self, ui: &mut Ui)