Struct egui::PointerState

pub struct PointerState { /* fields omitted */ }

Mouse or touch state.

Implementations

impl PointerState

pub fn delta(&self) -> Vec2

How much the pointer moved compared to last frame, in points.

pub fn velocity(&self) -> Vec2

Current velocity of pointer.

pub fn press_origin(&self) -> Option<Pos2>

Where did the current click/drag originate? None if no mouse button is down.

pub fn hover_pos(&self) -> Option<Pos2>

If it is a good idea to show a tooltip, where is pointer?

pub fn interact_pos(&self) -> Option<Pos2>

If you detect a click or drag and wants to know where it happened, use this.

Latest position of the mouse, but ignoring any Event::PointerGone if there were interactions this frame. When tapping a touch screen, this will be the location of the touch.

pub fn has_pointer(&self) -> bool

Do we have a pointer?

false if the mouse is not over the egui area, or if no touches are down on touch screens.

pub fn is_still(&self) -> bool

Is the pointer currently still? This is smoothed so a few frames of stillness is required before this returns true.

pub fn is_moving(&self) -> bool

Is the pointer currently moving? This is smoothed so a few frames of stillness is required before this returns false.

pub fn any_pressed(&self) -> bool

Was any pointer button pressed (!down -> down) this frame? This can sometimes return true even if any_down() == false because a press can be shorted than one frame.

pub fn any_released(&self) -> bool

Was any pointer button released (down -> !down) this frame?

pub fn any_down(&self) -> bool

Is any pointer button currently down?

pub fn any_click(&self) -> bool

Were there any type of click this frame?

pub fn button_down(&self, button: PointerButton) -> bool

Is this button currently down?

impl PointerState

pub fn ui(&self, ui: &mut Ui)