egui
Struct Rect

#[repr(C)]
pub struct Rect {
    pub min: Pos2,
    pub max: Pos2,
}

A rectangular region of space.

Usually a Rect has a positive (or zero) size, and then Self::min <= Self::max. In these cases Self::min is the left-top corner and Self::max is the right-bottom corner.

A rectangle is allowed to have a negative size, which happens when the order of min and max are swapped. These are usually a sign of an error.

Normally the unit is points (logical pixels) in screen space coordinates.

Rect does NOT implement Default, because there is no obvious default value. Rect::ZERO may seem reasonable, but when used as a bounding box, Rect::NOTHING is a better default - so be explicit instead!

Fields
min: Pos2

One of the corners of the rectangle, usually the left top one.

max: Pos2

The other corner, opposing Self::min. Usually the right bottom one.

Implementations

impl Rect

pub const EVERYTHING: Rect = _

Infinite rectangle that contains every point.

pub const NOTHING: Rect = _

The inverse of Self::EVERYTHING: stretches from positive infinity to negative infinity. Contains no points.

This is useful as the seed for bounding boxes.

Example:

let mut rect = Rect::NOTHING;
assert!(rect.size() == Vec2::splat(-f32::INFINITY));
assert!(rect.contains(pos2(0.0, 0.0)) == false);
rect.extend_with(pos2(2.0, 1.0));
rect.extend_with(pos2(0.0, 3.0));
assert_eq!(rect, Rect::from_min_max(pos2(0.0, 1.0), pos2(2.0, 3.0)))

pub const NAN: Rect = _

An invalid Rect filled with f32::NAN.

pub const ZERO: Rect = _

A Rect filled with zeroes.

pub const fn from_min_max(min: Pos2, max: Pos2) -> Rect

pub fn from_min_size(min: Pos2, size: Vec2) -> Rect

left-top corner plus a size (stretching right-down).

pub fn from_center_size(center: Pos2, size: Vec2) -> Rect

pub fn from_x_y_ranges( x_range: impl Into<Rangef>, y_range: impl Into<Rangef>, ) -> Rect

pub fn from_two_pos(a: Pos2, b: Pos2) -> Rect

Returns the bounding rectangle of the two points.

pub fn from_pos(point: Pos2) -> Rect

A zero-sized rect at a specific point.

pub fn from_points(points: &[Pos2]) -> Rect

Bounding-box around the points.

pub fn everything_right_of(left_x: f32) -> Rect

A Rect that contains every point to the right of the given X coordinate.

pub fn everything_left_of(right_x: f32) -> Rect

A Rect that contains every point to the left of the given X coordinate.

pub fn everything_below(top_y: f32) -> Rect

A Rect that contains every point below a certain y coordinate

pub fn everything_above(bottom_y: f32) -> Rect

A Rect that contains every point above a certain y coordinate

pub fn with_min_x(self, min_x: f32) -> Rect

pub fn with_min_y(self, min_y: f32) -> Rect

pub fn with_max_x(self, max_x: f32) -> Rect

pub fn with_max_y(self, max_y: f32) -> Rect

pub fn expand(self, amnt: f32) -> Rect

Expand by this much in each direction, keeping the center

pub fn expand2(self, amnt: Vec2) -> Rect

Expand by this much in each direction, keeping the center

pub fn scale_from_center(self, scale_factor: f32) -> Rect

Scale up by this factor in each direction, keeping the center

pub fn scale_from_center2(self, scale_factor: Vec2) -> Rect

Scale up by this factor in each direction, keeping the center

pub fn shrink(self, amnt: f32) -> Rect

Shrink by this much in each direction, keeping the center

pub fn shrink2(self, amnt: Vec2) -> Rect

Shrink by this much in each direction, keeping the center

pub fn translate(self, amnt: Vec2) -> Rect

pub fn rotate_bb(self, rot: Rot2) -> Rect

Rotate the bounds (will expand the Rect)

pub fn intersects(self, other: Rect) -> bool

pub fn set_width(&mut self, w: f32)

keep min

pub fn set_height(&mut self, h: f32)

keep min

pub fn set_center(&mut self, center: Pos2)

Keep size

pub fn contains(&self, p: Pos2) -> bool

pub fn contains_rect(&self, other: Rect) -> bool

pub fn clamp(&self, p: Pos2) -> Pos2

Return the given points clamped to be inside the rectangle Panics if Self::is_negative.

pub fn extend_with(&mut self, p: Pos2)

pub fn extend_with_x(&mut self, x: f32)

Expand to include the given x coordinate

pub fn extend_with_y(&mut self, y: f32)

Expand to include the given y coordinate

pub fn union(self, other: Rect) -> Rect

The union of two bounding rectangle, i.e. the minimum Rect that contains both input rectangles.

pub fn intersect(self, other: Rect) -> Rect

The intersection of two Rect, i.e. the area covered by both.

pub fn center(&self) -> Pos2

pub fn size(&self) -> Vec2

rect.size() == Vec2 { x: rect.width(), y: rect.height() }

pub fn width(&self) -> f32

pub fn height(&self) -> f32

pub fn aspect_ratio(&self) -> f32

Width / height

aspect_ratio < 1: portrait / high

aspect_ratio = 1: square

aspect_ratio > 1: landscape / wide

pub fn square_proportions(&self) -> Vec2

[2, 1] for wide screen, and [1, 2] for portrait, etc. At least one dimension = 1, the other >= 1 Returns the proportions required to letter-box a square view area.

pub fn area(&self) -> f32

pub fn distance_to_pos(&self, pos: Pos2) -> f32

The distance from the rect to the position.

The distance is zero when the position is in the interior of the rectangle.

Negative rectangles always return f32::INFINITY.

pub fn distance_sq_to_pos(&self, pos: Pos2) -> f32

The distance from the rect to the position, squared.

The distance is zero when the position is in the interior of the rectangle.

Negative rectangles always return f32::INFINITY.

pub fn signed_distance_to_pos(&self, pos: Pos2) -> f32

Signed distance to the edge of the box.

Negative inside the box.

Negative rectangles always return f32::INFINITY.

let rect = Rect::from_min_max(pos2(0.0, 0.0), pos2(1.0, 1.0));
assert_eq!(rect.signed_distance_to_pos(pos2(0.50, 0.50)), -0.50);
assert_eq!(rect.signed_distance_to_pos(pos2(0.75, 0.50)), -0.25);
assert_eq!(rect.signed_distance_to_pos(pos2(1.50, 0.50)), 0.50);

pub fn lerp_inside(&self, t: Vec2) -> Pos2

Linearly interpolate so that [0, 0] is Self::min and [1, 1] is Self::max.

pub fn lerp_towards(&self, other: &Rect, t: f32) -> Rect

Linearly self towards other rect.

pub fn x_range(&self) -> Rangef

pub fn y_range(&self) -> Rangef

pub fn bottom_up_range(&self) -> Rangef

pub fn is_negative(&self) -> bool

width < 0 || height < 0

pub fn is_positive(&self) -> bool

width > 0 && height > 0

pub fn is_finite(&self) -> bool

True if all members are also finite.

pub fn any_nan(self) -> bool

True if any member is NaN.

impl Rect

Convenience functions (assumes origin is towards left top):

pub fn left(&self) -> f32

min.x

pub fn left_mut(&mut self) -> &mut f32

min.x

pub fn set_left(&mut self, x: f32)

min.x

pub fn right(&self) -> f32

max.x

pub fn right_mut(&mut self) -> &mut f32

max.x

pub fn set_right(&mut self, x: f32)

max.x

pub fn top(&self) -> f32

min.y

pub fn top_mut(&mut self) -> &mut f32

min.y

pub fn set_top(&mut self, y: f32)

min.y

pub fn bottom(&self) -> f32

max.y

pub fn bottom_mut(&mut self) -> &mut f32

max.y

pub fn set_bottom(&mut self, y: f32)

max.y

pub fn left_top(&self) -> Pos2

pub fn center_top(&self) -> Pos2

pub fn right_top(&self) -> Pos2

pub fn left_center(&self) -> Pos2

pub fn right_center(&self) -> Pos2

pub fn left_bottom(&self) -> Pos2

pub fn center_bottom(&self) -> Pos2

pub fn right_bottom(&self) -> Pos2

pub fn split_left_right_at_fraction(&self, t: f32) -> (Rect, Rect)

Split rectangle in left and right halves. t is expected to be in the (0,1) range.

pub fn split_left_right_at_x(&self, split_x: f32) -> (Rect, Rect)

Split rectangle in left and right halves at the given x coordinate.

pub fn split_top_bottom_at_fraction(&self, t: f32) -> (Rect, Rect)

Split rectangle in top and bottom halves. t is expected to be in the (0,1) range.

pub fn split_top_bottom_at_y(&self, split_y: f32) -> (Rect, Rect)

Split rectangle in top and bottom halves at the given y coordinate.

impl Rect

pub fn intersects_ray(&self, o: Pos2, d: Vec2) -> bool

Does this Rect intersect the given ray (where d is normalized)?

A ray that starts inside the rect will return true.

pub fn intersects_ray_from_center(&self, d: Vec2) -> Pos2

Where does a ray from the center intersect the rectangle?

d is the direction of the ray and assumed to be normalized.