Struct egui::Response

pub struct Response {
    pub ctx: CtxRef,
    pub layer_id: LayerId,
    pub id: Id,
    pub rect: Rect,
    pub sense: Sense,
    // some fields omitted
}

The result of adding a widget to a Ui.

A Response lets you know whether or not a widget is being hovered, clicked or dragged. It also lets you easily show a tooltip on hover.

Whenever something gets added to a Ui, a Response object is returned. [ui.add] returns a Response, as does [ui.button], and all similar shortcuts.

Fields
ctx: CtxRef

Used for optionally showing a tooltip and checking for more interactions.

layer_id: LayerId

Which layer the widget is part of.

id: Id

The Id of the widget/area this response pertains.

rect: Rect

The area of the screen we are talking about.

sense: Sense

The senses (click and/or drag) that the widget was interested in (if any).

Implementations

impl Response

pub fn clicked(&self) -> bool

Returns true if this widget was clicked this frame by the primary button.

pub fn clicked_by(&self, button: PointerButton) -> bool

Returns true if this widget was clicked this frame by the given button.

pub fn secondary_clicked(&self) -> bool

Returns true if this widget was clicked this frame by the secondary mouse button (e.g. the right mouse button).

pub fn middle_clicked(&self) -> bool

Returns true if this widget was clicked this frame by the middle mouse button.

pub fn double_clicked(&self) -> bool

Returns true if this widget was double-clicked this frame by the primary button.

pub fn double_clicked_by(&self, button: PointerButton) -> bool

Returns true if this widget was double-clicked this frame by the given button.

pub fn clicked_elsewhere(&self) -> bool

true if there was a click outside this widget this frame.

pub fn enabled(&self) -> bool

Was the widget enabled? If false, there was no interaction attempted and the widget should be drawn in a gray disabled look.

pub fn hovered(&self) -> bool

The pointer is hovering above this widget or the widget was clicked/tapped this frame.

pub fn has_focus(&self) -> bool

This widget has the keyboard focus (i.e. is receiving key presses).

pub fn gained_focus(&self) -> bool

True if this widget has keyboard focus this frame, but didn't last frame.

pub fn lost_focus(&self) -> bool

The widget had keyboard focus and lost it, either because the user pressed tab or clicked somewhere else, or (in case of a crate::TextEdit) because the user pressed enter.

let response = ui.text_edit_singleline(&mut my_text);
if response.lost_focus() && ui.input().key_pressed(egui::Key::Enter) {
    do_request(&my_text);
}

pub fn lost_kb_focus(&self) -> bool

👎 Deprecated:
Renamed to lost_focus()

pub fn dragged(&self) -> bool

The widgets is being dragged.

To find out which button(s), query crate::PointerState::button_down (ui.input().pointer.button_down(…)).

pub fn dragged_by(&self, button: PointerButton) -> bool

pub fn drag_started(&self) -> bool

Did a drag on this widgets begin this frame?

pub fn drag_released(&self) -> bool

The widget was being dragged, but now it has been released.

pub fn drag_delta(&self) -> Vec2

If dragged, how many points were we dragged and in what direction?

pub fn interact_pointer_pos(&self) -> Option<Pos2>

Where the pointer (mouse/touch) were when when this widget was clicked or dragged. None if the widget is not being interacted with.

pub fn hover_pos(&self) -> Option<Pos2>

If it is a good idea to show a tooltip, where is pointer? None if the pointer is outside the response area.

pub fn is_pointer_button_down_on(&self) -> bool

Is the pointer button currently down on this widget? This is true if the pointer is pressing down or dragging a widget

pub fn changed(&self) -> bool

What the underlying data changed?

e.g. the slider was dragged, text was entered in a TextEdit etc. Always false for something like a Button. Can sometimes be true even though the data didn't changed (e.g. if the user entered a character and erased it the same frame).

pub fn mark_changed(&mut self)

Report the data shown by this widget changed.

This must be called by widgets that represent some mutable data, e.g. checkboxes, sliders etc.

pub fn on_hover_ui(self, add_contents: impl FnOnce(&mut Ui)) -> Self

Show this UI if the item was hovered (i.e. a tooltip). If you call this multiple times the tooltips will stack underneath the previous ones.

pub fn on_hover_ui_at_pointer(self, add_contents: impl FnOnce(&mut Ui)) -> Self

Like on_hover_ui, but show the ui next to cursor.

pub fn on_hover_text(self, text: impl Into<String>) -> Self

Show this text if the item was hovered (i.e. a tooltip). If you call this multiple times the tooltips will stack underneath the previous ones.

pub fn tooltip_text(self, text: impl Into<String>) -> Self

👎 Deprecated:
Deprecated 2020-10-01: use on_hover_text instead.

pub fn on_hover_cursor(self, cursor: CursorIcon) -> Self

When hovered, use this icon for the mouse cursor.

pub fn interact(&self, sense: Sense) -> Self

Check for more interactions (e.g. sense clicks on a Response returned from a label).

let response = ui.label("hello");
assert!(!response.clicked()); // labels don't sense clicks
let response = response.interact(egui::Sense::click());
if response.clicked() { /* … */ }

pub fn scroll_to_me(&self, align: Align)

Move the scroll to this UI with the specified alignment.

egui::ScrollArea::auto_sized().show(ui, |ui| {
    for i in 0..1000 {
        let response = ui.button(format!("Button {}", i));
        if response.clicked() {
            response.scroll_to_me(Align::Center);
        }
    }
});

pub fn widget_info(&self, make_info: impl Fn() -> WidgetInfo)

For accessibility.

Call after interacting and potential calls to Self::mark_changed.

impl Response

pub fn union(&self, other: Self) -> Self

A logical "or" operation. For instance a.union(b).hovered means "was either a or b hovered?".