egui
Struct UiStack

pub struct UiStack {
    pub id: Id,
    pub info: UiStackInfo,
    pub layout_direction: Direction,
    pub min_rect: Rect,
    pub max_rect: Rect,
    pub parent: Option<Arc<UiStack>>,
}

Information about a crate::Ui and its parents.

UiStack serves to keep track of the current hierarchy of crate::Uis, such that nested widgets or user code may adapt to the surrounding context or obtain layout information from a crate::Ui that might be several steps higher in the hierarchy.

Note: since UiStack contains a reference to its parent, it is both a stack, and a node within that stack. Most of its methods are about the specific node, but some methods walk up the hierarchy to provide information about the entire stack.

Fields
id: Idinfo: UiStackInfolayout_direction: Directionmin_rect: Rectmax_rect: Rectparent: Option<Arc<UiStack>>
Implementations

impl UiStack

pub fn kind(&self) -> Option<UiKind>

pub fn frame(&self) -> &Frame

pub fn tags(&self) -> &UiTags

User tags.

pub fn is_panel_ui(&self) -> bool

Is this crate::Ui a panel?

pub fn is_area_ui(&self) -> bool

Is this crate::Ui an crate::Area?

pub fn is_root_ui(&self) -> bool

Is this a root crate::Ui, i.e. created with crate::Ui::new()?

pub fn has_visible_frame(&self) -> bool

This this crate::Ui a crate::Frame with a visible stroke?

impl UiStack

pub fn iter(&self) -> UiStackIterator<'_>

Return an iterator that walks the stack from this node to the root.

pub fn contained_in(&self, kind: UiKind) -> bool

Check if this node is or is contained in a crate::Ui of a specific kind.