Module egui::style

egui theme (spacing, colors, etc).

Structs

DebugOptions

Options for help debug egui by adding extra visualization

Interaction

How and when interaction happens.

Selection

Selected text, selected elements etc

Spacing

Controls the sizes and distances between widgets.

Style

Specifies the look and feel of egui.

Visuals

Controls the visual style (colors etc) of egui.

WidgetVisuals

bg = background, fg = foreground.

Widgets

The visuals of widgets for different states of interaction.