egui::style
Struct Spacing

pub struct Spacing {

    pub item_spacing: Vec2,
    pub window_margin: Margin,
    pub button_padding: Vec2,
    pub menu_margin: Margin,
    pub indent: f32,
    pub interact_size: Vec2,
    pub slider_width: f32,
    pub slider_rail_height: f32,
    pub combo_width: f32,
    pub text_edit_width: f32,
    pub icon_width: f32,
    pub icon_width_inner: f32,
    pub icon_spacing: f32,
    pub default_area_size: Vec2,
    pub tooltip_width: f32,
    pub menu_width: f32,
    pub menu_spacing: f32,
    pub indent_ends_with_horizontal_line: bool,
    pub combo_height: f32,
    pub scroll: ScrollStyle,

}

Controls the sizes and distances between widgets.

Fields
item_spacing: Vec2

Horizontal and vertical spacing between widgets.

To add extra space between widgets, use Ui::add_space.

item_spacing is inserted after adding a widget, so to increase the spacing between widgets A and B you need to change item_spacing before adding A.

window_margin: Margin

Horizontal and vertical margins within a window frame.

button_padding: Vec2

Button size is text size plus this on each side

menu_margin: Margin

Horizontal and vertical margins within a menu frame.

indent: f32

Indent collapsing regions etc by this much.

interact_size: Vec2

Minimum size of a DragValue, color picker button, and other small widgets. interact_size.y is the default height of button, slider, etc. Anything clickable should be (at least) this size.

slider_width: f32

Default width of a Slider.

slider_rail_height: f32

Default rail height of a Slider.

combo_width: f32

Default (minimum) width of a ComboBox.

text_edit_width: f32

Default width of a crate::TextEdit.

icon_width: f32

Checkboxes, radio button and collapsing headers have an icon at the start. This is the width/height of the outer part of this icon (e.g. the BOX of the checkbox).

icon_width_inner: f32

Checkboxes, radio button and collapsing headers have an icon at the start. This is the width/height of the inner part of this icon (e.g. the check of the checkbox).

icon_spacing: f32

Checkboxes, radio button and collapsing headers have an icon at the start. This is the spacing between the icon and the text

default_area_size: Vec2

The size used for the Ui::max_rect the first frame.

Text will wrap at this width, and images that expand to fill the available space will expand to this size.

If the contents are smaller than this size, the area will shrink to fit the contents. If the contents overflow, the area will grow.

tooltip_width: f32

Width of a tooltip (on_hover_ui, on_hover_text etc).

menu_width: f32

The default wrapping width of a menu.

Items longer than this will wrap to a new line.

menu_spacing: f32

Horizontal distance between a menu and a submenu.

indent_ends_with_horizontal_line: bool

End indented regions with a horizontal line

combo_height: f32

Height of a combo-box before showing scroll bars.

scroll: ScrollStyle

Controls the spacing of a crate::ScrollArea.

Implementations

impl Spacing

pub fn icon_rectangles(&self, rect: Rect) -> (Rect, Rect)

Returns small icon rectangle and big icon rectangle

impl Spacing

pub fn ui(&mut self, ui: &mut Ui)