egui::style
Struct Visuals

pub struct Visuals {

    pub dark_mode: bool,
    pub override_text_color: Option<Color32>,
    pub widgets: Widgets,
    pub selection: Selection,
    pub hyperlink_color: Color32,
    pub faint_bg_color: Color32,
    pub extreme_bg_color: Color32,
    pub code_bg_color: Color32,
    pub warn_fg_color: Color32,
    pub error_fg_color: Color32,
    pub window_rounding: Rounding,
    pub window_shadow: Shadow,
    pub window_fill: Color32,
    pub window_stroke: Stroke,
    pub window_highlight_topmost: bool,
    pub menu_rounding: Rounding,
    pub panel_fill: Color32,
    pub popup_shadow: Shadow,
    pub resize_corner_size: f32,
    pub text_cursor: TextCursorStyle,
    pub clip_rect_margin: f32,
    pub button_frame: bool,
    pub collapsing_header_frame: bool,
    pub indent_has_left_vline: bool,
    pub striped: bool,
    pub slider_trailing_fill: bool,
    pub handle_shape: HandleShape,
    pub interact_cursor: Option<CursorIcon>,
    pub image_loading_spinners: bool,
    pub numeric_color_space: NumericColorSpace,

}

Controls the visual style (colors etc) of egui.

You can change the visuals of a Ui with Ui::visuals_mut and of everything with crate::Context::set_visuals_of.

If you want to change fonts, use crate::Context::set_fonts instead.

Fields
dark_mode: bool

If true, the visuals are overall dark with light text. If false, the visuals are overall light with dark text.

NOTE: setting this does very little by itself, this is more to provide a convenient summary of the rest of the settings.

override_text_color: Option<Color32>

Override default text color for all text.

This is great for setting the color of text for any widget.

If text_color is None (default), then the text color will be the same as the foreground stroke color (WidgetVisuals::fg_stroke) and will depend on whether or not the widget is being interacted with.

In the future we may instead modulate the text_color based on whether or not it is interacted with so that visuals.text_color is always used, but its alpha may be different based on whether or not it is disabled, non-interactive, hovered etc.

widgets: Widgets

Visual styles of widgets

selection: Selectionhyperlink_color: Color32

The color used for crate::Hyperlink,

faint_bg_color: Color32

Something just barely different from the background color. Used for crate::Grid::striped.

extreme_bg_color: Color32

Very dark or light color (for corresponding theme). Used as the background of text edits, scroll bars and others things that needs to look different from other interactive stuff.

code_bg_color: Color32

Background color behind code-styled monospaced labels.

warn_fg_color: Color32

A good color for warning text (e.g. orange).

error_fg_color: Color32

A good color for error text (e.g. red).

window_rounding: Roundingwindow_shadow: Shadowwindow_fill: Color32window_stroke: Strokewindow_highlight_topmost: bool

Highlight the topmost window.

menu_rounding: Roundingpanel_fill: Color32

Panel background color

popup_shadow: Shadowresize_corner_size: f32text_cursor: TextCursorStyle

How the text cursor acts.

clip_rect_margin: f32

Allow child widgets to be just on the border and still have a stroke with some thickness

button_frame: bool

Show a background behind buttons.

collapsing_header_frame: bool

Show a background behind collapsing headers.

indent_has_left_vline: bool

Draw a vertical lien left of indented region, in e.g. crate::CollapsingHeader.

striped: bool

Whether or not Grids and Tables should be striped by default (have alternating rows differently colored).

slider_trailing_fill: bool

Show trailing color behind the circle of a Slider. Default is OFF.

Enabling this will affect ALL sliders, and can be enabled/disabled per slider with Slider::trailing_fill.

handle_shape: HandleShape

Shape of the handle for sliders and similar widgets.

Changing this will affect ALL sliders, and can be enabled/disabled per slider with Slider::handle_shape.

interact_cursor: Option<CursorIcon>

Should the cursor change when the user hovers over an interactive/clickable item?

This is consistent with a lot of browser-based applications (vscode, github all turn your cursor into CursorIcon::PointingHand when a button is hovered) but it is inconsistent with native UI toolkits.

image_loading_spinners: bool

Show a spinner when loading an image.

numeric_color_space: NumericColorSpace

How to display numeric color values.

Implementations

impl Visuals

pub fn noninteractive(&self) -> &WidgetVisuals

pub fn text_color(&self) -> Color32

pub fn weak_text_color(&self) -> Color32

pub fn strong_text_color(&self) -> Color32

pub fn window_fill(&self) -> Color32

Window background color.

pub fn window_stroke(&self) -> Stroke

pub fn fade_out_to_color(&self) -> Color32

When fading out things, we fade the colors towards this.

pub fn gray_out(&self, color: Color32) -> Color32

Returned a โgrayed outโ version of the given color.

impl Visuals

pub fn dark() -> Self

Default dark theme.

pub fn light() -> Self

Default light theme.

impl Visuals

pub fn ui(&mut self, ui: &mut Ui)