Struct egui::style::Visuals

pub struct Visuals {

    pub dark_mode: bool,
    pub override_text_color: Option<Color32>,
    pub widgets: Widgets,
    pub selection: Selection,
    pub hyperlink_color: Color32,
    pub faint_bg_color: Color32,
    pub extreme_bg_color: Color32,
    pub code_bg_color: Color32,
    pub window_corner_radius: f32,
    pub window_shadow: Shadow,
    pub popup_shadow: Shadow,
    pub resize_corner_size: f32,
    pub text_cursor_width: f32,
    pub text_cursor_preview: bool,
    pub clip_rect_margin: f32,
    pub button_frame: bool,
    pub collapsing_header_frame: bool,

}

Controls the visual style (colors etc) of egui.

You can change the visuals of a Ui with Ui::visuals_mut and of everything with crate::Context::set_visuals.

If you want to change fonts, use crate::Context::set_fonts instead.

Fields
dark_mode: bool

If true, the visuals are overall dark with light text. If false, the visuals are overall light with dark text.

NOTE: setting this does very little by itself, this is more to provide a convenient summary of the rest of the settings.

override_text_color: Option<Color32>

Override default text color for all text.

This is great for setting the color of text for any widget.

If text_color is None (default), then the text color will be the same as the foreground stroke color (WidgetVisuals::fg_stroke) and will depend on wether or not the widget is being interacted with.

In the future we may instead modulate the text_color based on wether or not it is interacted with so that visuals.text_color is always used, but its alpha may be different based on whether or not it is disabled, non-interactive, hovered etc.

widgets: Widgets

Visual styles of widgets

selection: Selectionhyperlink_color: Color32

The color used for Hyperlink,

faint_bg_color: Color32

Something just barely different from the background color. Used for crate::Grid::striped.

extreme_bg_color: Color32

Very dark or light color (for corresponding theme). Used as the background of text edits, scroll bars and others things that needs to look different from other interactive stuff.

code_bg_color: Color32

Background color behind code-styled monospaced labels.

window_corner_radius: f32window_shadow: Shadowpopup_shadow: Shadowresize_corner_size: f32text_cursor_width: f32text_cursor_preview: bool

show where the text cursor would be if you clicked

clip_rect_margin: f32

Allow child widgets to be just on the border and still have a stroke with some thickness

button_frame: bool

Show a background behind buttons.

collapsing_header_frame: bool

Show a background behind collapsing headers.

Implementations

impl Visuals

pub fn noninteractive(&self) -> &WidgetVisuals

pub fn text_color(&self) -> Color32

pub fn weak_text_color(&self) -> Color32

pub fn strong_text_color(&self) -> Color32

pub fn window_fill(&self) -> Color32

pub fn window_stroke(&self) -> Stroke

impl Visuals

pub fn dark() -> Self

Default dark theme.

pub fn light() -> Self

Default light theme.

impl Visuals

pub fn light_dark_radio_buttons(&mut self, ui: &mut Ui)

Show radio-buttons to switch between light and dark mode.

pub fn light_dark_small_toggle_button(&self, ui: &mut Ui) -> Option<Self>

Show small toggle-button for light and dark mode.

pub fn ui(&mut self, ui: &mut Ui)