egui

Module text

Re-exports

pub use crate::text_selection::CCursorRange;

Structs

CCursor

Character cursor.

FontData

A .ttf or .otf file and a font face index.

FontDefinitions

Describes the font data and the sizes to use.

Fonts

The collection of fonts used by epaint.

Galley

Text that has been laid out, ready for painting.

LayoutJob

Describes the task of laying out text.

LayoutSection

TextFormat

Formatting option for a section of text.

TextWrapping

Controls the text wrapping and elision of a LayoutJob.

Enums

FontFamily

Font of unknown size.

Constants

TAB_SIZE

One \t character is this many spaces wide.