Struct egui::text::Fonts

pub struct Fonts { /* private fields */ }

The collection of fonts used by epaint.

Required in order to paint text.

Implementations

impl Fonts

pub fn new(pixels_per_point: f32, definitions: FontDefinitions) -> Fonts

Create a new Fonts for text layout. This call is expensive, so only create on Fonts and then reuse it.

pub fn pixels_per_point(&self) -> f32

pub fn definitions(&self) -> &FontDefinitions

pub fn round_to_pixel(&self, point: f32) -> f32

pub fn floor_to_pixel(&self, point: f32) -> f32

pub fn font_image(&self) -> Arc<FontImage>

Call each frame to get the latest available font texture data.

pub fn glyph_width(&self, text_style: TextStyle, c: char) -> f32

Width of this character in points.

pub fn row_height(&self, text_style: TextStyle) -> f32

Height of one row of text. In points

pub fn layout_job(&self, job: LayoutJob) -> Arc<Galley>

Layout some text. This is the most advanced layout function. See also Self::layout, Self::layout_no_wrap and Self::layout_delayed_color.

The implementation uses memoization so repeated calls are cheap.

pub fn layout(
&self,
text: String,
text_style: TextStyle,
color: Color32,
wrap_width: f32
) -> Arc<Galley>

Will wrap text at the given width and line break at \n.

The implementation uses memoization so repeated calls are cheap.

pub fn layout_no_wrap(
&self,
text: String,
text_style: TextStyle,
color: Color32
) -> Arc<Galley>

Will line break at \n.

The implementation uses memoization so repeated calls are cheap.

pub fn layout_delayed_color(
&self,
text: String,
text_style: TextStyle,
wrap_width: f32
) -> Arc<Galley>

Like Self::layout, made for when you want to pick a color for the text later.

The implementation uses memoization so repeated calls are cheap.

pub fn num_galleys_in_cache(&self) -> usize

pub fn end_frame(&self)

Must be called once per frame to clear the Galley cache.