Struct egui::text::Galley

pub struct Galley {
    pub job: Arc<LayoutJob>,
    pub rows: Vec<Row>,
    pub elided: bool,
    pub rect: Rect,
    pub mesh_bounds: Rect,
    pub num_vertices: usize,
    pub num_indices: usize,
    pub pixels_per_point: f32,
}

Text that has been laid out, ready for painting.

You can create a Galley using crate::Fonts::layout_job;

Needs to be recreated if the underlying font atlas texture changes, which happens under the following conditions:

pixels_per_point or max_texture_size change. These parameters are set in crate::text::Fonts::begin_frame. When using egui they are set from egui::InputState and can change at any time.

The atlas has become full. This can happen any time a new glyph is added to the atlas, which in turn can happen any time new text is laid out.

Fields
job: Arc<LayoutJob>

The job that this galley is the result of. Contains the original string and style sections.

rows: Vec<Row>

Rows of text, from top to bottom.

The number of characters in all rows sum up to job.text.chars().count() unless Self::elided is true.

Note that a paragraph (a piece of text separated with \n) can be split up into multiple rows.

elided: bool

Set to true the text was truncated due to TextWrapping::max_rows.

rect: Rect

Bounding rect.

rect.top() is always 0.0.

With LayoutJob::halign:

Align::LEFT: rect.left() == 0.0

Align::Center: rect.center() == 0.0

Align::RIGHT: rect.right() == 0.0

mesh_bounds: Rect

Tight bounding box around all the meshes in all the rows. Can be used for culling.

num_vertices: usize

Total number of vertices in all the row meshes.

num_indices: usize

Total number of indices in all the row meshes.

pixels_per_point: f32

The number of physical pixels for each logical point. Since this affects the layout, we keep track of it so that we can warn if this has changed once we get to tessellation.

Implementations

impl Galley

pub fn is_empty(&self) -> bool

pub fn text(&self) -> &str

The full, non-elided text of the input job.

pub fn size(&self) -> Vec2

impl Galley

Physical positions

pub fn pos_from_pcursor(&self, pcursor: PCursor) -> Rect

Returns a 0-width Rect.

pub fn pos_from_cursor(&self, cursor: &Cursor) -> Rect

Returns a 0-width Rect.

pub fn cursor_from_pos(&self, pos: Vec2) -> Cursor

Cursor at the given position within the galley

impl Galley

Cursor positions

pub fn end(&self) -> Cursor

Cursor to one-past last character.

pub fn end_rcursor(&self) -> RCursor

impl Galley

Cursor conversions

pub fn from_ccursor(&self, ccursor: CCursor) -> Cursor

pub fn from_rcursor(&self, rcursor: RCursor) -> Cursor

pub fn from_pcursor(&self, pcursor: PCursor) -> Cursor

impl Galley

Cursor positions

pub fn cursor_left_one_character(&self, cursor: &Cursor) -> Cursor

pub fn cursor_right_one_character(&self, cursor: &Cursor) -> Cursor

pub fn cursor_up_one_row(&self, cursor: &Cursor) -> Cursor

pub fn cursor_down_one_row(&self, cursor: &Cursor) -> Cursor

pub fn cursor_begin_of_row(&self, cursor: &Cursor) -> Cursor

pub fn cursor_end_of_row(&self, cursor: &Cursor) -> Cursor