Struct egui::util::History

pub struct History<T> { /* fields omitted */ }

This struct tracks recent values of some time series.

One use is to show a log of recent events, or show a graph over recent events.

It has both a maximum length and a maximum storage time. Elements are dropped when either max length or max age is reached.

Time difference between values can be zero, but never negative.

This can be used for things like smoothed averages (for e.g. FPS) or for smoothed velocity (e.g. mouse pointer speed). All times are in seconds.

Implementations

impl<T> History<T> where
T: Copy,

pub fn new(max_len: usize, max_age: f64) -> Self

pub fn from_max_len_age(max_len: usize, max_age: f64) -> Self

pub fn max_len(&self) -> usize

pub fn max_age(&self) -> f32

pub fn is_empty(&self) -> bool

pub fn len(&self) -> usize

Current number of values kept in history

pub fn total_count(&self) -> u64

Total number of values seen. Includes those that have been discarded due to max_len or max_age.

pub fn latest(&self) -> Option<T>

pub fn latest_mut(&mut self) -> Option<&mut T>

pub fn duration(&self) -> f32

Amount of time contained from start to end in this History.

pub fn iter(&self) -> impl Iterator<Item = (f64, T)> + '_

(time, value) pairs Time difference between values can be zero, but never negative.

pub fn values(&self) -> impl Iterator<Item = T> + '_

pub fn clear(&mut self)

pub fn add(&mut self, now: f64, value: T)

Values must be added with a monotonically increasing time, or at least not decreasing.

pub fn mean_time_interval(&self) -> Option<f32>

Mean time difference between values in this History.

pub fn flush(&mut self, now: f64)

Remove samples that are too old

impl<T> History<T> where
T: Copy,
T: Sum,
T: Div<f32, Output = T>,

pub fn sum(&self) -> T

pub fn average(&self) -> Option<T>

impl<T, Vel> History<T> where
T: Copy,
T: Sub<Output = Vel>,
Vel: Div<f32, Output = Vel>,

pub fn velocity(&self) -> Option<Vel>

Calculate a smooth velocity (per second) over the entire time span