egui::widget_text

Enum WidgetText

pub enum WidgetText {
    Text(String),
    RichText(Arc<RichText>),
    LayoutJob(Arc<LayoutJob>),
    Galley(Arc<Galley>),
}

This is how you specify text for a widget.

A lot of widgets use impl Into<WidgetText> as an argument, allowing you to pass in String, RichText, LayoutJob, and more.

Often a WidgetText is just a simple String, but it can be a RichText (text with color, style, etc), a LayoutJob (for when you want full control of how the text looks) or text that has already been laid out in a Galley.

You can color the text however you want, or use Color32::PLACEHOLDER which will be replaced with a color chosen by the widget that paints the text.

Variants

Text(String)

Plain unstyled text.

We have this as a special case, as it is the common-case, and it uses less memory than Self::RichText.

RichText(Arc<RichText>)

Text and optional style choices for it.

Prefer Self::Text if there is no styling, as it will be faster.

LayoutJob(Arc<LayoutJob>)

Use this LayoutJob when laying out the text.

Only LayoutJob::text and LayoutJob::sections are guaranteed to be respected.

TextWrapping::max_width, LayoutJob::halign, LayoutJob::justify and LayoutJob::first_row_min_height will likely be determined by the crate::Layout of the Ui the widget is placed in. If you want all parts of the LayoutJob respected, then convert it to a Galley and use Self::Galley instead.

You can color the text however you want, or use Color32::PLACEHOLDER which will be replaced with a color chosen by the widget that paints the text.

Galley(Arc<Galley>)

Use exactly this galley when painting the text.

You can color the text however you want, or use Color32::PLACEHOLDER which will be replaced with a color chosen by the widget that paints the text.

Implementations

impl WidgetText

pub fn is_empty(&self) -> bool

pub fn text(&self) -> &str

fn map_rich_text<F>(self, f: F) -> Self
where F: FnOnce(RichText) -> RichText,

Map the contents based on the provided closure.

Self::Text => convert to RichText and call f

Self::RichText => call f

else do nothing

pub fn text_style(self, text_style: TextStyle) -> Self

Override the TextStyle if, and only if, this is a RichText.

Prefer using RichText directly!

pub fn fallback_text_style(self, text_style: TextStyle) -> Self

Set the TextStyle unless it has already been set

Prefer using RichText directly!

pub fn color(self, color: impl Into<Color32>) -> Self

Override text color if, and only if, this is a RichText.

Prefer using RichText directly!

pub fn heading(self) -> Self

Prefer using RichText directly!

pub fn monospace(self) -> Self

Prefer using RichText directly!

pub fn code(self) -> Self

Prefer using RichText directly!

pub fn strong(self) -> Self

Prefer using RichText directly!

pub fn weak(self) -> Self

Prefer using RichText directly!

pub fn underline(self) -> Self

Prefer using RichText directly!

pub fn strikethrough(self) -> Self

Prefer using RichText directly!

pub fn italics(self) -> Self

Prefer using RichText directly!

pub fn small(self) -> Self

Prefer using RichText directly!

pub fn small_raised(self) -> Self

Prefer using RichText directly!

pub fn raised(self) -> Self

Prefer using RichText directly!

pub fn background_color(self, background_color: impl Into<Color32>) -> Self

Prefer using RichText directly!

pub(crate) fn font_height( &self, fonts: &mut FontsView<'_>, style: &Style, ) -> f32

Returns a value rounded to emath::GUI_ROUNDING.

pub fn into_layout_job( self, style: &Style, fallback_font: FontSelection, default_valign: Align, ) -> Arc<LayoutJob>

pub fn into_galley( self, ui: &Ui, wrap_mode: Option<TextWrapMode>, available_width: f32, fallback_font: impl Into<FontSelection>, ) -> Arc<Galley>

Layout with wrap mode based on the containing Ui.

wrap_mode: override for Ui::wrap_mode

pub fn into_galley_impl( self, ctx: &Context, style: &Style, text_wrapping: TextWrapping, fallback_font: FontSelection, default_valign: Align, ) -> Arc<Galley>