Module egui::widgets

Widgets are pieces of GUI such as Label, Button, Slider etc.

Example widget uses:

ui.add(Label::new("Text").text_color(color::red));

if ui.add(Button::new("Click me")).clicked() { ... }

Modules

color_picker

Color picker widgets.

plot

Simple plotting library.

Structs

Button

Clickable button with text.

Checkbox

Boolean on/off control with text label.

CursorPair

DragValue

A numeric value that you can change by dragging the number. More compact than a Slider.

Hyperlink

A clickable hyperlink, e.g. to "https://github.com/emilk/egui".

Image

An widget to show an image of a given size.

ImageButton

A clickable image within a frame.

Label

Static text.

RadioButton

One out of several alternatives, either selected or not.

SelectableLabel

One out of several alternatives, either selected or not. Will mark selected items with a different background color. An alternative to RadioButton and Checkbox.

Separator

A visual separator. A horizontal or vertical line (depending on Layout).

Slider

Control a number by a horizontal slider.

TextEdit

A text region that the user can edit the contents of.

Traits

TextBuffer

Trait constraining what types TextEdit may use as an underlying buffer.

Widget

Anything implementing Widget can be added to a Ui with Ui::add.

Functions

reset_button

Show a button to reset a value to its default. The button is only enabled if the value does not already have its original value.

stroke_ui