Module egui::widgets::drag_value

Structs

DragValue

A numeric value that you can change by dragging the number. More compact than a crate::Slider.

Functions

clamp_value_to_range ğŸ”’

Clamp the given value with careful handling of negative zero, and other corner cases.

default_parser ğŸ”’

The default egui parser of numbers.

get ğŸ”’

parse ğŸ”’

set ğŸ”’

Type Aliases

GetSetValue ğŸ”’

Combined into one function (rather than two) to make it easier for the borrow checker.

NumFormatter ğŸ”’

NumParser ğŸ”’