Module egui::widgets::plot

Simple plotting library.

Modules

items 🔒

Contains items that can be added to a plot.

legend 🔒

transform 🔒

Structs

Arrows

A set of arrows.

AxisBools

Bar

One bar in a BarChart. Potentially floating, allowing stacked bar charts. Width can be changed to allow variable-width histograms.

BarChart

A bar chart.

BoundsLinkGroups 🔒

BoxElem

A box in a BoxPlot diagram. This is a low level graphical element; it will not compute quartiles and whiskers, letting one use their preferred formula. Use Points to draw the outliers.

BoxPlot

A diagram containing a series of BoxElem elements.

BoxSpread

Contains the values of a single box in a box plot.

CoordinatesFormatter

Specifies the coordinates formatting when passed to Plot::coordinates_formatter.

CursorLinkGroups 🔒

GridInput

Input for “grid spacer” functions.

GridMark

One mark (horizontal or vertical line) in the background grid of a plot.

HLine

A horizontal line in a plot, filling the full width

Legend

The configuration for a plot legend.

Line

A series of values forming a path.

LinkedBounds 🔒

Plot

A 2D plot, e.g. a graph of a function.

PlotBounds

2D bounding box of f64 precision. The range of data values we show.

PlotFrameCursors 🔒

Contains the cursors drawn for a plot widget in a single frame.

PlotImage

An image in the plot.

PlotMemory 🔒

Information about the plot that has to persist between frames.

PlotPoint

A point coordinate in the plot.

PlotResponse

What Plot::show returns.

PlotTransform

Contains the screen rectangle and the plot bounds and provides methods to transform between them.

PlotUi

Provides methods to interact with a plot while building it. It is the single argument of the closure provided to Plot::show. See Plot for an example of how to use it.

Points

A set of points.

Polygon

A convex polygon.

PreparedPlot 🔒

Text

Text inside the plot.

VLine

A vertical line in a plot, filling the full width

Enums

BoundsModification 🔒

User-requested modifications to the plot bounds. We collect them in the plot build function to later apply them at the right time, as other modifications need to happen first.

Corner

Where to place the plot legend.

Cursor 🔒

Indicates a vertical or horizontal cursor line in plot coordinates.

LineStyle

MarkerShape

Orientation

Determines whether a plot element is vertically or horizontally oriented.

PlotPoints

Represents many PlotPoints.

Constants

MIN_LINE_SPACING_IN_POINTS 🔒

Functions

fill_marks_between 🔒

Fill in all values between [min, max] which are a multiple of step_size

format_number

Helper for formatting a number so that we always show at least a few decimals, unless it is an integer, in which case we never show any decimals.

generate_marks 🔒

Fill in all values between [min, max] which are a multiple of step_size

log_grid_spacer

Recursively splits the grid into base subdivisions (e.g. 100, 10, 1).

next_power 🔒

Returns next bigger power in given base e.g.

uniform_grid_spacer

Splits the grid into uniform-sized spacings (e.g. 100, 25, 1).

Type Aliases

AxisFormatter 🔒

AxisFormatterFn 🔒

CoordinatesFormatterFn 🔒

GridSpacer 🔒

GridSpacerFn 🔒

LabelFormatter 🔒

LabelFormatterFn 🔒