Module egui::widgets::plot

Simple plotting library.

Structs

Arrows

A set of arrows.

Bar

One bar in a BarChart. Potentially floating, allowing stacked bar charts. Width can be changed to allow variable-width histograms.

BarChart

A bar chart.

BoxElem

A box in a BoxPlot diagram. This is a low level graphical element; it will not compute quartiles and whiskers, letting one use their preferred formula. Use Points to draw the outliers.

BoxPlot

A diagram containing a series of BoxElem elements.

BoxSpread

Contains the values of a single box in a box plot.

HLine

A horizontal line in a plot, filling the full width

Legend

The configuration for a plot legend.

Line

A series of values forming a path.

Plot

A 2D plot, e.g. a graph of a function.

PlotImage

An image in the plot.

PlotUi

Provides methods to interact with a plot while building it. It is the single argument of the closure provided to Plot::show. See Plot for an example of how to use it.

Points

A set of points.

Polygon

A convex polygon.

Text

Text inside the plot.

VLine

A vertical line in a plot, filling the full width

Value

A value in the value-space of the plot.

Values

Enums

Corner

Where to place the plot legend.

LineStyle

MarkerShape