Struct egui::widgets::plot::Plot

pub struct Plot { /* private fields */ }

A 2D plot, e.g. a graph of a function.

Plot supports multiple lines and points.

use egui::plot::{Line, Plot, Value, Values};
let sin = (0..1000).map(|i| {
    let x = i as f64 * 0.01;
    Value::new(x, x.sin())
});
let line = Line::new(Values::from_values_iter(sin));
Plot::new("my_plot").view_aspect(2.0).show(ui, |plot_ui| plot_ui.line(line));

Implementations

impl Plot

pub fn new(id_source: impl Hash) -> Self

Give a unique id for each plot within the same Ui.

pub fn data_aspect(self, data_aspect: f32) -> Self

width / height ratio of the data. For instance, it can be useful to set this to 1.0 for when the two axes show the same unit. By default the plot window’s aspect ratio is used.

pub fn view_aspect(self, view_aspect: f32) -> Self

width / height ratio of the plot region. By default no fixed aspect ratio is set (and width/height will fill the ui it is in).

pub fn width(self, width: f32) -> Self

Width of plot. By default a plot will fill the ui it is in. If you set Self::view_aspect, the width can be calculated from the height.

pub fn height(self, height: f32) -> Self

Height of plot. By default a plot will fill the ui it is in. If you set Self::view_aspect, the height can be calculated from the width.

pub fn min_size(self, min_size: Vec2) -> Self

Minimum size of the plot view.

pub fn show_x(self, show_x: bool) -> Self

Show the x-value (e.g. when hovering). Default: true.

pub fn show_y(self, show_y: bool) -> Self

Show the y-value (e.g. when hovering). Default: true.

pub fn center_x_axis(self, on: bool) -> Self

Always keep the x-axis centered. Default: false.

pub fn center_y_axis(self, on: bool) -> Self

Always keep the y-axis centered. Default: false.

pub fn allow_zoom(self, on: bool) -> Self

Whether to allow zooming in the plot. Default: true.

pub fn allow_drag(self, on: bool) -> Self

Whether to allow dragging in the plot to move the bounds. Default: true.

pub fn custom_label_func<F: 'static + Fn(&str, &Value) -> String>(
self,
custom_lebel_func: F
) -> Self

Provide a function to customize the on-hovel label for the x and y axis

use egui::plot::{Line, Plot, Value, Values};
let sin = (0..1000).map(|i| {
    let x = i as f64 * 0.01;
    Value::new(x, x.sin())
});
let line = Line::new(Values::from_values_iter(sin));
Plot::new("my_plot").view_aspect(2.0)
.custom_label_func(|name, value| {
    if !name.is_empty() {
        format!("{}: {:.*}%", name, 1, value.y).to_string()
    } else {
        "".to_string()
    }
})
.show(ui, |plot_ui| plot_ui.line(line));

pub fn include_x(self, x: impl Into<f64>) -> Self

Expand bounds to include the given x value. For instance, to always show the y axis, call plot.include_x(0.0).

pub fn include_y(self, y: impl Into<f64>) -> Self

Expand bounds to include the given y value. For instance, to always show the x axis, call plot.include_y(0.0).

pub fn legend(self, legend: Legend) -> Self

Show a legend including all named items.

pub fn show_background(self, show: bool) -> Self

Whether or not to show the background Rect. Can be useful to disable if the plot is overlaid over existing content. Default: true.

pub fn show_axes(self, show: [bool; 2]) -> Self

Show the axes. Can be useful to disable if the plot is overlaid over an existing grid or content. Default: [true; 2].

pub fn show<R>(
self,
ui: &mut Ui,
build_fn: impl FnOnce(&mut PlotUi) -> R
) -> InnerResponse<R>

Interact with and add items to the plot and finally draw it.