Struct egui::widgets::plot::PlotUi

pub struct PlotUi { /* private fields */ }

Provides methods to interact with a plot while building it. It is the single argument of the closure provided to Plot::show. See Plot for an example of how to use it.

Implementations

impl PlotUi

pub fn ctx(&self) -> &Context

pub fn plot_bounds(&self) -> PlotBounds

The plot bounds as they were in the last frame. If called on the first frame and the bounds were not further specified in the plot builder, this will return bounds centered on the origin. The bounds do not change until the plot is drawn.

pub fn set_plot_bounds(&mut self, plot_bounds: PlotBounds)

Set the plot bounds. Can be useful for implementing alternative plot navigation methods.

pub fn translate_bounds(&mut self, delta_pos: Vec2)

Move the plot bounds. Can be useful for implementing alternative plot navigation methods.

pub fn plot_hovered(&self) -> bool

Returns true if the plot area is currently hovered.

pub fn plot_clicked(&self) -> bool

Returns true if the plot was clicked by the primary button.

pub fn plot_secondary_clicked(&self) -> bool

Returns true if the plot was clicked by the secondary button.

pub fn pointer_coordinate(&self) -> Option<PlotPoint>

The pointer position in plot coordinates. Independent of whether the pointer is in the plot area.

pub fn pointer_coordinate_drag_delta(&self) -> Vec2

The pointer drag delta in plot coordinates.

pub fn screen_from_plot(&self, position: PlotPoint) -> Pos2

Transform the plot coordinates to screen coordinates.

pub fn plot_from_screen(&self, position: Pos2) -> PlotPoint

Transform the screen coordinates to plot coordinates.

pub fn line(&mut self, line: Line)

Add a data line.

pub fn polygon(&mut self, polygon: Polygon)

Add a polygon. The polygon has to be convex.

pub fn text(&mut self, text: Text)

Add a text.

pub fn points(&mut self, points: Points)

Add data points.

pub fn arrows(&mut self, arrows: Arrows)

Add arrows.

pub fn image(&mut self, image: PlotImage)

Add an image.

pub fn hline(&mut self, hline: HLine)

Add a horizontal line. Can be useful e.g. to show min/max bounds or similar. Always fills the full width of the plot.

pub fn vline(&mut self, vline: VLine)

Add a vertical line. Can be useful e.g. to show min/max bounds or similar. Always fills the full height of the plot.

pub fn box_plot(&mut self, box_plot: BoxPlot)

Add a box plot diagram.

pub fn bar_chart(&mut self, chart: BarChart)

Add a bar chart.