Struct egui::widgets::Button

pub struct Button<'a> { /* private fields */ }

Clickable button with text.

See also Ui::button.

if ui.add(egui::Button::new("Click me")).clicked() {
    do_stuff();
}

// A greyed-out and non-interactive button:
if ui.add_enabled(false, egui::Button::new("Can't click this")).clicked() {
    unreachable!();
}

Implementations

impl<'a> Button<'a>

pub fn new(text: impl Into<WidgetText>) -> Self

pub fn image(image: impl Into<Image<'a>>) -> Self

Creates a button with an image. The size of the image as displayed is defined by the provided size.

pub fn image_and_text( image: impl Into<Image<'a>>, text: impl Into<WidgetText> ) -> Self

Creates a button with an image to the left of the text. The size of the image as displayed is defined by the provided size.

pub fn opt_image_and_text( image: Option<Image<'a>>, text: Option<WidgetText> ) -> Self

pub fn wrap(self, wrap: bool) -> Self

If true, the text will wrap to stay within the max width of the Ui.

By default Self::wrap will be true in vertical layouts and horizontal layouts with wrapping, and false on non-wrapping horizontal layouts.

Note that any \n in the text will always produce a new line.

pub fn fill(self, fill: impl Into<Color32>) -> Self

Override background fill color. Note that this will override any on-hover effects. Calling this will also turn on the frame.

pub fn stroke(self, stroke: impl Into<Stroke>) -> Self

Override button stroke. Note that this will override any on-hover effects. Calling this will also turn on the frame.

pub fn small(self) -> Self

Make this a small button, suitable for embedding into text.

pub fn frame(self, frame: bool) -> Self

Turn off the frame

pub fn sense(self, sense: Sense) -> Self

By default, buttons senses clicks. Change this to a drag-button with Sense::drag().

pub fn min_size(self, min_size: Vec2) -> Self

Set the minimum size of the button.

pub fn rounding(self, rounding: impl Into<Rounding>) -> Self

Set the rounding of the button.

pub fn shortcut_text(self, shortcut_text: impl Into<WidgetText>) -> Self

Show some text on the right side of the button, in weak color.

Designed for menu buttons, for setting a keyboard shortcut text (e.g. Ctrl+S).

The text can be created with Context::format_shortcut.

pub fn selected(self, selected: bool) -> Self

If true, mark this button as “selected”.