Struct egui::widgets::Button

pub struct Button { /* private fields */ }

Clickable button with text.

See also Ui::button.

if ui.add(egui::Button::new("Click me")).clicked() {
    do_stuff();
}

// A greyed-out and non-interactive button:
if ui.add_enabled(false, egui::Button::new("Can't click this")).clicked() {
    unreachable!();
}

Implementations

impl Button

pub fn new(text: impl Into<WidgetText>) -> Self

pub fn image_and_text(
texture_id: TextureId,
size: impl Into<Vec2>,
text: impl Into<WidgetText>
) -> Self

Creates a button with an image to the left of the text. The size of the image as displayed is defined by the size Vec2 provided.

pub fn wrap(self, wrap: bool) -> Self

If true, the text will wrap to stay within the max width of the Ui.

By default Self::wrap will be true in vertical layouts and horizontal layouts with wrapping, and false on non-wrapping horizontal layouts.

Note that any \n in the text will always produce a new line.

pub fn text_color(self, text_color: Color32) -> Self

👎Deprecated: Replaced by: Button::new(RichText::new(text).color(…))

pub fn text_style(self, text_style: TextStyle) -> Self

👎Deprecated: Replaced by: Button::new(RichText::new(text).text_style(…))

pub fn fill(self, fill: impl Into<Color32>) -> Self

Override background fill color. Note that this will override any on-hover effects. Calling this will also turn on the frame.

pub fn stroke(self, stroke: impl Into<Stroke>) -> Self

Override button stroke. Note that this will override any on-hover effects. Calling this will also turn on the frame.

pub fn small(self) -> Self

Make this a small button, suitable for embedding into text.

pub fn frame(self, frame: bool) -> Self

Turn off the frame

pub fn sense(self, sense: Sense) -> Self

By default, buttons senses clicks. Change this to a drag-button with Sense::drag().