Struct egui::widgets::DragValue

#[must_use = "You should put this widget in an ui with `ui.add(widget);`"]pub struct DragValue<'a> { /* fields omitted */ }

A numeric value that you can change by dragging the number. More compact than a Slider.

ui.add(egui::DragValue::new(&mut my_f32).speed(0.1));

Implementations

impl<'a> DragValue<'a>

pub fn new<Num: Numeric>(value: &'a mut Num) -> Self

pub fn f32(value: &'a mut f32) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn f64(value: &'a mut f64) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn i8(value: &'a mut i8) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn u8(value: &'a mut u8) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn i16(value: &'a mut i16) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn u16(value: &'a mut u16) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn i32(value: &'a mut i32) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn u32(value: &'a mut u32) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn i64(value: &'a mut i64) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn u64(value: &'a mut u64) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn isize(value: &'a mut isize) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn usize(value: &'a mut usize) -> Self

👎 Deprecated:
Use DragValue::new instead

pub fn from_get_set(get_set_value: impl 'a + FnMut(Option<f64>) -> f64) -> Self

pub fn speed(self, speed: impl Into<f64>) -> Self

How much the value changes when dragged one point (logical pixel).

pub fn clamp_range<Num: Numeric>(self, clamp_range: RangeInclusive<Num>) -> Self

Clamp incoming and outgoing values to this range.

pub fn clamp_range_f64(self, clamp_range: RangeInclusive<f64>) -> Self

👎 Deprecated:
Use clamp_range

pub fn range(self, clamp_range: RangeInclusive<f32>) -> Self

👎 Deprecated:
Renamed clamp_range

pub fn prefix(self, prefix: impl ToString) -> Self

Show a prefix before the number, e.g. "x: "

pub fn suffix(self, suffix: impl ToString) -> Self

Add a suffix to the number, this can be e.g. a unit ("°" or " m")

pub fn min_decimals(self, min_decimals: usize) -> Self

Set a minimum number of decimals to display. Normally you don't need to pick a precision, as the slider will intelligently pick a precision for you. Regardless of precision the slider will use "smart aim" to help the user select nice, round values.

pub fn max_decimals(self, max_decimals: usize) -> Self

Set a maximum number of decimals to display. Values will also be rounded to this number of decimals. Normally you don't need to pick a precision, as the slider will intelligently pick a precision for you. Regardless of precision the slider will use "smart aim" to help the user select nice, round values.

pub fn max_decimals_opt(self, max_decimals: Option<usize>) -> Self

pub fn fixed_decimals(self, num_decimals: usize) -> Self

Set an exact number of decimals to display. Values will also be rounded to this number of decimals. Normally you don't need to pick a precision, as the slider will intelligently pick a precision for you. Regardless of precision the slider will use "smart aim" to help the user select nice, round values.