egui::widgets
Struct Label

pub struct Label { /* private fields */ }

Static text.

Usually it is more convenient to use Ui::label.

ui.label("Equivalent");
ui.add(egui::Label::new("Equivalent"));
ui.add(egui::Label::new("With Options").truncate());
ui.label(egui::RichText::new("With formatting").underline());

For full control of the text you can use crate::text::LayoutJob as argument to Self::new.

Implementations

impl Label

pub fn new(text: impl Into<WidgetText>) -> Self

pub fn text(&self) -> &str

pub fn wrap_mode(self, wrap_mode: TextWrapMode) -> Self

Set the wrap mode for the text.

By default, crate::Ui::wrap_mode will be used, which can be overridden with crate::Style::wrap_mode.

Note that any \n in the text will always produce a new line.

pub fn wrap(self) -> Self

Set Self::wrap_mode to TextWrapMode::Wrap.

pub fn truncate(self) -> Self

Set Self::wrap_mode to TextWrapMode::Truncate.

pub fn extend(self) -> Self

Set Self::wrap_mode to TextWrapMode::Extend, disabling wrapping and truncating, and instead expanding the parent Ui.

pub fn halign(self, align: Align) -> Self

Sets the horizontal alignment of the Label to the given Align value.

pub fn selectable(self, selectable: bool) -> Self

Can the user select the text with the mouse?

Overrides crate::style::Interaction::selectable_labels.

pub fn sense(self, sense: Sense) -> Self

Make the label respond to clicks and/or drags.

By default, a label is inert and does not respond to click or drags. By calling this you can turn the label into a button of sorts. This will also give the label the hover-effect of a button, but without the frame.

if ui.add(Label::new("click me").sense(Sense::click())).clicked() {
    /* … */
}

impl Label

pub fn layout_in_ui(self, ui: &mut Ui) -> (Pos2, Arc<Galley>, Response)

Do layout and position the galley in the ui, without painting it or adding widget info.