Struct egui::widgets::Label

pub struct Label { /* fields omitted */ }

Static text.

ui.label("Equivalent");
ui.add(egui::Label::new("Equivalent"));
ui.add(egui::Label::new("With Options").text_color(egui::Color32::RED));

Implementations

impl Label

pub fn new(text: impl ToString) -> Self

pub fn text(&self) -> &str

pub fn wrap(self, wrap: bool) -> Self

If true, the text will wrap at the max_width. By default Self::wrap will be true in vertical layouts and horizontal layouts with wrapping, and false on non-wrapping horizontal layouts.

Note that any \n in the text label will always produce a new line.

pub fn multiline(self, multiline: bool) -> Self

👎 Deprecated:
Use Label::wrap instead

pub fn text_style(self, text_style: TextStyle) -> Self

The default is Style::body_text_style (generally TextStyle::Body).

pub fn heading(self) -> Self

pub fn monospace(self) -> Self

pub fn code(self) -> Self

Monospace label with gray background

pub fn strong(self) -> Self

Extra strong text (stronger color).

pub fn weak(self) -> Self

Extra weak text (fainter color).

pub fn underline(self) -> Self

draw a line under the text

pub fn strikethrough(self) -> Self

draw a line through the text, crossing it out

pub fn italics(self) -> Self

tilt the characters to the right.

pub fn small(self) -> Self

Smaller text

pub fn small_raised(self) -> Self

For e.g. exponents

pub fn raised(self) -> Self

Align text to top. Only applicable together with Self::small().

pub fn background_color(self, background_color: impl Into<Color32>) -> Self

Fill-color behind the text

pub fn text_color(self, text_color: impl Into<Color32>) -> Self

pub fn sense(self, sense: Sense) -> Self

Make the label respond to clicks and/or drags.

By default, a label is inert and does not respond to click or drags. By calling this you can turn the label into a button of sorts. This will also give the label the hover-effect of a button, but without the frame.

if ui.add(Label::new("click me").sense(Sense::click())).clicked() {
    /* … */
}

impl Label

pub fn layout(&self, ui: &Ui) -> Arc<Galley>

pub fn layout_width(&self, ui: &Ui, max_width: f32) -> Arc<Galley>

pub fn font_height(&self, fonts: &Fonts, style: &Style) -> f32

pub fn paint_galley(&self, ui: &mut Ui, pos: Pos2, galley: Arc<Galley>)

pub fn text_style_or_default(&self, style: &Style) -> TextStyle

Read the text style, or get the default for the current style