egui::widgets
Struct ProgressBar

pub struct ProgressBar { /* private fields */ }

A simple progress bar.

See also: crate::Spinner.

Implementations

impl ProgressBar

pub fn new(progress: f32) -> Self

Progress in the [0, 1] range, where 1 means “completed”.

pub fn desired_width(self, desired_width: f32) -> Self

The desired width of the bar. Will use all horizontal space if not set.

pub fn desired_height(self, desired_height: f32) -> Self

The desired height of the bar. Will use the default interaction size if not set.

pub fn fill(self, color: Color32) -> Self

The fill color of the bar.

pub fn text(self, text: impl Into<WidgetText>) -> Self

A custom text to display on the progress bar.

pub fn show_percentage(self) -> Self

Show the progress in percent on the progress bar.

pub fn animate(self, animate: bool) -> Self

Whether to display a loading animation when progress < 1. Note that this will cause the UI to be redrawn. Defaults to false.

If Self::rounding and Self::animate are used simultaneously, the animation is not rendered, since it requires a perfect circle to render correctly. However, the UI is still redrawn.

pub fn rounding(self, rounding: impl Into<Rounding>) -> Self

Set the rounding of the progress bar.

If Self::rounding and Self::animate are used simultaneously, the animation is not rendered, since it requires a perfect circle to render correctly. However, the UI is still redrawn.