Struct egui::widgets::TextEdit

#[must_use = "You should put this widget in an ui with `ui.add(widget);`"]pub struct TextEdit<'t> { /* fields omitted */ }

A text region that the user can edit the contents of.

See also Ui::text_edit_singleline and Ui::text_edit_multiline.

Example:

let response = ui.add(egui::TextEdit::singleline(&mut my_string));
if response.changed() {
    // …
}
if response.lost_focus() && ui.input().key_pressed(egui::Key::Enter) {
    // …
}

Implementations

impl<'t> TextEdit<'t>

pub fn cursor(ui: &Ui, id: Id) -> Option<CursorPair>

impl<'t> TextEdit<'t>

pub fn new(text: &'t mut String) -> Self

👎 Deprecated:
Use TextEdit::singleline or TextEdit::multiline (or the helper ui.text_edit_singleline, ui.text_edit_multiline) instead

pub fn singleline(text: &'t mut String) -> Self

No newlines (\n) allowed. Pressing enter key will result in the TextEdit losing focus (response.lost_focus).

pub fn multiline(text: &'t mut String) -> Self

A TextEdit for multiple lines. Pressing enter key will create a new line.

pub fn id(self, id: Id) -> Self

pub fn id_source(self, id_source: impl Hash) -> Self

A source for the unique Id, e.g. .id_source("second_text_edit_field") or .id_source(loop_index).

pub fn hint_text(self, hint_text: impl Into<String>) -> Self

Show a faint hint text when the text field is empty.

pub fn password(self, password: bool) -> Self

If true, hide the letters from view and prevent copying from the field.

pub fn text_style(self, text_style: TextStyle) -> Self

pub fn text_color(self, text_color: Color32) -> Self

pub fn text_color_opt(self, text_color: Option<Color32>) -> Self

pub fn enabled(self, enabled: bool) -> Self

Default is true. If set to false then you cannot edit the text.

pub fn frame(self, frame: bool) -> Self

Default is true. If set to false there will be no frame showing that this is editable text!

pub fn desired_width(self, desired_width: f32) -> Self

Set to 0.0 to keep as small as possible

pub fn desired_rows(self, desired_height_rows: usize) -> Self

Set the number of rows to show by default. The default for singleline text is 1. The default for multiline text is 4.