Trait egui::widgets::TextBuffer

pub trait TextBuffer: AsRef<str> + Into<String> {
    fn insert_text(&mut self, text: &str, ch_idx: usize) -> usize;

    fn delete_char_range(&mut self, ch_range: Range<usize>);

    fn as_str(&self) -> &str { ... }

    fn clear(&mut self) { ... }

    fn replace(&mut self, text: &str) { ... }

    fn take(&mut self) -> String { ... }
}

Trait constraining what types TextEdit may use as an underlying buffer.

Most likely you will use a String which implements TextBuffer.

Required methods

fn insert_text(&mut self, text: &str, ch_idx: usize) -> usize

Inserts text text into this buffer at character index ch_idx.

Notes

ch_idx is a character index, not a byte index.

Return

Returns how many characters were successfully inserted

fn delete_char_range(&mut self, ch_range: Range<usize>)

Deletes a range of text ch_range from this buffer.

Notes

ch_range is a character range, not a byte range.

Provided methods

fn as_str(&self) -> &str

Returns this buffer as a str.

This is an utility method, as it simply relies on the AsRef<str> implementation.

fn clear(&mut self)

Clears all characters in this buffer

fn replace(&mut self, text: &str)

Replaces all contents of this string with text

fn take(&mut self) -> String

Clears all characters in this buffer and returns a string of the contents.