Trait egui::widgets::Widget

#[must_use = "You should put this widget in an ui with `ui.add(widget);`"]pub trait Widget {
    fn ui(self, ui: &mut Ui) -> Response;
}

Anything implementing Widget can be added to a Ui with Ui::add.

Examples include [Button], [Label] and Slider.

|ui: &mut Ui| -> Response { … } also implemented Widget.

Required methods

fn ui(self, ui: &mut Ui) -> Response

Allocate space, interact, paint, and return a Response.

Loading content...